% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{addXInfo}
\alias{addXInfo}
\title{Add experiment info to the given matrix.}
\usage{
addXInfo(x, xinfo)
}
\arguments{
\item{xinfo}{A \code{matrix} of experiment info.}

\item{tab}{A \code{matrix} to which experiment info will be added.}

\item{col.index}{Index of the column in \code{tab} at which experiment
info will be inserted. Columns in the original \code{tab} matrix that
have an index equal to or greater than \code{col.index} will be moved
to the right of the inserted experiment info in the returned matrix.}
}
\value{
The input matrix with experiment info added at the specified column.
}
\description{
Add experiment info to the given matrix.
}
\keyword{internal}

