% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{allKwargs}
\alias{allKwargs}
\title{Test if ellipsis arguments are all keyword arguments.}
\usage{
allKwargs(...)
}
\arguments{
\item{...}{Ellipsis arguments.}
}
\value{
TRUE if all ellipsis arguments are keyword arguments; FALSE otherwise.
}
\description{
Test if ellipsis arguments are all keyword arguments.
}
\keyword{internal}

