% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{allNA}
\alias{allNA}
\title{Test if all elements are NA values.}
\usage{
allNA(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if object has nonzero length and all elements are NA values; 
FALSE otherwise.
}
\description{
Test if all elements are NA values.
}
\keyword{internal}

