% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{allWhite}
\alias{allWhite}
\title{Test if vector is all whitespace.}
\usage{
allWhite(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
TRUE if character vector is of length zero or contains only
whitespace characters; FALSE otherwise.
}
\description{
Test if vector is all whitespace.
}
\keyword{internal}

