% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{anyKwargs}
\alias{anyKwargs}
\title{Test if any ellipsis arguments are keyword arguments.}
\usage{
anyKwargs(...)
}
\arguments{
\item{...}{Ellipsis arguments.}
}
\value{
TRUE if any ellipsis arguments are keyword arguments; FALSE otherwise.
}
\description{
Test if any ellipsis arguments are keyword arguments.
}
\keyword{internal}

