% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{as.data.frame.geno}
\alias{as.data.frame.geno}
\title{Coerce \code{geno} object to \code{data.frame}.}
\usage{
\method{as.data.frame}{geno}(x, ..., chr = NULL, digits = NULL,
  missing.value = "-", include.mapunit = TRUE)
}
\arguments{
\item{x}{A \code{geno} object.}

\item{...}{Unused arguments.}

\item{chr}{Vector of sequences for which genotype data should be included in 
the genotype \code{data.frame}. If none are specified, a genotype 
\code{data.frame} is returned for all available sequences.}

\item{digits}{If specified, round genetic map positions to the specified 
number of digits.}

\item{missing.value}{Missing data value. This can be any single character
that is not a possible phenotype or genotype value.}

\item{include.mapunit}{Include map unit information in map positions.}
}
\value{
A \code{data.frame} corresponding to the input \code{geno} object.
}
\description{
Coerce \code{geno} object to \code{data.frame}.
}
\keyword{internal}

