% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{as.data.frame.map}
\alias{as.data.frame.map}
\title{Convert \code{map} to \code{data.frame}.}
\usage{
\method{as.data.frame}{map}(x, ..., map.unit = NULL)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{map} object.}

\item{...}{Unused arguments.}

\item{map.unit}{Explicitly sets \code{'map.unit'} attribute.}
}
\value{
A \code{data.frame} corresponding to the input \code{map} object.
}
\description{
Convert \code{map} to \code{data.frame}.
}
\seealso{
Other map utility functions: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{hasMapCSV}}, \code{\link{inMapOrder}},
  \code{\link{inferMapStep}}, \code{\link{mapframe}},
  \code{\link{orderMap}}, \code{\link{readMapCSV}},
  \code{\link{readMapHDF5}}, \code{\link{readMapframeCSV}},
  \code{\link{setMapUnit}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapHDF5}},
  \code{\link{writeMapframeCSV}}
}

