% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{as.data.frame.mapframe}
\alias{as.data.frame.mapframe}
\title{Coerce \code{mapframe} to \code{data.frame}.}
\usage{
\method{as.data.frame}{mapframe}(x, ...)
}
\arguments{
\item{x}{A \code{mapframe} object.}

\item{...}{Unused arguments.}
}
\value{
A \code{data.frame} corresponding to the input \code{mapframe} object.
}
\description{
Coerce \code{mapframe} to \code{data.frame}.
}
\seealso{
Other map utilities: \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{inMapOrder}}, \code{\link{inferMapStep}},
  \code{\link{mapframe}}, \code{\link{orderMap}},
  \code{\link{setMapUnit}}
}

