% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{as.data.frame.pheno}
\alias{as.data.frame.pheno}
\title{Convert \code{pheno} object to \code{data.frame}.}
\usage{
\method{as.data.frame}{pheno}(x, ..., digits = NULL)
}
\arguments{
\item{x}{A \code{pheno} object.}

\item{...}{Unused arguments.}

\item{digits}{If specified, round numeric phenotype values to the given
number of digits.}
}
\value{
A \code{data.frame} corresponding to the input \code{pheno} object.
}
\description{
Convert \code{pheno} object to \code{data.frame}.
}
\keyword{internal}

