% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{as.geno}
\alias{as.geno}
\alias{as.geno.data.frame}
\title{Convert to \code{geno} object.}
\usage{
as.geno(from, require.mapunit = TRUE)

\method{as.geno}{data.frame}(from, require.mapunit = TRUE)
}
\arguments{
\item{from}{Object containing genotype data.}

\item{require.mapunit}{Require map unit information in map positions.}
}
\value{
A \code{geno} object corresponding to the input object.
}
\description{
Convert to \code{geno} object.
}
\keyword{internal}
