% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{as.map}
\alias{as.map}
\alias{as.map.data.frame}
\alias{as.map.list}
\alias{as.map.mapframe}
\alias{as.map.scanone}
\title{Coerce object to \code{map}.}
\usage{
as.map(from, map.unit = NULL)

\method{as.map}{data.frame}(from, map.unit = NULL)

\method{as.map}{list}(from, map.unit = NULL)

\method{as.map}{mapframe}(from, map.unit = NULL)

\method{as.map}{scanone}(from, map.unit = NULL)
}
\arguments{
\item{from}{Object containing map data.}

\item{map.unit}{Explicitly sets \code{'map.unit'}. This must be specified if
an object does not have an existing \code{'map.unit'} attribute.}
}
\value{
An \pkg{R/qtl} \code{map} corresponding to the input object.
}
\description{
Coerce object to \code{map}.
}
\seealso{
Other map utilities: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{inMapOrder}}, \code{\link{inferMapStep}},
  \code{\link{mapframe}}, \code{\link{orderMap}},
  \code{\link{setMapUnit}}
}

