% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{as.mapping}
\alias{as.mapping}
\title{Convert to a \code{mapping} object.}
\usage{
as.mapping(x)
}
\arguments{
\item{x}{Named vector.}
}
\value{
A \code{mapping} corresponding to the input object.
}
\description{
Convert to a \code{mapping} object.
}
\seealso{
Other mapping object functions: \code{\link{is.mapping}},
  \code{\link{mapping}}, \code{\link{values}}
}

