% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{as.pheno}
\alias{as.pheno}
\alias{as.pheno.data.frame}
\title{Convert to \code{pheno} object.}
\usage{
as.pheno(from)

\method{as.pheno}{data.frame}(from)
}
\arguments{
\item{from}{Object containing phenotype data.}
}
\value{
A \code{pheno} object corresponding to the input object.
}
\description{
Convert to \code{pheno} object.
}
\keyword{internal}

