% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScanone.R
\name{batchPermScanone}
\alias{batchPermScanone}
\title{Run \code{qtl::scanone} on a batch of permuted \code{cross} objects.}
\usage{
batchPermScanone(cross, pheno.col = NULL, n.cluster = 1, iseed = NULL,
  n.perm = 1000, perm.pheno = TRUE, perm.geno = FALSE,
  perm.type = c("max", "bins"), ...)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{pheno.col}{Phenotype columns for which QTL analysis should be run.
If no phenotypes are specified, all are used.}

\item{n.cluster}{Number of nodes to request for parallel execution. Any
number of processes can be requested with the argument \code{n.cluster},
but the number actually returned can be limited to the number of processes
available. On a local computer, this limit is the number of physical cores
in the computer. In a PBS batch system, the maximum number of processes
will be set from the PBS node file.}

\item{iseed}{Seed for random number generator.}

\item{n.perm}{Number of permutations.}

\item{perm.pheno}{Permute phenotype data.}

\item{perm.geno}{Permute genotype data.}

\item{perm.type}{Type of permutation data (see below).}

\item{...}{Additional keyword arguments passed to \code{scanone}.}
}
\value{
If \code{perm.type} is set to \code{'max'}, a regular
\code{scanoneperm} object is returned, containing the maximum
LOD values from the given permutations. If \code{perm.type} is
set to \code{'bins'}, a \code{scanonebins} array is returned.
Each row of this array corresponds to a permutation, each array
column corresponds to a bin spanning an interval of LOD values,
and each slice corresponds to a LOD column. Each element contains
the number of loci in the given bin interval for that LOD column
in that permutation.
}
\description{
Run \code{qtl::scanone} on a batch of permuted \code{cross} objects.
}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})
}
\seealso{
Other scan utility functions: \code{\link{batchPermScanoneF}},
  \code{\link{batchPermScantwoF}},
  \code{\link{batchPermScantwo}},
  \code{\link{batchPhenoScanone}},
  \code{\link{batchPhenoScantwo}}, \code{\link{batchScan}},
  \code{\link{nodePermScanoneF}},
  \code{\link{nodePermScanone}},
  \code{\link{nodePermScantwoF}},
  \code{\link{nodePermScantwo}},
  \code{\link{nodePhenoScanone}},
  \code{\link{nodePhenoScantwo}}
}

