% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScanoneF.R
\name{batchPermScanoneF}
\alias{batchPermScanoneF}
\title{Run \code{funqtl::scanoneF} on a batch of permuted \code{cross} objects.}
\usage{
batchPermScanoneF(cross, pheno.col = NULL, n.cluster = 1, iseed = NULL,
  n.perm = 1000, perm.pheno = TRUE, perm.geno = FALSE, ...)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{pheno.col}{Phenotype columns for which QTL analysis should be 
performed. If specified, phenotypes should contain measurements for 
consecutive values of the parameter of the function-valued trait 
(e.g. times). If no phenotypes are specified, all are used.}

\item{n.cluster}{Number of nodes to request for parallel execution. Any
number of processes can be requested with the argument \code{n.cluster},
but the number actually returned can be limited to the number of processes
available. On a local computer, this limit is the number of physical cores
in the computer. In a PBS batch system, the maximum number of processes
will be set from the PBS node file.}

\item{iseed}{Seed for random number generator.}

\item{n.perm}{Number of permutations.}

\item{perm.pheno}{Permute phenotype data.}

\item{perm.geno}{Permute genotype data.}

\item{...}{Additional keyword arguments passed to \code{scanoneF}.}
}
\value{
A \code{scanoneperm} matrix containing the results of \code{scanoneF}
for all permutations.
}
\description{
Run \code{funqtl::scanoneF} on a batch of permuted \code{cross} objects.
}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})

Kwak IY, Moore CR, Spalding EP, Broman KW (2014) A simple
regression-based method to map quantitative trait loci underlying
function-valued phenotypes. \emph{Genetics} \bold{197}(4):1409-16.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/24931408}{PubMed})
}
\seealso{
Other scan utility functions: \code{\link{batchPermScanone}},
  \code{\link{batchPermScantwoF}},
  \code{\link{batchPermScantwo}},
  \code{\link{batchPhenoScanone}},
  \code{\link{batchPhenoScantwo}}, \code{\link{batchScan}},
  \code{\link{nodePermScanoneF}},
  \code{\link{nodePermScanone}},
  \code{\link{nodePermScantwoF}},
  \code{\link{nodePermScantwo}},
  \code{\link{nodePhenoScanone}},
  \code{\link{nodePhenoScantwo}}
}

