% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScan.R
\name{batchScan}
\alias{batchScan}
\title{Perform a batch of QTL analysis.}
\usage{
batchScan(x, scanfunction, cross, n.cluster = 1, iseed = NULL, ...)
}
\arguments{
\item{x}{An atomic vector to which the scan function is applied elementwise.}

\item{scanfunction}{QTL scan function. The first argument of this function 
must be the element of \code{x} that is passed by \code{clusterApply}, and 
its second argument should be the \code{cross} object.}

\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{n.cluster}{Number of nodes to request for parallel execution. Any
number of processes can be requested with the argument \code{n.cluster},
but the number actually returned can be limited to the number of processes
available. On a local computer, this limit is the number of physical cores
in the computer. In a PBS batch system, the maximum number of processes
will be set from the PBS node file.}

\item{iseed}{Seed for random number generator stream.}

\item{...}{Additional keyword arguments passed to the scan function.}
}
\value{
List of QTL scan results.
}
\description{
The specified \code{scanfunction} is applied to the \code{cross} for every 
element in vector \code{x}. This function uses the \pkg{parallel} package to
set up a cluster, initialiase a random-number generator stream, and run the
batch in parallel.
}
\seealso{
Other scan utility functions: \code{\link{batchPermScanoneF}},
  \code{\link{batchPermScanone}},
  \code{\link{batchPermScantwoF}},
  \code{\link{batchPermScantwo}},
  \code{\link{batchPhenoScanone}},
  \code{\link{batchPhenoScantwo}},
  \code{\link{nodePermScanoneF}},
  \code{\link{nodePermScanone}},
  \code{\link{nodePermScantwoF}},
  \code{\link{nodePermScantwo}},
  \code{\link{nodePhenoScanone}},
  \code{\link{nodePhenoScantwo}}
}

