% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binLODValues.R
\name{binLODValues}
\alias{binLODValues}
\title{Bin LOD values of \code{scanone} object.}
\usage{
binLODValues(x, lodcolumns = NULL)
}
\arguments{
\item{x}{A \code{scanone} object or equivalent \code{data.frame} in which the
data columns contain LOD score values.}

\item{lodcolumns}{In a \code{scanone} or equivalent object, this parameter 
indicates which LOD columns to consider. These must be either a character
vector of column names or a numeric vector of indices \emph{with respect
to the set of LOD columns}. If no LOD columns are specified, all are used.}
}
\value{
A \code{scanonebins} array of LOD bin counts, with a single row, each
column corresponding to a bin spanning an interval of LOD values, and each
slice corresponding to a LOD column. Each column name indicates the half-open
interval that defines the given bin (e.g. \code{'LOD[1,2)'} for a bin that
contains the number of loci with LOD scores greater than or equal to one and
less than two). Returned bins cover the range of LOD values from 1.0 until the
lowest integer that is greater than the maximum observed LOD value.
}
\description{
Bin LOD values of \code{scanone} object.
}

