% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bstripBlankRows}
\alias{bstripBlankRows}
\title{Strip blank rows from bottom of \code{data.frame}.}
\usage{
bstripBlankRows(x)
}
\arguments{
\item{x}{A \code{data.frame} with columns of type \code{character}.}
}
\value{
Input object in which bottommost blank rows (i.e. rows
containing no non-whitespace characters) have been stripped.
}
\description{
Strip blank rows from bottom of \code{data.frame}.
}
\keyword{internal}
