% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clamp}
\alias{clamp}
\title{Clamp numbers within a range.}
\usage{
clamp(n, interval)
}
\arguments{
\item{n}{Numeric vector.}

\item{interval}{Numeric vector containing the minimum and maximum values of
the range, respectively.}
}
\value{
Input vector with values clamped within the specified range.
}
\description{
Clamp numbers within a range.
}
\keyword{internal}

