% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5Stack-class.R
\docType{methods}
\name{closeStack}
\alias{closeStack}
\alias{closeStack,H5Stack-method}
\title{Close all HDF5 objects in \code{H5Stack}.}
\usage{
closeStack(h5stack)

\S4method{closeStack}{H5Stack}(h5stack)
}
\arguments{
\item{h5stack}{A \code{H5Stack} object.}
}
\value{
Empty \code{H5Stack} object.
}
\description{
Close all HDF5 objects in \code{H5Stack}.
}
\keyword{internal}

