% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{coerceDataFrame}
\alias{coerceDataFrame}
\title{Coerce \code{data.frame} columns to the specified classes.}
\usage{
coerceDataFrame(x, classes)
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{classes}{Classes to which columns of \code{data.frame} are to be 
coerced. This should be a character vector containing class names, with 
the same number of elements as there are columns in the \code{data.frame}.}
}
\value{
Coerced \code{data.frame}.
}
\description{
Coerce \code{data.frame} columns to the specified classes.
}
\keyword{internal}

