% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCrossInfo.R
\docType{methods}
\name{compareCrossInfo}
\alias{compareCrossInfo}
\alias{compareCrossInfo,CrossInfo-method}
\alias{compareCrossInfo.CrossInfo}
\alias{compareCrossInfo.cross}
\title{Compare \code{cross} and \code{CrossInfo} objects.}
\usage{
compareCrossInfo(...)

compareCrossInfo.cross(cross)

compareCrossInfo.CrossInfo(cross.info, cross)

compareCrossInfo(...)

\S4method{compareCrossInfo}{CrossInfo}(cross.info, cross)
}
\arguments{
\item{...}{Arguments (see below).}

\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
TRUE if \code{CrossInfo} and \pkg{R/qtl} \code{cross} objects are 
concordant; otherwise returns a character vector of mismatch errors.
}
\description{
Test concordance of an \pkg{R/qtl} \code{cross} and a CrossInfo 
object. If an \pkg{R/qtl} \code{cross} is the first argument, it should have 
attribute \code{'info'} containing a \code{CrossInfo} object. If a 
\code{CrossInfo} object is the first argument, the second argument should be 
the \pkg{R/qtl} \code{cross} object with which it is being compared.
}
\seealso{
\code{\linkS4class{CrossInfo}}
}
\keyword{internal}

