% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{convertMapUnit}
\alias{convertMapUnit}
\title{Convert map unit and rescale map positions.}
\usage{
convertMapUnit(x, map.unit)
}
\arguments{
\item{x}{Object containing map data.}

\item{map.unit}{New map unit.}
}
\value{
Input object with converted map unit and rescaled map positions.
}
\description{
Convert map unit and rescale map positions.
}
\section{Map Units}{
 
A \code{map} or \code{mapframe} must express positions in either genetic or 
physical map units. Genetic map positions must be in centiMorgans (cM), while 
physical map positions must be in base pairs (bp), kilobases (kb), or 
Megabases (Mb).
}
\keyword{internal}

