% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\docType{package}
\name{Cross Utilities}
\alias{Cross Utilities}
\alias{Cross Utilities-package}
\title{Utilities for handling \pkg{R/qtl} \code{cross} objects.}
\description{
Cross utilities can query or modify an existing \pkg{R/qtl} 
\code{cross} object.
}
\details{
An \pkg{R/qtl} \code{cross} object contains phenotype data for a set 
of individual samples, along with genotype data for the same samples for 
multiple marker loci across one or more sequences.

These utilities include functions to pull sequences
(\code{\link{pull.chr}}), individual sample IDs (\code{\link{pull.ind}}), or
alleles (\code{\link{pull.alleles}}). They also include functions to get
column indices within an \pkg{R/qtl} \code{cross$pheno} \code{data.frame}, 
for the ID column (\code{\link{getIdColIndex}}) and phenotype columns 
\code{\link{getPhenoColIndices}}.

There are functions to infer tetrad indices 
(\code{\link{inferTetradIndices}}) or strain indices 
(\code{\link{inferStrainIndices}}) for a \code{cross} object, based on 
sample IDs (if available) and genotype data.

A \code{cross} object can be modified by permuting either its phenotype or 
genotype data (\code{\link{permCross}}). When permuting phenotypes, it may
also be necessary to permute any associated data (e.g. covariates) in the 
same way. In such cases, \code{\link{permIndices}} can be used to generate
a set of permutation indices, which can in turn be passed to 
\code{\link{permCross}} or used to permute associated data.

There are several functions for handling a \code{cross} with a time-series 
phenotype, such that phenotypes have been observed at regular time points, 
and successive phenotype columns are named for the respective observation 
times (e.g. '0.0' for an observation at time zero). These allow for the time 
step to be inferred automatically (\code{\link{inferTimeStep}}), for the 
phenotype data to be padded in gaps of the time series 
(\code{\link{padTimeSeries}}), and for the phenotype values to 
be interpolated in time series gaps (\code{\link{interpTimeSeries}}).
}

