% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{class}
\name{CrossInfo-class}
\alias{CrossInfo}
\alias{CrossInfo-class}
\title{An S4 class to hold yeast cross information.}
\description{
A CrossInfo object holds yeast cross information for a specific \code{cross}
object. The contents of its slots should match its corresponding object.
}
\section{Slots}{

\describe{
\item{\code{seq}}{A character vector of sequence identifiers, with the name
of each element being the name of the given sequence. These must be
unique.}

\item{\code{pheno}}{A vector of cross phenotypes, with the name of each element being 
the syntactically valid name of the phenotype ID (as produced by the 
function \code{make.names}). Phenotype IDs must be unique.}

\item{\code{markers}}{A \code{data.frame} with one required column - 'marker' - 
containing the marker ID, and one optional column: 'seq'. If present, the 
'seq' column contains the sequence corresponding to the given marker. 
Each row name contains the syntactically valid name of the marker ID (as 
produced by the function \code{make.names}). Marker IDs must be unique.}

\item{\code{samples}}{A \code{data.frame} with one required column - 'sample.index' - 
that contains indices of the samples in the given cross dataset. Other optional 
columns can be present, including 'sample.id', 'sample.name', 'strain.index' 
and 'tetrad.index'. 

The 'sample.id' column contains sample IDs, which can be any valid item ID.
If this column is present, another column called 'sample.name' will
contain the syntactically valid name of the sample ID (as produced by the 
function \code{make.names}). Duplicate sample IDs are permissible, but only
if referring to replicate samples of the same strain. Different strains can 
have different numbers of replicates, but samples from a given strain must 
be in consecutive rows.
 
The 'strain.index' column, if present, contains strain indices denoting 
which strain each sample is taken from. This is only necessary if the cross
contains sample replicates. Replicate samples of a given strain must have 
the same genotype data, and their sample IDs must match, if present. 

If the samples are tetradic, this can also include a 'tetrad.index' column 
containing the tetrad index of each sample, indicating which tetrad that 
sample is from. Tetrads are assumed to be formed of each group of four 
consecutive strains, unless sample IDs are present that indicate tetrad 
membership. Samples can be missing from a tetrad, but in such cases the 
sample IDs should be such that it is possible to assign each sample to
a tetrad. In any case, samples from the same tetrad must be in consecutive 
rows of the cross data.}

\item{\code{alleles}}{A vector of cross allele symbols.}

\item{\code{genotypes}}{A vector of cross genotype symbols.}

\item{\code{crosstype}}{Cross type.}
}}
\keyword{internal}

