% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\docType{package}
\name{CSV Utilities}
\alias{CSV Utilities}
\alias{CSV Utilities-package}
\title{CSV input/output utilities.}
\description{
Functions for validated input and output of \pkg{R/qtl} data in CSV format.
These include functions for reading (\code{\link{readCrossCSV}}) or writing 
(\code{\link{writeCrossCSV}}) an \pkg{R/qtl} \code{cross} object. There are 
also functions for reading (\code{\link{readMapCSV}}) or writing 
(\code{\link{writeMapCSV}}) an \pkg{R/qtl} \code{map} object.
}

