% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{deleteColumn}
\alias{deleteColumn}
\title{Delete column from object.}
\usage{
deleteColumn(x, col.index = NULL, col.name = NULL)
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}.}

\item{col.index}{Column index of column to remove.}

\item{col.name}{Column name of column to remove.}
}
\value{
Input object with column removed.
}
\description{
Delete column from object.
}
\keyword{internal}

