% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dispatchFromClassS3}
\alias{dispatchFromClassS3}
\title{Dispatch method with respect to the given class vector.}
\usage{
dispatchFromClassS3(generic, class.vector, package)
}
\arguments{
\item{generic}{Generic function name.}

\item{class.vector}{Class vector of an R object.}

\item{package}{Package from which available methods are taken.}
}
\value{
Method suited to the specified class.
}
\description{
Dispatch method with respect to the given class vector.
}
\keyword{internal}
