% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ellipt}
\alias{ellipt}
\title{Ellipt string to the given length.}
\usage{
ellipt(s, n, left = FALSE, right = FALSE)
}
\arguments{
\item{s}{String to ellipt.}

\item{n}{Maximum number of characters in the output string. If the string is
ellipted, this includes the number of characters in the ellipses themselves.}

\item{left}{Indicates if the string should be
ellipted at the start ('left') of the string.}

\item{right}{Indicates if the string should be
ellipted at the end ('right') of the string.}
}
\value{
Input string ellipted to the specified length.
}
\description{
Ellipt string to the given length.
}
\keyword{internal}

