% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeGenotypes.R
\name{encodeGenotypes}
\alias{encodeGenotypes}
\alias{encodeGenotypes.character}
\alias{encodeGenotypes.data.frame}
\alias{encodeGenotypes.matrix}
\title{Encode genotype strings as genotype integers.}
\usage{
encodeGenotypes(x, genotypes)

\method{encodeGenotypes}{character}(x, genotypes)

\method{encodeGenotypes}{data.frame}(x, genotypes)

\method{encodeGenotypes}{matrix}(x, genotypes)
}
\arguments{
\item{x}{Object containing genotype strings.}

\item{genotypes}{Sorted character vector of genotype symbols.
Input genotype strings should be elements of this vector.}
}
\value{
Object of the same class and dimensions as input object, but with
genotype strings encoded as genotype integers. Missing values (\code{'-'})
are replaced with the character \code{NA} value.
}
\description{
Encode genotype strings as genotype integers.
}
\keyword{internal}
