% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlintervals.R
\name{estPhysicalPositions}
\alias{estPhysicalPositions}
\title{Add estimated physical positions to QTL intervals.}
\usage{
estPhysicalPositions(qtl.intervals, map.key = NULL)
}
\arguments{
\item{qtl.intervals}{A \code{qtlintervals} object.}

\item{map.key}{An optional \code{mapkey} function for converting genetic map
positions to physical map positions. If none is specified, the default
\code{mapkey} for the current genome is used to estimate physical positions.}
}
\value{
A copy of the input \code{qtlintervals} object, with an additional
column - \code{'pos (bp)'} - containing physical map positions of the QTL
intervals, as estimated from their genetic map positions.
}
\description{
Add estimated physical positions to QTL intervals.
}

