% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{extractMarkers}
\alias{extractMarkers}
\title{Extract marker loci.}
\usage{
extractMarkers(x)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
Input object with pseudomarkers removed.
}
\description{
Extract marker loci.
}
\seealso{
Other map utilities: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{inMapOrder}}, \code{\link{inferMapStep}},
  \code{\link{mapframe}}, \code{\link{orderMap}},
  \code{\link{setMapUnit}}
}

