% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{extractPseudomarkers}
\alias{extractPseudomarkers}
\title{Extract pseudomarker loci.}
\usage{
extractPseudomarkers(x)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
Input object with only pseudomarkers remaining.
}
\description{
Extract pseudomarker loci.
}
\seealso{
Other map utility functions: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}}, \code{\link{findFlanking}},
  \code{\link{findLoci}}, \code{\link{findMarkers}},
  \code{\link{getMapSteps}}, \code{\link{getMapUnit}},
  \code{\link{gmapframe}}, \code{\link{hasMapCSV}},
  \code{\link{inMapOrder}}, \code{\link{inferMapStep}},
  \code{\link{mapframe}}, \code{\link{orderMap}},
  \code{\link{readMapCSV}}, \code{\link{readMapHDF5}},
  \code{\link{readMapframeCSV}}, \code{\link{setMapUnit}},
  \code{\link{writeMapCSV}}, \code{\link{writeMapHDF5}},
  \code{\link{writeMapframeCSV}}
}

