% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5Stack-class.R
\docType{methods}
\name{fileID}
\alias{fileID}
\alias{fileID,H5Stack-method}
\title{Get HDF5 object ID of HDF5 file.}
\usage{
fileID(h5stack)

\S4method{fileID}{H5Stack}(h5stack)
}
\arguments{
\item{h5stack}{A \code{H5Stack} object.}
}
\value{
HDF5 object ID of HDF5 file. Returns \code{NULL} if stack is empty.
}
\description{
Get HDF5 object ID of HDF5 file.
}
\keyword{internal}
