% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{findFlanking}
\alias{findFlanking}
\alias{findFlanking.map}
\alias{findFlanking.mapframe}
\alias{findFlanking.scanone}
\title{Find map loci flanking map positions.}
\usage{
findFlanking(x, loc, expandtomarkers = FALSE)

\method{findFlanking}{map}(x, loc, expandtomarkers = FALSE)

\method{findFlanking}{mapframe}(x, loc, expandtomarkers = FALSE)

\method{findFlanking}{scanone}(x, loc, expandtomarkers = FALSE)
}
\arguments{
\item{x}{Object containing map data.}

\item{loc}{Locus \code{mapframe} specifying map positions.}

\item{expandtomarkers}{Expand the flanking interval to the nearest flanking
markers, or to the respective terminal loci.}
}
\value{
A \code{mapframe} containing two loci marking respectively the
lower and upper limit of the flanking interval.
}
\description{
Find map loci flanking map positions.
}
\seealso{
Other map utility functions: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findLoci}}, \code{\link{findMarkers}},
  \code{\link{getMapSteps}}, \code{\link{getMapUnit}},
  \code{\link{gmapframe}}, \code{\link{hasMapCSV}},
  \code{\link{inMapOrder}}, \code{\link{inferMapStep}},
  \code{\link{mapframe}}, \code{\link{mapsEqual}},
  \code{\link{orderMap}}, \code{\link{readMapCSV}},
  \code{\link{readMapHDF5}}, \code{\link{readMapframeCSV}},
  \code{\link{setMapUnit}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapHDF5}},
  \code{\link{writeMapframeCSV}}
}
