% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{findFlankingRowIndices}
\alias{findFlankingRowIndices}
\title{Find row index of flanking \code{mapframe} loci.}
\usage{
findFlankingRowIndices(x, loc, expandtomarkers = FALSE)
}
\arguments{
\item{x}{A \code{mapframe} or \code{scanone} object.}

\item{loc}{Locus \code{mapframe} specifying map positions.}

\item{expandtomarkers}{Expand the flanking interval to the nearest flanking
markers, or to the respective terminal loci.}
}
\value{
Vector containing the row indices of flanking loci.
}
\description{
Find row index of flanking \code{mapframe} loci.
}
\keyword{internal}

