% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{findMarkers}
\alias{findMarkers}
\title{Find closest markers.}
\usage{
findMarkers(x, loc)
}
\arguments{
\item{x}{Object containing map data.}

\item{loc}{Locus \code{mapframe} specifying map positions.}
}
\value{
A \code{mapframe} whose rows each contain the closest marker to the
map position specified in the corresponding row of the locus \code{mapframe}.
}
\description{
Find closest markers.
}
\seealso{
Other map utilities: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{getMapSteps}}, \code{\link{getMapUnit}},
  \code{\link{gmapframe}}, \code{\link{inMapOrder}},
  \code{\link{inferMapStep}}, \code{\link{mapframe}},
  \code{\link{orderMap}}, \code{\link{setMapUnit}}
}

