% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{formatChr}
\alias{formatChr}
\title{Format chromosome labels.}
\usage{
formatChr(x, prefix = c("", "c", "chr"), use.roman = TRUE)
}
\arguments{
\item{x}{Vector of chromosome labels.}

\item{prefix}{Chromosome prefix.}

\item{use.roman}{Use the Roman numeral form of the chromosome.}
}
\value{
Vector of formatted chromosome labels.
}
\description{
Format chromosome labels.
}
\keyword{internal}

