% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{genomeOpt}
\alias{genomeOpt}
\title{Set/get genome option.}
\usage{
genomeOpt(value)
}
\arguments{
\item{value}{New value of genome option.}
}
\value{
Current value of genome option.
}
\description{
This function returns the current value of the \pkg{shmootl} genome option.
If a genome is specified, the \pkg{shmootl} genome option is updated with the
given value. Call the function \code{\link{listGenomes}} to see available
genomes.
}
\seealso{
Other genome functions: \code{\link{listGenomes}}

Other package option functions: \code{\link{mapkeyOpt}}
}

