% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getAlleles}
\alias{getAlleles}
\alias{getAlleles,CrossInfo-method}
\title{Get allele symbols.}
\usage{
getAlleles(cross.info)

\S4method{getAlleles}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
Character vector of allele symbols.
}
\description{
Get allele symbols.
}
\keyword{internal}

