% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getAnalysisTitle}
\alias{getAnalysisTitle}
\title{Get analysis title for the given \pkg{shmootl} analysis pipeline.}
\usage{
getAnalysisTitle(pipeline)
}
\arguments{
\item{pipeline}{Character vector of \pkg{shmootl} analysis pipeline titles.}
}
\value{
Character vector in which each element contains the title of the
\pkg{shmootl} analysis pipeline in the corresponding element of the input
vector.
}
\description{
Get analysis title for the given \pkg{shmootl} analysis pipeline.
}
\keyword{internal}

