% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getCoercionFromClassS3}
\alias{getCoercionFromClassS3}
\title{Get coercion function for the given class.}
\usage{
getCoercionFromClassS3(class.vector)
}
\arguments{
\item{class.vector}{Class vector of an R object.}
}
\value{
Coercion function.
}
\description{
Get coercion function for the given class.
}
\keyword{internal}

