% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getColMask}
\alias{getColMask}
\title{Get logical mask of object columns.}
\usage{
getColMask(x, requested = NULL)
}
\arguments{
\item{x}{An object with columns.}

\item{requested}{A character vector of column names, a logical vector of
length equal to the number of object columns, or a numeric vector of column
indices of the input object. If this parameter is not specified, the returned
logical mask is \code{TRUE} for all columns.}
}
\value{
Logical mask of object columns.
}
\description{
Get logical mask of the columns of an object, as constrained by the
\code{requested} parameter. If using this function without \code{requested}
constraints, consider using the command \code{rep(TRUE, ncol(x))} instead.
}
\keyword{internal}

