% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{getDatColIndices}
\alias{getDatColIndices}
\title{Get data column indices.}
\usage{
getDatColIndices(x, datcolumns = NULL, strict = FALSE)
}
\arguments{
\item{x}{A \code{mapframe} or equivalent \code{data.frame}.}

\item{datcolumns}{Optional parameter indicating which data columns to
consider. This must be a character vector of data column names, a logical
vector with length equal to the number of data columns, or a numeric vector
of indices \emph{with respect to the set of data columns}. If no data columns
are specified, all are considered.}

\item{strict}{Option indicating that the specified data columns must
be in the same order as they are in the \code{mapframe} object.}
}
\value{
Vector of data column indices.
}
\description{
Get data column indices for a \code{mapframe} or equivalent \code{data.frame}.
Data columns are taken to be any columns that aren't a sequence or position
column.
}
\keyword{internal}

