% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{getExcelDigestInfo}
\alias{getExcelDigestInfo}
\title{Get info for digest Excel file.}
\usage{
getExcelDigestInfo()
}
\value{
A list containing descriptions
and headings for Excel digest worksheets.
}
\description{
Get info for digest Excel file.
}
\keyword{internal}

