% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getFileExtension}
\alias{getFileExtension}
\title{Get file extensions.}
\usage{
getFileExtension(x)
}
\arguments{
\item{x}{Character vector of file paths.}
}
\value{
Vector in which each element contains the file extension of the 
corresponding element of the input vector, or an empty string if no file
extension was found.
}
\description{
Get file extensions.
}
\keyword{internal}

