% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getFilePrefix}
\alias{getFilePrefix}
\title{Get file prefixes.}
\usage{
getFilePrefix(x)
}
\arguments{
\item{x}{Character vector of file paths.}
}
\value{
Vector in which each element contains the file prefix of the 
corresponding element of the input vector, or the input file path if 
no file extension was found.
}
\description{
Get file prefixes.
}
\keyword{internal}

