% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{getFlankingPhenoColIndices}
\alias{getFlankingPhenoColIndices}
\title{Get flanking phenotype column indices.}
\usage{
getFlankingPhenoColIndices(cross, i)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{i}{Column index for which the flanking phenotype column indices 
should be returned.}
}
\value{
Flanking phenotype column indices. If the specified index refers to
the first or last column, an NA value is returned for the lower or upper 
flanking index, respectively.
}
\description{
Get the indices of the phenotype columns in the \code{cross} phenotype data 
frame that are flanking the specified index, excluding special columns such 
as the sample ID column.
}
\keyword{internal}

