% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{getGenomes}
\alias{getGenomes}
\title{Get \pkg{shmootl} genome names.}
\usage{
getGenomes()
}
\value{
Character vector of \pkg{shmootl} genome names.
}
\description{
Get \pkg{shmootl} genome names.
}
\keyword{internal}

