% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getGenotypes}
\alias{getGenotypes}
\alias{getGenotypes,CrossInfo-method}
\title{Get genotype symbols.}
\usage{
getGenotypes(cross.info)

\S4method{getGenotypes}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
Character vector of genotype symbols.
}
\description{
Get genotype symbols.
}
\keyword{internal}

