% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{getGroupMemberNamesHDF5}
\alias{getGroupMemberNamesHDF5}
\title{Get HDF5 group member names.}
\usage{
getGroupMemberNamesHDF5(infile, h5name)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{h5name}{HDF5 group name.}
}
\value{
Vector of group member names. Returns \code{NULL}
if group is not present or has no members.
}
\description{
Get HDF5 group member names.
}
\keyword{internal}

