% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getIndices}
\alias{getIndices}
\title{Get indices of object.}
\usage{
getIndices(x)
}
\arguments{
\item{x}{An object with length.}
}
\value{
Integer vector of all object indices. Returns an empty integer vector
if the object does not have length.
}
\description{
Get indices of object.
}
\keyword{internal}

