% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getLodColIndex}
\alias{getLodColIndex}
\title{Get LOD column index.}
\usage{
getLodColIndex(x, lodcolumn = NULL)
}
\arguments{
\item{x}{A \code{scanone} or equivalent object,
or any object with LOD-column-associated elements.}

\item{lodcolumn}{In a \code{scanone} or equivalent object, this
parameter indicates which LOD column to consider. In other objects
with LOD-column-associated elements, this parameter indicates which
LOD-column-associated element to consider. This must be either a LOD
column name or an index \emph{with respect to the set of LOD columns}.
If no LOD column is specified and one such column is found, that column
is used by default; otherwise a LOD column must be specified.}
}
\value{
LOD column index.
}
\description{
Get LOD column index.
}
\keyword{internal}

