% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getLodColIndices}
\alias{getLodColIndices}
\alias{getLodColIndices.mapframe}
\alias{getLodColIndices.scanone}
\alias{getLodColIndices.scanonebins}
\alias{getLodColIndices.scanoneperm}
\alias{getLodColIndices.scantwo}
\alias{getLodColIndices.scantwoperm}
\alias{getLodColIndices.summary.scanonebins}
\alias{getLodColIndices.summary.scanoneperm}
\alias{getLodColIndices.summary.scantwoperm}
\title{Get LOD column indices.}
\usage{
getLodColIndices(x, lodcolumns = NULL, strict = FALSE)

\method{getLodColIndices}{mapframe}(x, lodcolumns = NULL, strict = FALSE)

\method{getLodColIndices}{scanone}(x, lodcolumns = NULL, strict = FALSE)

\method{getLodColIndices}{scanonebins}(x, lodcolumns = NULL, strict = FALSE)

\method{getLodColIndices}{scanoneperm}(x, lodcolumns = NULL, strict = FALSE)

\method{getLodColIndices}{scantwo}(x, lodcolumns = NULL, strict = FALSE)

\method{getLodColIndices}{scantwoperm}(x, lodcolumns = NULL, strict = FALSE)

\method{getLodColIndices}{summary.scanoneperm}(x, lodcolumns = NULL,
  strict = FALSE)

\method{getLodColIndices}{summary.scanonebins}(x, lodcolumns = NULL,
  strict = FALSE)

\method{getLodColIndices}{summary.scantwoperm}(x, lodcolumns = NULL,
  strict = FALSE)
}
\arguments{
\item{x}{A \code{scanone} or equivalent object,
or any object with LOD-column-associated elements.}

\item{lodcolumns}{In a \code{scanone} or equivalent object, this
parameter indicates which LOD columns to consider. In other objects
with LOD-column-associated elements, this parameter indicates which
of those LOD-column-associated elements to consider. The specified
LOD columns must be a character vector of LOD column names, a logical
vector with length equal to the number of LOD columns, or a numeric
vector of indices \emph{with respect to the set of LOD columns}. If
no LOD columns are specified, all are used.}

\item{strict}{Option indicating that the LOD columns, if specified, must
be in the same order as the corresponding elements of the input object.}
}
\value{
Vector of LOD column indices.
}
\description{
Get LOD column indices.
}
\keyword{internal}

