% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLODMatrix.R
\name{getLODMatrix}
\alias{getLODMatrix}
\title{Get LOD matrix.}
\usage{
getLODMatrix(x, chr = NULL, incl.markers = FALSE, lodcolumn = NULL,
  lower = c("full", "cond-int", "int", "cond-add", "add"), upper = c("add",
  "cond-add", "int", "cond-int", "full"))
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{scantwo} object.}

\item{chr}{Vector indicating which sequences to include in the LOD matrix.}

\item{incl.markers}{When \code{scantwo} has been run with a fixed non-zero
step width, this indicates if markers (as distinct from pseudomarkers) should
be included in the returned LOD matrix.}

\item{lodcolumn}{In a \code{scanone} or equivalent object, this
parameter indicates which LOD column to consider. In other objects
with LOD-column-associated elements, this parameter indicates which
LOD-column-associated element to consider. This must be either a LOD
column name or an index \emph{with respect to the set of LOD columns}.
If no LOD column is specified and one such column is found, that column
is used by default; otherwise a LOD column must be specified.}

\item{lower}{Type of LOD score to put in the lower triangle of the LOD matrix.}

\item{upper}{Type of LOD score to put in the upper triangle of the LOD matrix.}
}
\value{
A LOD matrix taken from the input \code{scantwo} object, containing
LOD scores for a single phenotype specified by the \code{lodcolumn} parameter.
}
\description{
Given a \code{scantwo} object, this function extracts a LOD matrix, in which
the upper and lower triangle contain LOD values of a type specified by the
parameters \code{upper} and \code{lower}, respectively. The method used to
obtain derived LOD values is that of \pkg{R/qtl} function \code{plot.scantwo}
(Broman \emph{et al.} 2003).
}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})
}
\seealso{
Function \code{plot.scantwo} in the
\href{http://www.rqtl.org/manual/qtl-manual.pdf}{R/qtl manual}.
}

