% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{getMapNamesHDF5}
\alias{getMapNamesHDF5}
\title{Get names of maps in HDF5 file.}
\usage{
getMapNamesHDF5(infile)
}
\arguments{
\item{infile}{An input HDF5 file.}
}
\value{
Vector of map names. Returns \code{NULL}
if there are no maps present.
}
\description{
Get names of maps in HDF5 file.
}
\keyword{internal}

