% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{getMapSteps}
\alias{getMapSteps}
\alias{getMapSteps.map}
\alias{getMapSteps.mapframe}
\alias{getMapSteps.scanone}
\title{Get steps between loci.}
\usage{
getMapSteps(x)

\method{getMapSteps}{mapframe}(x)

\method{getMapSteps}{map}(x)

\method{getMapSteps}{scanone}(x)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
List of vectors, each giving the steps between markers on a given 
sequence.
}
\description{
Get steps between loci.
}
\seealso{
Other map utilities: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapUnit}},
  \code{\link{gmapframe}}, \code{\link{inMapOrder}},
  \code{\link{inferMapStep}}, \code{\link{mapframe}},
  \code{\link{orderMap}}, \code{\link{setMapUnit}}
}

