% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getMarkerNames}
\alias{getMarkerNames}
\alias{getMarkerNames,CrossInfo-method}
\title{Get marker names.}
\usage{
getMarkerNames(cross.info, markers = NULL)

\S4method{getMarkerNames}{CrossInfo}(cross.info, markers = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{markers}{Vector indicating markers to consider. This may be a logical 
vector of markers, an integer vector of marker indices, or a character vector
that contains marker IDs or their equivalent syntactically valid name. If 
none are specified, results are returned for all markers.}
}
\value{
Character vector of marker names.
}
\description{
Get marker names.
}
\keyword{internal}

