% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getMarkerSeqs}
\alias{getMarkerSeqs}
\alias{getMarkerSeqs,CrossInfo-method}
\title{Get sequences by marker.}
\usage{
getMarkerSeqs(cross.info, markers = NULL)

\S4method{getMarkerSeqs}{CrossInfo}(cross.info, markers = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{markers}{Vector indicating markers to consider. This may be a logical 
vector of markers, an integer vector of marker indices, or a character vector
that contains marker IDs or their equivalent syntactically valid name. If 
none are specified, results are returned for all markers.}
}
\value{
Vector of sequences corresponding to the given markers. Returns 
\code{NULL} if no sequence-marker information is available.
}
\description{
Get sequences by marker.
}
\keyword{internal}

