% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getMask}
\alias{getMask}
\title{Get logical mask of object elements.}
\usage{
getMask(x, requested = NULL)
}
\arguments{
\item{x}{An object that is subsettable by a logical mask.}

\item{requested}{A character vector of names, a logical vector of the same
length as the input object, or a numeric vector containing indices of the
input object. If this parameter is not specified, the returned logical mask
is \code{TRUE} for all elements.}
}
\value{
Logical mask of object elements.
}
\description{
Get logical mask of the elements of an object, as constrained by the
\code{requested} parameter. If using this function without \code{requested}
constraints, consider using the command \code{rep(TRUE, length(x))} instead.
}
\keyword{internal}

