% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{getMetadataCSV}
\alias{getMetadataCSV}
\title{Get parameters of \pkg{R/qtl} input data.}
\usage{
getMetadataCSV(x)
}
\arguments{
\item{x}{An \pkg{R/qtl} input CSV file or
an \pkg{R/qtl} input \code{data.frame}.}
}
\value{
A list containing parameters of \pkg{R/qtl} input data.
}
\description{
Given the path to an \pkg{R/qtl} input CSV file, or a \code{data.frame}
loaded from such a file, this function determines the parameters of that
data (e.g. phenotype columns).
}
\details{
The returned list includes information on the following parameters:

\itemize{
 \item{\code{pheno.cols}}{Indices of phenotype columns,
 or \code{NULL} if none found.}
 \item{\code{id.col}}{Index of ID column, or \code{NA} if not found.}
 \item{\code{geno.cols}}{Indices of genotype columns,
 or \code{NULL} if none found.}
 \item{\code{dat.rows}}{Row indices of data.}
 \item{\code{map.present}}{\code{TRUE} if the data contains a map;
 \code{FALSE} otherwise.}
 \item{\code{class}}{Expected class of the input data.
 This is \code{'unknown'} by default.}
}

This information can then guide further processing of the input data.
}
\keyword{internal}
