% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getMissingValueFromClassS3}
\alias{getMissingValueFromClassS3}
\title{Get missing value for the given class.}
\usage{
getMissingValueFromClassS3(class.vector)
}
\arguments{
\item{class.vector}{Class vector of an R object.}
}
\value{
Missing value.
}
\description{
Get missing value for the given class.
}
\keyword{internal}

