% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getNumMarkers}
\alias{getNumMarkers}
\alias{getNumMarkers,CrossInfo-method}
\title{Get the number of markers.}
\usage{
getNumMarkers(cross.info)

\S4method{getNumMarkers}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
Number of markers.
}
\description{
Get the number of markers.
}
\keyword{internal}

