% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getNumSamples}
\alias{getNumSamples}
\alias{getNumSamples,CrossInfo-method}
\title{Get the number of samples.}
\usage{
getNumSamples(cross.info)

\S4method{getNumSamples}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
Number of samples.
}
\description{
Get the number of samples.
}
\keyword{internal}

