% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getNumSeqs}
\alias{getNumSeqs}
\alias{getNumSeqs,CrossInfo-method}
\title{Get the number of sequences.}
\usage{
getNumSeqs(cross.info)

\S4method{getNumSeqs}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
Number of sequences.
}
\description{
Get the number of sequences.
}
\keyword{internal}

