% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{getObjectNamesHDF5}
\alias{getObjectNamesHDF5}
\title{Get names of HDF5 objects from HDF5 file and name.}
\usage{
getObjectNamesHDF5(infile, h5name = NULL, include.requested = FALSE,
  relative = FALSE, recursive = FALSE)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{h5name}{Requested HDF5 object name. By default, results
are returned with respect to the HDF5 root object.}

\item{include.requested}{Include the name of the
requested HDF5 object among returned object names.}

\item{relative}{Option indicating that the returned HDF5 object
names should be given relative to the requested HDF5 object.}

\item{recursive}{Option indicating whether
to get HDF5 object names recursively.}
}
\value{
Character vector of names of HDF5 objects
that lie below the requested HDF5 object.
}
\description{
Get names of HDF5 objects from HDF5 file and name.
}
\keyword{internal}

