% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getPhenotypeIndices}
\alias{getPhenotypeIndices}
\alias{getPhenotypeIndices,CrossInfo-method}
\title{Get phenotype indices.}
\usage{
getPhenotypeIndices(cross.info, phenotypes = NULL)

\S4method{getPhenotypeIndices}{CrossInfo}(cross.info, phenotypes = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{phenotypes}{Vector indicating phenotypes to consider. This may be a 
logical vector of phenotypes, an integer vector of phenotype indices, or 
a character vector that contains phenotype IDs or their equivalent 
syntactically valid name. If none are specified, results are returned 
for all phenotypes.}
}
\value{
Integer vector of phenotype indices.
}
\description{
Get phenotype indices.
}
\keyword{internal}

