% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getPhenotypes.summary.scantwoperm}
\alias{getPhenotypes.summary.scantwoperm}
\title{Get phenotype names.}
\usage{
getPhenotypes.summary.scantwoperm(x)
}
\arguments{
\item{x}{Object that may contain phenotype names.}
}
\value{
Character vector of phenotype names. Returns
\code{NULL} if the object does not contain phenotype names.
}
\description{
Get phenotype names.
}
\keyword{internal}

