% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{getPhenotypesHDF5}
\alias{getPhenotypesHDF5}
\title{Get names of phenotypes in HDF5 file.}
\usage{
getPhenotypesHDF5(infile)
}
\arguments{
\item{infile}{An input HDF5 file.}
}
\value{
Vector of phenotype names. Returns \code{NULL}
if there are no phenotypes present.
}
\description{
Get names of phenotypes in HDF5 file.
}
\keyword{internal}

