% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPipelineArgparser}
\alias{getPipelineArgparser}
\title{Get argument parser for \pkg{shmootl} pipeline.}
\usage{
getPipelineArgparser(pipeline)
}
\arguments{
\item{pipeline}{Name of a \pkg{shmootl} pipeline.}
}
\value{
Argument parser for the specified \pkg{shmootl} pipeline.
}
\description{
Get argument parser for \pkg{shmootl} pipeline.
}
\keyword{internal}

