% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPipelineFunction}
\alias{getPipelineFunction}
\title{Get \pkg{shmootl} pipeline function.}
\usage{
getPipelineFunction(pipeline)
}
\arguments{
\item{pipeline}{Name of a \pkg{shmootl} pipeline.}
}
\value{
Function for the given \pkg{shmootl} pipeline.
}
\description{
Get \pkg{shmootl} pipeline function.
}
\keyword{internal}

