% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPipelineFunctionName}
\alias{getPipelineFunctionName}
\title{Get function name for the given \pkg{shmootl} pipeline.}
\usage{
getPipelineFunctionName(pipeline)
}
\arguments{
\item{pipeline}{Character vector of \pkg{shmootl} pipeline names.}
}
\value{
Character vector in which each element contains the name
of the \pkg{shmootl} pipeline function called by the pipeline
named in the corresponding element of the input vector.
}
\description{
Get function name for the given \pkg{shmootl} pipeline.
}
\keyword{internal}
