% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPipelineInfo}
\alias{getPipelineInfo}
\title{Get info for \pkg{shmootl} pipeline.}
\usage{
getPipelineInfo(pipeline)
}
\arguments{
\item{pipeline}{Name of \pkg{shmootl} pipeline.}
}
\value{
A \code{list} containing info for the given \pkg{shmootl} pipeline.
}
\description{
This function interrogates the function definition and documentation for the 
specified pipeline and returns the resulting information as a list. The 
elements of this list are as follows:
}
\details{
\itemize{
\item{\code{'command'}} {Shell command used to call pipeline.}
\item{\code{'title'}} {Pipeline function title.}
\item{\code{'description'}} {Pipeline function description. This is not 
returned if it duplicates the function title.}
\item{\code{'details'}} {Pipeline function details.}
\item{\code{'params'}} {A \code{data.frame} containing pipeline function 
parameter info. If the function has no parameters, this takes a \code{NULL} 
value.}
}
 
Elements \code{'command'} and \code{'title'} are guaranteed to contain their
respective values. The remaining elements are only returned if corresponding 
documentation is available; otherwise, they are assigned a \code{NULL} value.
}
\keyword{internal}

