% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPipelineName}
\alias{getPipelineName}
\title{Get \pkg{shmootl} pipeline name for the given function.}
\usage{
getPipelineName(pipe.func.name)
}
\arguments{
\item{pipe.func.name}{Character vector of
\pkg{shmootl} pipeline function names.}
}
\value{
Character vector in which each element contains the name
of the \pkg{shmootl} pipeline implemented by the function named
in the corresponding element of the input vector.
}
\description{
Get \pkg{shmootl} pipeline name for the given function.
}
\keyword{internal}
