% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPipelines}
\alias{getPipelines}
\title{Get \pkg{shmootl} pipeline names.}
\usage{
getPipelines()
}
\value{
Character vector of \pkg{shmootl} pipeline names.
}
\description{
Get \pkg{shmootl} pipeline names.
}

