% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPipelineTitle}
\alias{getPipelineTitle}
\title{Get title of the given \pkg{shmootl} pipeline.}
\usage{
getPipelineTitle(pipeline)
}
\arguments{
\item{pipeline}{Character vector of \pkg{shmootl} pipeline names.}
}
\value{
Character vector in which each element contains the title of the
\pkg{shmootl} pipeline in the corresponding element of the input vector.
}
\description{
Get title of the given \pkg{shmootl} pipeline.
}
\keyword{internal}

