% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPkgAnalysisNames}
\alias{getPkgAnalysisNames}
\title{Get \pkg{shmootl} analysis names.}
\usage{
getPkgAnalysisNames()
}
\value{
Character vector of \pkg{shmootl} analysis names.
}
\description{
Get \pkg{shmootl} analysis names.
}
\keyword{internal}

