% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPkgAnalysisPipelineNames}
\alias{getPkgAnalysisPipelineNames}
\title{Get \pkg{shmootl} analysis pipeline names.}
\usage{
getPkgAnalysisPipelineNames()
}
\value{
Character vector of \pkg{shmootl} analysis pipeline names.
}
\description{
Get \pkg{shmootl} analysis pipeline names.
}
\keyword{internal}

