% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPkgPipelineDocs}
\alias{getPkgPipelineDocs}
\title{Get \pkg{shmootl} pipeline docs.}
\usage{
getPkgPipelineDocs(pipelines = NULL)
}
\arguments{
\item{pipelines}{Pipelines for which pipeline docs should be retrieved.
If no pipeline is specified, all pipeline docs are returned.}
}
\value{
Named list of \pkg{shmootl} pipeline docs.
}
\description{
Package pipeline docs are taken from the loaded \pkg{shmootl} package docs,
if available; otherwise they are taken from installed \pkg{shmootl} manual
pages.
}
\keyword{internal}
