% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPkgPipelineFunctionNames}
\alias{getPkgPipelineFunctionNames}
\title{Get \pkg{shmootl} pipeline function names.}
\usage{
getPkgPipelineFunctionNames(group = NULL)
}
\arguments{
\item{group}{Pipeline group for which pipeline function names should be
retrieved. If no pipeline group is specified, all pipeline function names
are returned.}
}
\value{
Character vector of \pkg{shmootl} pipeline function names.
}
\description{
Get \pkg{shmootl} pipeline function names.
}

