% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPkgPipelineInfo}
\alias{getPkgPipelineInfo}
\title{Get info for \pkg{shmootl} pipelines.}
\usage{
getPkgPipelineInfo(pipelines = NULL)
}
\arguments{
\item{pipelines}{Names of one or more \pkg{shmootl} pipelines for which info
should be retrieved. If none are specified, pipeline info is returned for all
pipelines.}
}
\value{
A \code{list} of \code{list} objects, each containing info for a
specific \pkg{shmootl} pipeline.
}
\description{
This function interrogates the function definition and documentation for the 
specified pipelines and returns the resulting information as a \code{list}.
Each element of this \code{list} is another \code{list} containing
information for a specific pipeline, the elements of which are as follows:
}
\details{
\itemize{
\item{\code{'command'}} {Shell command used to call pipeline.}
\item{\code{'title'}} {Pipeline function title.}
\item{\code{'group'}} {Category to which the pipeline belongs.}
\item{\code{'description'}} {Pipeline function description. This is not 
returned if it duplicates the function title.}
\item{\code{'details'}} {Pipeline function details.}
\item{\code{'params'}} {A \code{data.frame} containing pipeline function 
parameter info. If the function has no parameters, this takes a \code{NULL} 
value.}
}
 
Elements \code{'command'}, \code{'title'}, and \code{'group'} are guaranteed
to contain their respective values. The remaining elements are only returned
if corresponding documentation is available; otherwise, they are assigned a
\code{NULL} value.
}
\keyword{internal}

