% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPkgPipelineUsage}
\alias{getPkgPipelineUsage}
\title{Get default usage string for \pkg{shmootl} pipelines.}
\usage{
getPkgPipelineUsage()
}
\value{
Character vector containing the \pkg{shmootl} pipeline default usage 
string.
}
\description{
Get default usage string for \pkg{shmootl} pipelines.
}
\keyword{internal}
