% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{getPosColDataMapUnit}
\alias{getPosColDataMapUnit}
\alias{getPosColDataMapUnit.character}
\alias{getPosColDataMapUnit.data.frame}
\alias{getPosColDataMapUnit.numeric}
\title{Get map unit from position column values.}
\usage{
getPosColDataMapUnit(x)

\method{getPosColDataMapUnit}{character}(x)

\method{getPosColDataMapUnit}{data.frame}(x)

\method{getPosColDataMapUnit}{numeric}(x)
}
\arguments{
\item{x}{A \code{mapframe} or equivalent \code{data.frame},
or a vector of map positions.}
}
\value{
Map unit indicated by the position column data. Returns
\code{NA} if map unit information not found in position column data.
}
\description{
Get map unit from position column values.
}
\section{Map Units}{


A \code{map} or \code{mapframe} must express positions in either genetic or
physical map units. Genetic map positions must be in centiMorgans (cM), while
physical map positions must be in base pairs (bp), kilobases (kb), or
Megabases (Mb).
}

\keyword{internal}
