% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{getPosColIndex}
\alias{getPosColIndex}
\title{Get \code{mapframe} position column index.}
\usage{
getPosColIndex(x)
}
\arguments{
\item{x}{A \code{mapframe} or equivalent \code{data.frame}.}
}
\value{
Position column index.
}
\description{
Get position column index for a \code{mapframe} or equivalent \code{data.frame}.
The position column is taken to be the leftmost column whose heading contains
the word 'pos'. This can also be in uppercase (i.e. 'POS'), but can't be part
of a larger word, such as 'position'. An error is raised if a position column
cannot be found.
}
\keyword{internal}

