% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{getPosColNameMapUnit}
\alias{getPosColNameMapUnit}
\alias{getPosColNameMapUnit.character}
\alias{getPosColNameMapUnit.data.frame}
\title{Get map unit from position column heading.}
\usage{
getPosColNameMapUnit(x)

\method{getPosColNameMapUnit}{data.frame}(x)

\method{getPosColNameMapUnit}{character}(x)
}
\arguments{
\item{x}{A \code{mapframe} or equivalent \code{data.frame},
or a position column heading.}
}
\value{
Map unit indicated by the position column heading. Returns NA if map
unit information not found in position column heading.
}
\description{
Get map unit from position column heading.
}
\section{Map Units}{
 
A \code{map} or \code{mapframe} must express positions in either genetic or 
physical map units. Genetic map positions must be in centiMorgans (cM), while 
physical map positions must be in base pairs (bp), kilobases (kb), or 
Megabases (Mb).
}
\keyword{internal}

