% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQTLPeaks.R
\name{getQTLPeaks}
\alias{getQTLPeaks}
\alias{getQTLPeaks.mapframe}
\alias{getQTLPeaks.qtl}
\alias{getQTLPeaks.scanone}
\title{Get QTL peaks.}
\usage{
## Generic method.
getQTLPeaks(x, chr=NULL, ...)

\method{getQTLPeaks}{mapframe}(x, chr = NULL, threshold = NULL,
  lodcolumn = NULL, ...)

\method{getQTLPeaks}{qtl}(x, chr = NULL, qtl.indices = NULL, ...)

\method{getQTLPeaks}{scanone}(x, chr = NULL, threshold = NULL,
  lodcolumn = NULL, ...)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{scanone} or \code{qtl} object.}

\item{chr}{Vector indicating which sequences to consider. If no sequences
are specified, QTL peaks are returned for all available sequences.}

\item{...}{Further arguments (see below).}

\item{threshold}{For a \code{scanone} or equivalent object, this
contains a single \code{numeric} LOD significance threshold, or
an object (e.g. \code{summary.scanoneperm}) containing one such
threshold and its associated significance level.}

\item{lodcolumn}{In a \code{scanone} or equivalent object, this
parameter indicates which LOD column to consider. In other objects
with LOD-column-associated elements, this parameter indicates which
LOD-column-associated element to consider. This must be either a LOD
column name or an index \emph{with respect to the set of LOD columns}.
If no LOD column is specified and one such column is found, that column
is used by default; otherwise a LOD column must be specified.}

\item{qtl.indices}{In a \code{qtl} object, this option indicates the QTLs 
for which a QTL peak should be returned.}
}
\value{
A \code{mapframe} object in which each row indicates the locus of a 
QTL peak, with row names containing respective QTL names. Returns an empty
\code{mapframe} if there are no significant QTLs.
}
\description{
Get QTL peaks.
}
\seealso{
Other QTL functions: \code{\link{estPhysicalPositions}},
  \code{\link{getQTLFeatures}},
  \code{\link{getQTLIntervals}},
  \code{\link{plotQTLScanone}}
}
