% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{getResultAnalysesHDF5}
\alias{getResultAnalysesHDF5}
\title{Get names of result analyses for a given phenotype.}
\usage{
getResultAnalysesHDF5(infile, phenotype)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{phenotype}{Name of phenotype (or equivalent analysis units).}
}
\value{
Vector of result analysis names for the given phenotype.
Returns an empty character vector if no result analyses are found.
}
\description{
Get names of result analyses for a given phenotype.
}
\keyword{internal}

