% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{getResultInfoHDF5}
\alias{getResultInfoHDF5}
\title{Get result info for the given HDF5 scan files.}
\usage{
getResultInfoHDF5(infiles, phenotypes = NULL, analyses = NULL)
}
\arguments{
\item{phenotypes}{Phenotypes (or equivalent analysis units) for which result
info should be obtained. If none are specified, result info is returned for
all available phenotypes.}

\item{analyses}{Analyses for which result info should be obtained. If none
are specified, result info is returned for all available analyses.}

\item{infile}{Vector of input HDF5 scan file paths.}
}
\value{
Hierarchical list containing result info for each of the input HDF5
scan files. Each element of the returned list is named for a specific HDF5
scan file, and contains a sub-list of result info for the given scan file.
Each element of the sub-list is named for a specific phenotype, and contains
a sub-sub-list of result info for the given scan file and phenotype. Every
element of that sub-sub-list is named for a specific analysis, and is a
character vector of results for the given scan file, phenotype, and analysis.
}
\description{
Get result info for the given HDF5 scan files.
}
\keyword{internal}
