% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{getResultNamesHDF5}
\alias{getResultNamesHDF5}
\title{Get names of results for a given phenotype.}
\usage{
getResultNamesHDF5(infile, phenotype)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{phenotype}{Name of phenotype (or equivalent analysis unit).}
}
\value{
Vector of result names for the given phenotype. Returns \code{NULL}
if there are no results present for the given phenotype.
}
\description{
Get names of results for a given phenotype.
}
\keyword{internal}

