% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{getResultPhenotypesHDF5}
\alias{getResultPhenotypesHDF5}
\title{Get names of result phenotypes in HDF5 file.}
\usage{
getResultPhenotypesHDF5(infile)
}
\arguments{
\item{infile}{An input HDF5 file.}
}
\value{
Vector of result phenotype names. Returns an empty
character vector if there are no result phenotypes present.
}
\description{
Get names of result phenotypes in HDF5 file.
}
\keyword{internal}
