% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getRowIndices}
\alias{getRowIndices}
\title{Get row indices of object.}
\usage{
getRowIndices(x)
}
\arguments{
\item{x}{An object with rows.}
}
\value{
Integer vector of all row indices. Returns an empty integer vector if
the object has zero rows.
}
\description{
Get row indices of object.
}
\keyword{internal}

