% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getRowMask}
\alias{getRowMask}
\title{Get logical mask of object rows.}
\usage{
getRowMask(x, requested = NULL)
}
\arguments{
\item{x}{An object with rows.}

\item{requested}{A character vector of row names, a logical vector of length
equal to the number of object rows, or a numeric vector of row indices of the
input object. If this parameter is not specified, the returned logical mask
is \code{TRUE} for all rows.}
}
\value{
Logical mask of object rows.
}
\description{
Get logical mask of the rows of an object, as constrained by the
\code{requested} parameter. If using this function without \code{requested}
constraints, consider using the command \code{rep(TRUE, nrow(x))} instead.
}
\keyword{internal}
