% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getRunIndexList}
\alias{getRunIndexList}
\title{Get index list of successive runs in a vector.}
\usage{
getRunIndexList(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}
}
\value{
List of integer vectors, each containing indices for a run of
repeated values in the input vector. Each list element takes its name
from the corresponding repeated value. Returns an empty list if the
input vector is of length zero.
}
\description{
Get index list of successive runs in a vector.
}
\keyword{internal}
