% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getRunIndices}
\alias{getRunIndices}
\title{Get indices of successive runs in a run-length encoding.}
\usage{
getRunIndices(x)
}
\arguments{
\item{x}{An \code{rle} object.}
}
\value{
Integer vector of all run indices in the given run-length encoding,
which can be used to index into the \code{lengths} and \code{values} of the
given \code{rle} object. Returns an empty integer vector if the run-length
encoding has zero runs.
}
\description{
Get indices of successive runs in a run-length encoding.
}
\keyword{internal}

