% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getSampleNames}
\alias{getSampleNames}
\alias{getSampleNames,CrossInfo-method}
\title{Get sample names.}
\usage{
getSampleNames(cross.info, samples = NULL, strains = NULL, tetrads = NULL,
  simplify = FALSE)

\S4method{getSampleNames}{CrossInfo}(cross.info, samples = NULL,
  strains = NULL, tetrads = NULL, simplify = FALSE)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{samples}{Vector indicating samples to consider. This may be a 
logical vector of samples, an integer vector of sample indices, or 
a character vector that contains sample IDs or their equivalent 
syntactically valid name.}

\item{strains}{Vector indicating strains to consider. This may be a 
logical vector of strains, an integer vector of strain indices, or 
a character vector that contains strain IDs or their equivalent 
syntactically valid name.}

\item{tetrads}{Vector indicating tetrads to consider. This may be a 
logical vector of tetrads, or an integer vector of tetrad indices.}

\item{simplify}{Simplify list return value to a vector.}
}
\value{
If samples are specified by strain or tetrad, and if the option 
\code{simplify} is not TRUE, this method returns a list of character vectors,
each containing the sample names corresponding to a given strain/tetrad.
Otherwise, a vector of sample names is returned.
}
\description{
Get names of the specified samples. Samples can be specified at one of the 
following levels: sample, strain, or tetrad. If none are specified, all 
sample names are returned.
}
\keyword{internal}

