% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{getSamplesVCF}
\alias{getSamplesVCF}
\title{Get sample IDs from VCF file.}
\usage{
getSamplesVCF(file)
}
\arguments{
\item{file}{A VCF file path.}
}
\value{
Vector of the sample IDs in the given VCF file.
}
\description{
Get sample IDs from VCF file.
}
\keyword{internal}

