% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getScanoneThresholdInfo}
\alias{getScanoneThresholdInfo}
\alias{getScanoneThresholdInfo.numeric}
\alias{getScanoneThresholdInfo.qtlintervals}
\alias{getScanoneThresholdInfo.summary.scanonebins}
\alias{getScanoneThresholdInfo.summary.scanoneperm}
\title{Get \code{scanone} threshold info from object.}
\usage{
getScanoneThresholdInfo(x)

\method{getScanoneThresholdInfo}{numeric}(x)

\method{getScanoneThresholdInfo}{qtlintervals}(x)

\method{getScanoneThresholdInfo}{summary.scanonebins}(x)

\method{getScanoneThresholdInfo}{summary.scanoneperm}(x)
}
\arguments{
\item{x}{An object containing a single \code{scanone} threshold,
or an object associated with a single \code{scanone} threshold.}
}
\value{
A named list containing the \code{scanone} threshold
info that was extracted from the input object.
}
\description{
Get \code{scanone} threshold info from object.
}
\keyword{internal}

