% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{getSeqColIndex}
\alias{getSeqColIndex}
\title{Get sequence column index.}
\usage{
getSeqColIndex(x)
}
\arguments{
\item{x}{A \code{mapframe} or equivalent \code{data.frame}.}
}
\value{
Sequence column index.
}
\description{
Get sequence column index for a \code{mapframe} or equivalent
\code{data.frame}. The sequence column is taken to be the column
whose heading is \code{'chr'}. An error is raised if there is not
exactly one sequence column.
}
\keyword{internal}

