% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getSeqIndices}
\alias{getSeqIndices}
\alias{getSeqIndices,CrossInfo-method}
\title{Get sequence indices.}
\usage{
getSeqIndices(cross.info, sequences = NULL)

\S4method{getSeqIndices}{CrossInfo}(cross.info, sequences = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{sequences}{Vector indicating sequences to consider. Sequences may 
alternatively be specified by a logical vector. If none are specified, 
results are returned for all sequences.}
}
\value{
Integer vector of sequence indices.
}
\description{
Get sequence indices.
}
\keyword{internal}

