% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getSeqinfo}
\alias{getSeqinfo}
\title{Get \pkg{GenomeInfoDb} \code{Seqinfo} object.}
\usage{
getSeqinfo(genome = NULL)
}
\arguments{
\item{genome}{Genome for which a \pkg{GenomeInfoDb} \code{Seqinfo} object
should be returned. If this is not specified, a \code{Seqinfo} object is
returned for the current genome.}
}
\value{
A \code{Seqinfo} object for the given genome.
}
\description{
Get \pkg{GenomeInfoDb} \code{Seqinfo} object.
}
\keyword{internal}
