% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getSeqMarkers}
\alias{getSeqMarkers}
\alias{getSeqMarkers,CrossInfo-method}
\title{Get markers by sequence.}
\usage{
getSeqMarkers(cross.info, sequences = NULL, simplify = FALSE)

\S4method{getSeqMarkers}{CrossInfo}(cross.info, sequences = NULL,
  simplify = FALSE)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{sequences}{Vector indicating sequences to consider. Sequences may 
alternatively be specified by a logical vector. If none are specified, 
results are returned for all sequences.}

\item{simplify}{Simplify list return value to a vector.}
}
\value{
List of character vectors, each containing the marker IDs
corresponding to a given sequence. If option \code{simplify} is TRUE, 
this is simplified to a vector. Returns \code{NULL} if no sequence-marker 
information is available.
}
\description{
Get markers by sequence.
}
\keyword{internal}

