% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getSeqTable}
\alias{getSeqTable}
\title{Get genome sequence info table.}
\usage{
getSeqTable(genome = NULL)
}
\arguments{
\item{genome}{Genome for which sequence info should be returned. If this
is not specified, sequence info is returned for the current genome.}
}
\value{
A \code{data.frame} containing sequence info for the given genome.
This can be used to create, but is distinct from, a \pkg{GenomeInfoDb}
\code{Seqinfo} object.
}
\description{
Get genome sequence info table.
}
\keyword{internal}
