% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getSequences}
\alias{getSequences}
\alias{getSequences,CrossInfo-method}
\title{Get vector of normalised sequence labels.}
\usage{
getSequences(cross.info, sequences = NULL)

\S4method{getSequences}{CrossInfo}(cross.info, sequences = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{sequences}{Vector indicating sequences to consider. Sequences may 
alternatively be specified by a logical vector. If none are specified, 
results are returned for all sequences.}
}
\value{
Character vector of normalised sequence labels.
}
\description{
Get vector of normalised sequence labels.
}
\keyword{internal}

