% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getSpecialAttributeNames}
\alias{getSpecialAttributeNames}
\title{Get special attributes for the given object.}
\usage{
getSpecialAttributeNames(x)
}
\arguments{
\item{x}{R object.}
}
\value{
Vector of special attribute names.
}
\description{
Get special attributes for the given object.
}
\keyword{internal}

