% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getStrainIndices}
\alias{getStrainIndices}
\alias{getStrainIndices,CrossInfo-method}
\title{Get strain indices for the given samples.}
\usage{
getStrainIndices(cross.info, samples = NULL)

\S4method{getStrainIndices}{CrossInfo}(cross.info, samples = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{samples}{Vector indicating samples to consider. This may be a 
logical vector of samples, an integer vector of sample indices, or 
a character vector that contains sample IDs or their equivalent 
syntactically valid name.}
}
\value{
Integer vector of strain indices for the given samples. If no samples 
are specified, all strain indices are returned. Returns \code{NULL} if strain 
indices are not present.
}
\description{
Get strain indices for the given samples.
}
\keyword{internal}

