% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getTetradIndices}
\alias{getTetradIndices}
\alias{getTetradIndices,CrossInfo-method}
\title{Get tetrad indices for the given samples.}
\usage{
getTetradIndices(cross.info, samples = NULL)

\S4method{getTetradIndices}{CrossInfo}(cross.info, samples = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{samples}{Vector indicating samples to consider. This may be a 
logical vector of samples, an integer vector of sample indices, or 
a character vector that contains sample IDs or their equivalent 
syntactically valid name.}
}
\value{
Integer vector of tetrad indices for the given samples. If no samples 
are specified, all tetrad indices are returned. Returns \code{NULL} if tetrad 
indices are not present.
}
\description{
Get tetrad indices for the given samples.
}
\keyword{internal}

