% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getThresholdAlpha}
\alias{getThresholdAlpha}
\title{Get alpha of LOD threshold.}
\usage{
getThresholdAlpha(threshold)
}
\arguments{
\item{threshold}{Numeric vector whose only element is a LOD threshold value,
and whose name is the corresponding alpha value expressed as a percentage.}
}
\value{
Numeric vector whose only element is the alpha value of the
input LOD threshold, expressed as a real number in the range 0-1.
}
\description{
Get alpha of LOD threshold.
}
\keyword{internal}

