% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getXInfoFromFilenames}
\alias{getXInfoFromFilenames}
\title{Get experiment info from filenames.}
\usage{
getXInfoFromFilenames(filenames, pattern)
}
\arguments{
\item{filenames}{Character vector of file names.}

\item{pattern}{Pattern for extracting experiment info from file names. This
must be a valid Perl regex with named capture groups. Neither the capture
groups nor the pattern itself are required to match any given filename, but
all capture groups must have a name, and that name cannot clash with other
names that might be used alongside the extracted info.}
}
\value{
Character matrix in which each row contains the values of capture
groups for a given filename, and each column contains values of a given
capture group across filenames. Unmatched capture groups are represented by
\code{NA} values.
}
\description{
Get experiment info from filenames.
}
\keyword{internal}

