% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{gmapframe}
\alias{gmapframe}
\title{Create a new genetic \code{mapframe}.}
\usage{
gmapframe(...)
}
\arguments{
\item{...}{Further arguments. These are passed to the \code{data.frame}
constructor. All arguments should be passed as keyword arguments.}
}
\value{
A new \code{mapframe} with genetic map positions.
}
\description{
Create a new genetic \code{mapframe}.
}
\section{Mapframe}{


A \code{mapframe} object is a class of \code{data.frame} that contains mapped
information, or a \code{data.frame} with equivalent properties. The leftmost
two columns of every \code{mapframe} object must be as follows:

\itemize{
\item{\code{chr}:}   {Chromosome/sequence labels in
the order defined for the current reference genome.}
\item{\code{pos}:}   {Map positions in increasing order within each
chromosome/sequence. The map unit used in a \code{mapframe} position
column must be indicated by setting its \code{'map.unit'} attribute.}
}

Any number of additional data columns can be included. If present, the row
names indicate the locus IDs, which should be unique within the given
\code{mapframe}.
}

\seealso{
Other map utility functions: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{hasMapCSV}},
  \code{\link{inMapOrder}}, \code{\link{inferMapStep}},
  \code{\link{mapframe}}, \code{\link{mapsEqual}},
  \code{\link{orderMap}}, \code{\link{readMapCSV}},
  \code{\link{readMapHDF5}}, \code{\link{readMapframeCSV}},
  \code{\link{setMapUnit}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapHDF5}},
  \code{\link{writeMapframeCSV}}
}
