% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5Stack-class.R
\docType{class}
\name{H5Stack-class}
\alias{H5Stack}
\alias{H5Stack-class}
\title{An S4 class to manage a HDF5 object stack.}
\description{
An S4 class to manage a HDF5 object stack.
}
\section{Slots}{

\describe{
\item{\code{file}}{File containing the HDF5 objects to be accessed.}

\item{\code{ids}}{List of HDF5 object IDs.}
}}
\keyword{internal}

