% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{h5writeAttributes}
\alias{h5writeAttributes}
\title{Write attributes to HDF5 object.}
\usage{
h5writeAttributes(h5obj, object = NULL, attrs = NULL)
}
\arguments{
\item{h5obj}{HDF5 object to which attributes will be assigned.}

\item{object}{An R object whose attributes will be written.}

\item{attrs}{List of attributes to be written.}
}
\description{
Write attributes to HDF5 object.
}
\keyword{internal}

