% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{hasCrossHDF5}
\alias{hasCrossHDF5}
\title{Test if HDF5 file contains an \pkg{R/qtl} \code{cross} object.}
\usage{
hasCrossHDF5(infile)
}
\arguments{
\item{infile}{An input HDF5 file.}
}
\value{
\code{TRUE} if the given HDF5 file contains an \pkg{R/qtl}
\code{cross} object; \code{FALSE} otherwise.
}
\description{
Test if HDF5 file contains an \pkg{R/qtl} \code{cross} object.
}
\keyword{internal}

