% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasEnumGenotypes.R
\name{hasEnumGenotypes}
\alias{hasEnumGenotypes}
\alias{hasEnumGenotypes.cross}
\alias{hasEnumGenotypes.geno}
\title{Test if object contains enumerated genotypes.}
\usage{
hasEnumGenotypes(x)

\method{hasEnumGenotypes}{cross}(x)

\method{hasEnumGenotypes}{geno}(x)
}
\arguments{
\item{x}{Object containing genotype data.}
}
\value{
\code{TRUE} if object contains enumerated genotypes;
\code{FALSE} otherwise.
}
\description{
Test if object contains enumerated genotypes.
}
\keyword{internal}

