% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasFounderGenotypes.R
\name{hasFounderGenotypes}
\alias{hasFounderGenotypes}
\alias{hasFounderGenotypes.cross}
\alias{hasFounderGenotypes.geno}
\title{Test if object contains founder genotypes.}
\usage{
hasFounderGenotypes(x)

\method{hasFounderGenotypes}{cross}(x)

\method{hasFounderGenotypes}{geno}(x)
}
\arguments{
\item{x}{Object containing genotype data.}
}
\value{
\code{TRUE} if object contains founder genotypes;
\code{FALSE} otherwise.
}
\description{
Test if object contains founder genotypes.
}
\keyword{internal}

