% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{hasGenoQualVCF}
\alias{hasGenoQualVCF}
\title{Test if VCF file contains GQ scores.}
\usage{
hasGenoQualVCF(file)
}
\arguments{
\item{file}{A VCF file path.}
}
\value{
TRUE if the specified VCF file contains genotype quality (GQ) scores;
FALSE otherwise.
}
\description{
Test if VCF file contains GQ scores.
}
\keyword{internal}

