% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{hasMapHDF5}
\alias{hasMapHDF5}
\title{Test if HDF5 file contains the named map.}
\usage{
hasMapHDF5(infile, name)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{name}{Map name.}
}
\value{
\code{TRUE} if the given HDF5 file contains the named map;
\code{FALSE} otherwise.
}
\description{
Test if HDF5 file contains the named map.
}
\keyword{internal}
