% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{hasMarkerSeqs}
\alias{hasMarkerSeqs}
\alias{hasMarkerSeqs,CrossInfo-method}
\title{Test if \code{CrossInfo} object has a marker-sequence mapping.}
\usage{
hasMarkerSeqs(cross.info)

\S4method{hasMarkerSeqs}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
TRUE if marker-sequence mapping is present; FALSE otherwise.
}
\description{
Test if \code{CrossInfo} object has a marker-sequence mapping.
}
\keyword{internal}

