% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hasNames}
\alias{hasNames}
\title{Test if object has names.}
\usage{
hasNames(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if object has nonempty names with no \code{NA} values;
\code{FALSE} otherwise.
}
\description{
Test if object has names.
}
\keyword{internal}
