% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{hasObjectHDF5}
\alias{hasObjectHDF5}
\title{Test if HDF5 file contains the named object.}
\usage{
hasObjectHDF5(infile, h5name)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{h5name}{HDF5 object name.}
}
\value{
TRUE if the given HDF5 file contains the named object;
FALSE otherwise.
}
\description{
Test if HDF5 file contains the named object.
}
\keyword{internal}

