% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlintervals.R
\name{hasPhysicalPositions}
\alias{hasPhysicalPositions}
\title{Test if QTL intervals contain physical positions.}
\usage{
hasPhysicalPositions(qtl.intervals)
}
\arguments{
\item{qtl.intervals}{A \code{qtlintervals} object.}
}
\value{
TRUE if \code{qtlintervals} object contains physical positions
in terms of base-pair units; FALSE otherwise.
}
\description{
Test if QTL intervals contain physical positions.
}
\keyword{internal}

