% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{hasResultsOverviewHDF5}
\alias{hasResultsOverviewHDF5}
\title{Test if HDF5 file contains a QTL analysis results overview.}
\usage{
hasResultsOverviewHDF5(infile)
}
\arguments{
\item{infile}{An input HDF5 file.}
}
\value{
\code{TRUE} if the given HDF5 file contains a
QTL analysis results overview; \code{FALSE} otherwise.
}
\description{
Test if HDF5 file contains a QTL analysis results overview.
}
\keyword{internal}
