% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hasRownames}
\alias{hasRownames}
\title{Test if object has rownames.}
\usage{
hasRownames(x)
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}.}
}
\value{
TRUE if input object has non-default row names; FALSE otherwise.
}
\description{
Test if object has rownames.
}
\keyword{internal}

