% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{hasSampleIDs}
\alias{hasSampleIDs}
\alias{hasSampleIDs,CrossInfo-method}
\title{Test if \code{CrossInfo} object has sample IDs.}
\usage{
hasSampleIDs(cross.info)

\S4method{hasSampleIDs}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
TRUE if sample IDs are present; FALSE otherwise.
}
\description{
Test if \code{CrossInfo} object has sample IDs.
}
\keyword{internal}

