% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{hasStrainIndices}
\alias{hasStrainIndices}
\alias{hasStrainIndices,CrossInfo-method}
\title{Test if \code{CrossInfo} object has strain indices.}
\usage{
hasStrainIndices(cross.info)

\S4method{hasStrainIndices}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
TRUE if strain indices are present; FALSE otherwise.
}
\description{
Test if \code{CrossInfo} object has strain indices.
}
\keyword{internal}

