% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{hasTetradIndices}
\alias{hasTetradIndices}
\alias{hasTetradIndices,CrossInfo-method}
\title{Test if \code{CrossInfo} object has sample tetrad indices.}
\usage{
hasTetradIndices(cross.info)

\S4method{hasTetradIndices}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
TRUE if sample tetrad indices are present; FALSE otherwise.
}
\description{
Test if \code{CrossInfo} object has sample tetrad indices.
}
\keyword{internal}

