% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{hasTimeSeriesPhenotypes}
\alias{hasTimeSeriesPhenotypes}
\title{Test if \code{cross} contains time-series phenotypes.}
\usage{
hasTimeSeriesPhenotypes(cross, allow.gaps = TRUE, tol = 1e-05)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{allow.gaps}{Allow gaps in time series, provided
that any gaps are a multiple of the inferred time step.}

\item{tol}{Tolerance for time step equality.}
}
\value{
\code{TRUE} if \code{cross} seems to have time-series phenotypes,
given the specified constraints; \code{FALSE} otherwise.
}
\description{
Test if \code{cross} contains time-series phenotypes.
}
\section{Time-Series Phenotypes}{


A set of phenotypes can be designated as a time-series by naming each
phenotype with the time point at which phenotype observations were made
(e.g. \code{'0.0'}, \code{'1.0'}, \code{'2.0'}). Time points can be in
any unit, but must be non-negative, monotonically increasing, and have
a consistent time step. If some time points are missing, the resulting
gap in time must be a multiple of the time step.
}
\seealso{
Other cross object functions: \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{permIndices}}, \code{\link{pull.alleles}},
  \code{\link{pull.chr}}, \code{\link{pull.crosstype}},
  \code{\link{pull.ind}}, \code{\link{readCrossCSV}},
  \code{\link{writeCrossCSV}}

Other time-series functions: \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}
}

