% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{inferFormatFromFilename}
\alias{inferFormatFromFilename}
\title{Infer format from filename.}
\usage{
inferFormatFromFilename(filename)
}
\arguments{
\item{filename}{A file name.}
}
\value{
File format inferred from extension of file name.
Returns \code{unknown} if format could not be inferred.
}
\description{
Infer format from filename.
}
\keyword{internal}

