% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{inferMapStep}
\alias{inferMapStep}
\title{Infer map step size.}
\usage{
inferMapStep(x, tol = 1e-05)
}
\arguments{
\item{x}{Object containing map data.}

\item{tol}{Tolerance for step equality.}
}
\value{
Inferred map step. Returns \code{NULL}
if the map step could not be inferred.
}
\description{
Infer map step size.
}
\seealso{
Other map utility functions: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{hasMapCSV}}, \code{\link{inMapOrder}},
  \code{\link{mapframe}}, \code{\link{orderMap}},
  \code{\link{readMapCSV}}, \code{\link{readMapHDF5}},
  \code{\link{readMapframeCSV}}, \code{\link{setMapUnit}},
  \code{\link{writeMapCSV}}, \code{\link{writeMapHDF5}},
  \code{\link{writeMapframeCSV}}
}

