% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{inferStrainIndices}
\alias{inferStrainIndices}
\alias{inferStrainIndices.character}
\alias{inferStrainIndices.cross}
\title{Infer sample strain indices.}
\usage{
inferStrainIndices(x)

\method{inferStrainIndices}{character}(x)

\method{inferStrainIndices}{cross}(x)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{cross} object, or a character vector of sample
IDs.}
}
\value{
Vector of sample strain indices inferred from the input object.
}
\description{
Infers strain indices of a set of samples from the given object. A character 
vector of sample IDs can be passed as an argument, in which case the strain 
indices will be inferred from each set of identical consecutive sample IDs. 
If a \code{cross} object is passed as an argument, strain indices will be 
inferred from sample IDs, if present, or otherwise by comparison of sample 
genotypes.
}
\seealso{
Other cross utilities: \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{pull.alleles}}, \code{\link{pull.chr}},
  \code{\link{pull.crosstype}}, \code{\link{pull.ind}}
}

