% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{inferTetradIndices}
\alias{inferTetradIndices}
\alias{inferTetradIndices.character}
\alias{inferTetradIndices.cross}
\title{Infer sample tetrad indices.}
\usage{
inferTetradIndices(x)

\method{inferTetradIndices}{character}(x)

\method{inferTetradIndices}{cross}(x)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{cross} object, or a character vector of sample IDs.}
}
\value{
Vector of sample tetrad indices inferred from the input object. 
Returns \code{NULL} if tetrad pattern could not be found.
}
\description{
Infers tetrad indices of a set of samples from the given object. 
A character vector of sample IDs can be passed as the first argument, in 
which case the tetrad indices will be inferred from the sample IDs themselves.
If sample IDs have a numeric suffix, these are taken as segregant numbers 
and used to infer which tetrad each sample ID might belong to. 
  
If a \code{cross} is the first argument, tetrad indices will be inferred 
from sample IDs, if present, or otherwise by comparison of sample genotypes.
}
\seealso{
Other cross utilities: \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{pull.alleles}}, \code{\link{pull.chr}},
  \code{\link{pull.ind}}
}

