% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{inferTimeStep}
\alias{inferTimeStep}
\title{Infer time step of time-series phenotypes.}
\usage{
inferTimeStep(cross, allow.gaps = TRUE, tol = 1e-05)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{allow.gaps}{Allow gaps in time series, provided that any gaps are a 
multiple of the inferred time step.}

\item{tol}{Tolerance for time step equality.}
}
\value{
Inferred time step. Returns \code{NULL} if time step could not be inferred.
}
\description{
Infer time step of time-series phenotypes.
}
\seealso{
Other cross utilities: \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{pull.alleles}}, \code{\link{pull.chr}},
  \code{\link{pull.ind}}
}

