% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5Stack-class.R
\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize,H5Stack-method}
\title{Init \code{H5Stack} object.}
\usage{
\S4method{initialize}{H5Stack}(.Object, file, h5name = NULL)
}
\arguments{
\item{.Object}{A \code{H5Stack} object.}

\item{file}{A HDF5 file path.}

\item{h5name}{Name of HDF5 object to access.}
}
\value{
A prepared \code{H5Stack} object.
}
\description{
Init \code{H5Stack} object.
}
\keyword{internal}

