% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{inMapOrder}
\alias{inMapOrder}
\alias{inMapOrder.data.frame}
\alias{inMapOrder.map}
\title{Test if object is in map order.}
\usage{
inMapOrder(x)

\method{inMapOrder}{data.frame}(x)

\method{inMapOrder}{map}(x)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
\code{TRUE} if map data of object is ordered by
sequence, then map position; \code{FALSE} otherwise.
}
\description{
Test if object is in map order.
}
\seealso{
Other map utility functions: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{hasMapCSV}}, \code{\link{inferMapStep}},
  \code{\link{mapframe}}, \code{\link{orderMap}},
  \code{\link{readMapCSV}}, \code{\link{readMapHDF5}},
  \code{\link{readMapframeCSV}}, \code{\link{setMapUnit}},
  \code{\link{writeMapCSV}}, \code{\link{writeMapHDF5}},
  \code{\link{writeMapframeCSV}}
}

