% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{inRange}
\alias{inRange}
\title{Test if numbers lie within a range.}
\usage{
inRange(n, interval)
}
\arguments{
\item{n}{Numeric vector.}

\item{interval}{Numeric vector containing the minimum and maximum values of
the range, respectively.}
}
\value{
Logical vector indicating which elements of the input numeric vector
lie within the specified range.
}
\description{
Test if numbers lie within a range.
}
\keyword{internal}
