% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{insertColumn}
\alias{insertColumn}
\title{Insert column in an object.}
\usage{
insertColumn(x, col.index, col.name = NULL, data = NA)
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}.}

\item{col.index}{Column index of inserted column.}

\item{col.name}{Column name of inserted column. If a column name is 
specified, but the object has no existing column names, default names 
(e.g. 'COL05') will be assigned to the existing columns. If a column 
name is not specified, but the object has existing column names, the 
inserted column will be assigned a default column name.}

\item{data}{Optional vector of data to insert in the new column. The length 
of this vector should be evenly divisible by the number of rows in the data 
frame.}
}
\value{
Input object with column inserted.
}
\description{
Insert column in an object.
}
\keyword{internal}

