% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{interpTimeSeries}
\alias{interpTimeSeries}
\title{Interpolate gaps in a time-series.}
\usage{
interpTimeSeries(cross, tol = 1e-05)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{tol}{Tolerance for time step equality.}
}
\value{
The input \code{cross} object is returned with gaps in time-series 
phenotypes filled with values interpolated from the gap endpoints.
}
\description{
Interpolate values for gaps in time-series phenotypes of a \code{cross} 
object.
}
\seealso{
Other cross utilities: \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}}, \code{\link{padTimeSeries}},
  \code{\link{permCross}}, \code{\link{pull.alleles}},
  \code{\link{pull.chr}}, \code{\link{pull.ind}}
}

