% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{intersectLoci}
\alias{intersectLoci}
\title{Find intersection set of loci.}
\usage{
intersectLoci(...)
}
\arguments{
\item{...}{One or more objects containing map data.}
}
\value{
Locus \code{mapframe} with loci common to all input objects.
}
\description{
Find intersection set of loci.
}
\keyword{internal}
