% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{is.mapping}
\alias{is.mapping}
\title{Test if object is a \code{mapping}.}
\usage{
is.mapping(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if object is a \code{mapping};
\code{FALSE} otherwise.
}
\description{
Test if object is a \code{mapping}.
}
\seealso{
Other mapping object functions: \code{\link{as.mapping}},
  \code{\link{mappingKeys}}, \code{\link{mappingValues}},
  \code{\link{mapping}}
}
