% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isBOOL}
\alias{isBOOL}
\title{Test for a single logical value.}
\usage{
isBOOL(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if object is a single logical value; FALSE otherwise.
}
\description{
Test for a single logical value.
}
\keyword{internal}

