% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isDefaultMarkerID}
\alias{isDefaultMarkerID}
\title{Test for default marker IDs.}
\usage{
isDefaultMarkerID(ids)
}
\arguments{
\item{ids}{Vector of locus IDs.}
}
\value{
Logical vector indicating which elements are valid IDs that follow 
the pattern of a default marker ID.
}
\description{
Test for default marker IDs.
}

