% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isEnumAllele}
\alias{isEnumAllele}
\title{Test if symbol is a valid enumerated allele.}
\usage{
isEnumAllele(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if symbol is a valid enumerated allele; FALSE otherwise.
}
\description{
Test if symbol is a valid enumerated allele.
}
\keyword{internal}

