% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isEnumGenotype}
\alias{isEnumGenotype}
\title{Test if symbol is a valid enumerated genotype.}
\usage{
isEnumGenotype(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if symbol is a valid enumerated genotype; FALSE otherwise.
}
\description{
Test if symbol is a valid enumerated genotype.
}
\keyword{internal}

