% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isFALSE}
\alias{isFALSE}
\title{Test for a single FALSE value.}
\usage{
isFALSE(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if object is a single FALSE value; FALSE otherwise.
}
\description{
Test for a single FALSE value.
}
\keyword{internal}

