% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isFounderAllele}
\alias{isFounderAllele}
\title{Test if symbol is a valid founder allele.}
\usage{
isFounderAllele(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if symbol is a valid founder allele; FALSE otherwise.
}
\description{
Test if symbol is a valid founder allele.
}
\keyword{internal}

