% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isFounderGenotype}
\alias{isFounderGenotype}
\title{Test if symbol is a valid founder genotype.}
\usage{
isFounderGenotype(x, strict = FALSE)
}
\arguments{
\item{x}{Test object.}

\item{strict}{Return \code{TRUE} only for complete genotypes.}
}
\value{
\code{TRUE} if symbol is a valid founder genotype;
\code{FALSE} otherwise.
}
\description{
Test if symbol is a valid founder genotype.
}
\section{Founder Alleles and Genotypes}{


A founder allele is a single letter corresponding to a specific founder
strain, indicating that the region containing the marker derives from that
founder. For example, \code{'W'} might be used to represent the Western
Africa founder strain \code{DBVPG6044}, or \code{'S'} might be used for
the Sake strain \code{Y12}).

For a given ploidy \code{N}, a founder genotype is the concatenation of the
set of \code{N} founder alleles. For example, given one founder allele
\code{'W'} and another \code{'S'}, the founder genotype is \code{'WS'}.
}
\keyword{internal}

