% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{isGeneticMap}
\alias{isGeneticMap}
\title{Test if object is a genetic \code{map}.}
\usage{
isGeneticMap(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if object is a genetic \code{map} object; FALSE otherwise.
}
\description{
Test if object is a genetic \code{map}.
}
\keyword{internal}

