% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{isGeneticMapframe}
\alias{isGeneticMapframe}
\title{Test if object is a genetic \code{mapframe}.}
\usage{
isGeneticMapframe(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if object is a \code{mapframe} whose pos column contains
genetic map positions; \code{FALSE} otherwise.
}
\description{
Test if object is a genetic \code{mapframe}.
}
\keyword{internal}
