% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{isGeneticMapUnit}
\alias{isGeneticMapUnit}
\title{Test if object is a known genetic map unit.}
\usage{
isGeneticMapUnit(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if object is a known genetic map unit; FALSE otherwise.
}
\description{
Test if object is a known genetic map unit.
}
\keyword{internal}

