% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isNegativeNumber}
\alias{isNegativeNumber}
\title{Test for negative numbers.}
\usage{
isNegativeNumber(n)
}
\arguments{
\item{n}{Test vector.}
}
\value{
Logical vector indicating which elements are less than zero.
}
\description{
Test for negative numbers.
}
\keyword{internal}
