% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isNonNegativeNumber}
\alias{isNonNegativeNumber}
\title{Test for non-negative numbers.}
\usage{
isNonNegativeNumber(n)
}
\arguments{
\item{n}{Test vector.}
}
\value{
Logical vector indicating which elements are greater than or equal to 
zero.
}
\description{
Test for non-negative numbers.
}
\keyword{internal}

