% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{isNormSeq}
\alias{isNormSeq}
\title{Test for normalised sequence labels.}
\usage{
isNormSeq(x)
}
\arguments{
\item{x}{Vector of sequence labels.}
}
\value{
Logical vector indicating which sequence labels are normalised.
}
\description{
Test for normalised sequence labels.
}
\seealso{
Other chromosome/sequence functions: \code{\link{formatChr}},
  \code{\link{formatSeq}}, \code{\link{isNormChr}},
  \code{\link{normChr}}, \code{\link{normSeq}},
  \code{\link{orderChr}}, \code{\link{orderSeq}},
  \code{\link{rankChr}}, \code{\link{rankSeq}},
  \code{\link{sortChr}}, \code{\link{sortSeq}}
}

