% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{isPhysicalMap}
\alias{isPhysicalMap}
\title{Test if object is a physical \code{map}.}
\usage{
isPhysicalMap(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if object is a physical \code{map};
\code{FALSE} otherwise.
}
\description{
Test if object is a physical \code{map}.
}
\keyword{internal}

