% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{isPhysicalMapframe}
\alias{isPhysicalMapframe}
\title{Test if object is a physical \code{mapframe}.}
\usage{
isPhysicalMapframe(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if object is a \code{mapframe} whose pos column contains
physical map positions; \code{FALSE} otherwise.
}
\description{
Test if object is a physical \code{mapframe}.
}
\keyword{internal}
