% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{isPhysicalMapUnit}
\alias{isPhysicalMapUnit}
\title{Test if object is a known physical map unit.}
\usage{
isPhysicalMapUnit(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if object is a known physical map unit;
\code{FALSE} otherwise.
}
\description{
Test if object is a known physical map unit.
}
\keyword{internal}
