% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isPositiveNumber}
\alias{isPositiveNumber}
\title{Test for positive numbers.}
\usage{
isPositiveNumber(n)
}
\arguments{
\item{n}{Test vector.}
}
\value{
Logical vector indicating which elements are positive numbers.
}
\description{
Test for positive numbers.
}
\keyword{internal}

