% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isPositiveWholeNumber}
\alias{isPositiveWholeNumber}
\title{Test for positive whole numbers.}
\usage{
isPositiveWholeNumber(n, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{n}{Test vector.}

\item{tol}{Numeric tolerance.}
}
\value{
Logical vector indicating which elements are positive whole numbers.
}
\description{
Test for positive whole numbers.
}
\keyword{internal}
