% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isProbability}
\alias{isProbability}
\title{Test for valid probabilities.}
\usage{
isProbability(n)
}
\arguments{
\item{n}{Test vector.}
}
\value{
Logical vector indicating which elements are valid probabilities.
}
\description{
Test for valid probabilities.
}
\keyword{internal}

