% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isRawAllele}
\alias{isRawAllele}
\title{Test if symbol is a valid raw allele.}
\usage{
isRawAllele(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if symbol is a valid raw allele; FALSE otherwise.
}
\description{
Test if symbol is a valid raw allele.
}
\keyword{internal}

