% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isSingleChar}
\alias{isSingleChar}
\title{Test for a single character.}
\usage{
isSingleChar(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if the object is a single character;
\code{FALSE} otherwise.
}
\description{
Test for a single character.
}
\keyword{internal}

