% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isSingleNonNegativeNumber}
\alias{isSingleNonNegativeNumber}
\title{Test for a single non-negative number.}
\usage{
isSingleNonNegativeNumber(n)
}
\arguments{
\item{n}{Test object.}
}
\value{
TRUE if the object is a single non-negative number; 
FALSE otherwise.
}
\description{
Test for a single non-negative number.
}
\keyword{internal}

