% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isSingleNonNegativeWholeNumber}
\alias{isSingleNonNegativeWholeNumber}
\title{Test for a single non-negative whole number.}
\usage{
isSingleNonNegativeWholeNumber(n, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{n}{Test object.}

\item{tol}{Numeric tolerance.}
}
\value{
\code{TRUE} if the object is a single non-negative whole number;
\code{FALSE} otherwise.
}
\description{
Test for a single non-negative whole number.
}
\keyword{internal}
