% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isSinglePositiveNumber}
\alias{isSinglePositiveNumber}
\title{Test for a single positive number.}
\usage{
isSinglePositiveNumber(n)
}
\arguments{
\item{n}{Test object.}
}
\value{
\code{TRUE} if the object is a single positive number;
\code{FALSE} otherwise.
}
\description{
Test for a single positive number.
}
\keyword{internal}

