% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isSingleProbability}
\alias{isSingleProbability}
\title{Test for a single valid probability.}
\usage{
isSingleProbability(n)
}
\arguments{
\item{n}{Test object.}
}
\value{
TRUE if the object is a single valid probability; 
FALSE otherwise.
}
\description{
Test for a single valid probability.
}
\keyword{internal}

