% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isSingleString}
\alias{isSingleString}
\title{Test for a single character string.}
\usage{
isSingleString(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if the object is a single string; FALSE otherwise.
}
\description{
Test for a single character string.
}
\keyword{internal}

