% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isSingleWholeNumber}
\alias{isSingleWholeNumber}
\title{Test for a single whole number.}
\usage{
isSingleWholeNumber(n, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{n}{Test object.}

\item{tol}{Numeric tolerance.}
}
\value{
\code{TRUE} if the object is a single whole number;
\code{FALSE} otherwise.
}
\description{
Test for a single whole number.
}
\keyword{internal}

