% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isValidAllele}
\alias{isValidAllele}
\title{Test if symbol is a valid allele.}
\usage{
isValidAllele(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if symbol is a valid allele; FALSE otherwise.
}
\description{
Test if symbol is a valid allele.
}
\keyword{internal}

