% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isValidGenotype}
\alias{isValidGenotype}
\title{Test if symbol is a valid genotype.}
\usage{
isValidGenotype(x, strict = FALSE)
}
\arguments{
\item{x}{Test object.}

\item{strict}{Return TRUE only for complete genotypes.}
}
\value{
TRUE if symbol is a valid enumerated or founder genotype;
FALSE otherwise.
}
\description{
Test if symbol is a valid genotype.
}
\keyword{internal}

