% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isValidID}
\alias{isValidID}
\title{Test identifier validity.}
\usage{
isValidID(x)
}
\arguments{
\item{x}{Test vector.}
}
\value{
Logical vector indicating which elements are valid identifiers.
}
\description{
Test identifier validity.
}
\section{Item IDs}{


Various items in \pkg{shmootl} (e.g. chromosomes/sequences, markers, samples,
phenotypes) have an associated identifier (ID). A valid item ID in
\pkg{shmootl} is a string of nonzero length, containing any combination
of printable ASCII characters (including letters and numbers), except for a
single quote, double quotes, backtick, comma, forward slash, or backslash.
}

