% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{isValidMapUnit}
\alias{isValidMapUnit}
\title{Test if object is a known map unit.}
\usage{
isValidMapUnit(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if object is a known genetic or physical map unit;
FALSE otherwise.
}
\description{
Test if object is a known map unit.
}
\keyword{internal}

