% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isValidMissingValue}
\alias{isValidMissingValue}
\title{Test if object is a valid missing value.}
\usage{
isValidMissingValue(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if object is a valid missing value; FALSE otherwise.
}
\description{
Test if object is a valid missing value.
}
\keyword{internal}

