% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isValidName}
\alias{isValidName}
\title{Test for syntactically valid names.}
\usage{
isValidName(x)
}
\arguments{
\item{x}{Test vector.}
}
\value{
Logical vector indicating which elements are syntactically valid R
names.
}
\description{
Test for syntactically valid names.
}
\section{Syntactic Names}{


A syntactically valid R name contains only letters, numbers, dots, and
underscores. The name should start with a letter, or a dot that is not
followed by a number. In addition, specific reserved words are not
permitted. Many R packages and functions automatically render item names
syntactically valid, so for example, a phenotype name may be changed when
added to an \pkg{R/qtl} \code{cross} object. For more information, see the
relevant \href{https://cran.r-project.org/manuals.html}{R documentation}.
See also the base R function \code{make.names}.
}

