% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isWholeNumber}
\alias{isWholeNumber}
\title{Test for whole numbers.}
\usage{
isWholeNumber(n, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{n}{Test vector.}

\item{tol}{Numeric tolerance.}
}
\value{
Logical vector indicating which elements are whole numbers.
}
\description{
Test for whole numbers.
}
\keyword{internal}
