% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{joinH5ObjectNameParts}
\alias{joinH5ObjectNameParts}
\title{Join components of HDF5 object name.}
\usage{
joinH5ObjectNameParts(components)
}
\arguments{
\item{components}{HDF5 name components.}
}
\value{
HDF5 object name.
}
\description{
Join components of HDF5 object name.
}
\keyword{internal}

