% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{joinSeqLabels}
\alias{joinSeqLabels}
\title{Join sequence labels.}
\usage{
joinSeqLabels(x)
}
\arguments{
\item{x}{A \code{data.frame} representing split sequence labels.}
}
\value{
Character vector of joined sequence labels.
}
\description{
Join sequence labels.
}
\keyword{internal}
