% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{listGenomes}
\alias{listGenomes}
\title{List available genomes.}
\usage{
listGenomes()
}
\value{
Character vector containing the names
of genomes that are available in \pkg{shmootl}.
}
\description{
List available genomes.
}
\seealso{
Other genome functions: \code{\link{genomeOpt}}
}
