% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{listWorksheets}
\alias{listWorksheets}
\title{List available output worksheets.}
\usage{
listWorksheets()
}
\value{
Character vector containing the names of worksheets that can be set
as output to a spreadsheet file in \pkg{shmootl} spreadsheet functions. This
excludes summary worksheets such as \code{'README'} and \code{'Overview'},
as these are included in every output spreadsheet file.
}
\description{
List available output worksheets.
}
\seealso{
Other Excel functions: \code{\link{writeDigestExcel}}
}

