% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadChrTable}
\alias{loadChrTable}
\title{Load chromosome info.}
\usage{
loadChrTable()
}
\value{
A \code{data.frame} object containing chromosome info.
}
\description{
This function loads standard chromosome info.
}
\keyword{internal}

