% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadListFromLine}
\alias{loadListFromLine}
\title{Load a list from a line of text.}
\usage{
loadListFromLine(line)
}
\arguments{
\item{line}{Character vector of length one, containing the line to be loaded.}
}
\value{
List or vector of loaded data.
}
\description{
The input line is loaded as a YAML flow-style list.
}
\keyword{internal}

