% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadMapping}
\alias{loadMapping}
\title{Load a mapping from a line or file.}
\usage{
loadMapping(line = NULL, file = NULL)
}
\arguments{
\item{line}{Character string containing the line to be loaded.}

\item{file}{Character string containing the name of the file to be loaded.}
}
\value{
A \code{mapping} of unique string keys to individual values.
}
\description{
Given a specified line of text or filename, the input text is loaded
as a YAML mapping, and returned as a simple \code{mapping} object.
}
\keyword{internal}

