% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadSeqInfo}
\alias{loadSeqInfo}
\title{Load genome sequence info.}
\usage{
loadSeqInfo()
}
\value{
A \code{list} of \code{data.frame} objects, each element named for a
given genome and containing sequence info for that genome.
}
\description{
This function loads the sequence info of genomes for which package data
is available.
}
\keyword{internal}

