% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadVector}
\alias{loadVector}
\title{Load a vector from a line or file.}
\usage{
loadVector(line = NULL, file = NULL, type = NULL)
}
\arguments{
\item{line}{Character string containing the line to be loaded.}

\item{file}{Character string containing the name of the file to be loaded.}
}
\value{
Vector of loaded data.
}
\description{
Given a specified line of text or filename, the input text is loaded
as a YAML list, and returned as an R \code{vector} object.
}
\keyword{internal}
