% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makeAlleleSet}
\alias{makeAlleleSet}
\title{Make allele set from the given genotypes.}
\usage{
makeAlleleSet(genotypes)
}
\arguments{
\item{genotypes}{Sorted character vector of genotype symbols.}
}
\value{
Character vector of allele symbols generated from the given genotypes.
}
\description{
Make allele set from the given genotypes.
}
\keyword{internal}
