% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{makeCross}
\alias{makeCross}
\title{Make an \pkg{R/qtl} \code{cross} object.}
\usage{
makeCross(geno, pheno)
}
\arguments{
\item{geno}{A \code{geno} object.}

\item{pheno}{A \code{pheno} object.}
}
\value{
An \pkg{R/qtl} \code{cross} object.
}
\description{
Make an \pkg{R/qtl} \code{cross} object.
}
\keyword{internal}

