% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{makeDefaultElementNames}
\alias{makeDefaultElementNames}
\title{Make default HDF5 group element names.}
\usage{
makeDefaultElementNames(n)
}
\arguments{
\item{n}{Number of default group element names to generate.}
}
\value{
Character vector of default group element names.
}
\description{
Following the convention of the \pkg{rhdf5} package, default HDF5 group 
element names are of the form 'ELT1', where 1 can be any positive integer. 
Leading zeros are used to pad element numbers, depending on the group size, 
so that the generated name strings will sort in numerical order.
}
\keyword{internal}

