% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{makeDefaultMapName}
\alias{makeDefaultMapName}
\title{Get default map name.}
\usage{
makeDefaultMapName(map)
}
\arguments{
\item{map}{An \pkg{R/qtl} \code{map} object.}
}
\value{
Default map name.
}
\description{
Get default map name.
}
\keyword{internal}

