% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makeDefaultMarkerIDs}
\alias{makeDefaultMarkerIDs}
\title{Make default marker IDs for loci.}
\usage{
makeDefaultMarkerIDs(loc, sep = c(":", "-"))
}
\arguments{
\item{loc}{Locus \code{data.frame} specifying physical map positions.}

\item{sep}{Separator between the two parts of a default marker ID.}
}
\value{
Character vector of default marker IDs.
}
\description{
Make default marker IDs for loci.
}

