% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{makeDefaultResultName}
\alias{makeDefaultResultName}
\title{Get default QTL analysis result name.}
\usage{
makeDefaultResultName(result)
}
\arguments{
\item{result}{Object containing a QTL analysis result.}
}
\value{
Default QTL analysis result name.
}
\description{
Get default QTL analysis result name.
}
\keyword{internal}

