% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{makeDummyMap}
\alias{makeDummyMap}
\title{Make a dummy map.}
\usage{
makeDummyMap(locus.seqs, locus.ids = NULL, map.unit = "cM", step = 5)
}
\arguments{
\item{locus.seqs}{Sequences corresponding to individual loci.}

\item{locus.ids}{Individual locus IDs.}

\item{map.unit}{Map unit.}

\item{step}{Map step size.}
}
\description{
Make a dummy map from the given combination of locus sequences, map unit, and
map step size. If locus IDs are specified, the locus IDs of the dummy map
will be set from these. Otherwise, locus IDs are set to pseudomarker IDs or
default marker IDs, according to whether map units are for a genetic or
physical map, respectively.
}
\details{
Locus sequences and locus IDs are assumed to be ordered. As in R/qtl, any
sequence with too few loci is extended, to ensure that all sequences have
enough loci.
}
\keyword{internal}

