% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{makeFounderGenoMatrix}
\alias{makeFounderGenoMatrix}
\title{Make founder genotype matrix from genotype data.}
\usage{
makeFounderGenoMatrix(x, y)
}
\arguments{
\item{x}{Raw sample genotype \code{array}.}

\item{y}{Raw founder genotype \code{array}.}
}
\value{
A founder genotype matrix, with genotypes encoded as integers and
their corresponding genotype symbols in the attribute \code{'genotypes'}.
}
\description{
Given input genotype data for cross segregant samples and their founder 
strains, this function assigns a symbol to each locus according to the 
inferred founder genotype.
}
\keyword{internal}

