% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{makeGeno}
\alias{makeGeno}
\title{Make an \pkg{R/qtl} \code{cross} \code{geno} object.}
\usage{
makeGeno(x, y = NULL)
}
\arguments{
\item{x}{Raw sample genotype \code{array}.}

\item{y}{Raw founder genotype \code{array}.}
}
\value{
A \code{geno} object.
}
\description{
Make an \pkg{R/qtl} \code{cross} \code{geno} object.
}
\keyword{internal}

