% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{makeGenoTable}
\alias{makeGenoTable}
\alias{makeGenoTable.list}
\title{Make genotype table from genotype data.}
\usage{
makeGenoTable(x, chr = NULL, digits = NULL, include.mapunit = TRUE)

\method{makeGenoTable}{list}(x, chr = NULL, digits = NULL,
  include.mapunit = TRUE)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{cross} \code{geno} object.}

\item{chr}{Vector of sequences for which genotype data should be included in 
the genotype table. If none are specified, a genotype table is returned for 
all available sequences.}

\item{digits}{If specified, round genetic map positions to the specified 
number of digits.}

\item{include.mapunit}{Include map unit information in map positions.}
}
\value{
A \code{data.frame} containing a genotype table.
}
\description{
Make genotype table from genotype data.
}
\keyword{internal}

