% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makeGenotypeSet}
\alias{makeGenotypeSet}
\title{Make genotype set from the given alleles and cross type.}
\usage{
makeGenotypeSet(alleles, crosstype, strict = FALSE)
}
\arguments{
\item{alleles}{Sorted character vector containing a set of allele symbols.}

\item{crosstype}{Cross type for which genotype set should be made.}

\item{strict}{Make genotype set containing only complete genotypes.}
}
\value{
Character vector of genotype symbols generated from the
given alleles in a way determined by the specified cross type.
}
\description{
Make genotype set from the given alleles and cross type.
}
\keyword{internal}
