% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{makeGroupObjectNames}
\alias{makeGroupObjectNames}
\title{Make valid HDF5 group object names.}
\usage{
makeGroupObjectNames(group.names = NULL, group.size = NULL)
}
\arguments{
\item{group.names}{HDF5 group object names.}

\item{group.size}{Number of elements in HDF5 group.}
}
\value{
HDF5 group object names.
}
\description{
Group names are made based on the proposed group names or the group size. 
Default group element names will replace every group name that is either 
an empty string or \code{NA} value. Default group element names are of the
form \code{'ELT1'}, where the number indicates the position of the element
within the given group.
}
\seealso{
makeDefaultElementNames
}
\keyword{internal}

