% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonebins.R
\name{makeLODBinLabels}
\alias{makeLODBinLabels}
\title{Make LOD bin labels from bin starts.}
\usage{
makeLODBinLabels(bin.starts)
}
\arguments{
\item{bin.starts}{Numeric vector of bin starts.}
}
\value{
Character vector of LOD bin labels corresponding to the input bins.
}
\description{
Make LOD bin labels from bin starts.
}
\keyword{internal}
