% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makeNumbers}
\alias{makeNumbers}
\title{Make numbers from numeric names.}
\usage{
makeNumbers(x)
}
\arguments{
\item{x}{Vector of syntactically valid names that represent numbers.}
}
\value{
Numeric vector of numbers corresponding to the input names.
}
\description{
Make numbers from numeric names.
}
\keyword{internal}

