% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{makePlaceholderPheno}
\alias{makePlaceholderPheno}
\title{Make a placeholder \code{pheno} object.}
\usage{
makePlaceholderPheno(samples)
}
\arguments{
\item{samples}{Integer vector of sample indices or character vector of
sample IDs.}
}
\value{
A placeholder \code{pheno} object with the given samples.
}
\description{
Make a placeholder \code{pheno} object.
}
\keyword{internal}
