% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\docType{methods}
\name{makeRawGenoMatrix}
\alias{makeRawGenoMatrix}
\alias{makeRawGenoMatrix,DNAStringSet-method}
\alias{makeRawGenoMatrix,QualityScaledDNAStringSet-method}
\alias{makeRawGenoMatrix.DNAStringSet}
\alias{makeRawGenoMatrix.QualityScaledDNAStringSet}
\title{Make raw genotype matrix from genotype data.}
\usage{
makeRawGenoMatrix(x)

makeRawGenoMatrix.DNAStringSet(x)

makeRawGenoMatrix.QualityScaledDNAStringSet(x)

makeRawGenoMatrix(x)

\S4method{makeRawGenoMatrix}{DNAStringSet}(x)

\S4method{makeRawGenoMatrix}{QualityScaledDNAStringSet}(x)
}
\arguments{
\item{x}{Sample genotype data.}
}
\value{
A genotype matrix, with genotypes encoded as integers and their
corresponding allele symbols in the attribute \code{'alleles'}.
}
\description{
Given input sample genotype data, this function assigns an arbitrary symbol 
at each locus according to the observed raw SNP genotype. So for example, if
the SNVs at a given locus are 'A' and 'C', samples are assigned the genotypes 
'X1' and 'X2', respectively. The mapping of SNV to genotype is performed 
independently for each locus, so a given raw genotype does not have the same
meaning across loci.
}
\keyword{internal}

