% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makeResultsOverview}
\alias{makeResultsOverview}
\title{Make QTL analysis results overview.}
\usage{
makeResultsOverview(phenotypes, analysis, results = NULL)
}
\arguments{
\item{phenotypes}{Vector of phenotype IDs.}

\item{analysis}{Name of analysis from which results were generated.}

\item{results}{Character vector summarising results of analysis. If the
results vector has names, results are matched to phenotypes by name.}
}
\value{
A results overview \code{data.frame}, with two columns:
\code{'Phenotype'}, listing phenotypes for which the analysis was
done; and a second column summarising the results of that analysis.
}
\description{
Make QTL analysis results overview.
}
\keyword{internal}

