% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makeScanoneThreshold}
\alias{makeScanoneThreshold}
\title{Make \code{scanone} threshold object.}
\usage{
makeScanoneThreshold(threshold, alpha = NULL, fdr = NULL)
}
\arguments{
\item{threshold}{LOD threshold value.}

\item{alpha}{Significance level (alpha) associated with the specified LOD
threshold. (Incompatible with \code{fdr}.)}

\item{fdr}{False-discovery rate associated with the specified LOD
threshold. (Incompatible with \code{alpha}.)}
}
\value{
If \code{alpha} is specified, a \code{summary.scanoneperm} object is
returned, with the given threshold at the specified \code{alpha} for the
single LOD column. If \code{fdr} is specified, a \code{summary.scanonebins}
object is returned, with the given threshold at the specified \code{fdr}
for the single LOD column. In any case, because no permutations are used in
generating the threshold, the returned object has attribute \code{'n.perm'}
containing an \code{NA} value.
}
\description{
Make \code{scanone} threshold object.
}
\keyword{internal}

