% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makeScantwoThresholdObject}
\alias{makeScantwoThresholdObject}
\title{Make \code{scantwo} threshold object.}
\usage{
makeScantwoThresholdObject(thresholds, phenotypes, alpha)
}
\arguments{
\item{thresholds}{Mapping or vector of \code{scantwo} LOD threshold values.
If a mapping or named vector, threshold values are accessed by name,
regardless of order. If an unnamed vector, threshold values are accessed
by position, and are assumed to be in the same order as those of a
\code{summary.scantwoperm} object.}

\item{phenotypes}{Character vector of phenotypes for which
the \code{scantwo} threshold object is being created.}

\item{alpha}{Significance level (alpha) associated
with the specified \code{scantwo} LOD thresholds.}
}
\value{
A \code{summary.scantwoperm} object with the given \code{scantwo}
thresholds at the specified significance level (\code{alpha}) for each LOD
column. Because no permutations are used in generating the \code{scantwo}
thresholds object, the returned object has attribute \code{'n.perm'}
containing the value \code{0}.

Note that phenotype names are not included in the \code{summary.scantwoperm}
object. This is a feature, not a bug, and is intended to match the behaviour
of \pkg{R/qtl} in the case of a single significance level (\code{alpha}).
}
\description{
Make \code{scantwo} threshold object.
}
\keyword{internal}

