% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{package}
\name{Map Utilities}
\alias{Map Utilities}
\alias{Map Utilities-package}
\title{Genetic and physical map utilities.}
\description{
These utilities include functions for \pkg{R/qtl} \code{map} objects, 
\pkg{shmootl} \code{mapframe} objects. A \code{mapframe} can be created 
in a similar manner to a \code{data.frame} with \code{\link{mapframe}}.
All arguments are passed to the \code{data.frame} constructor, except for
an addition \code{map.unit} parameter, which can be used to set the map unit
of the new \code{mapframe}. The function \code{\link{gmapframe}} creates a
genetic \code{mapframe} with map unit set automatically to \code{'cM'}.
Coercion functions exist to convert a \code{map} to a \code{mapframe} 
(\code{\link{as.mapframe}}), and vice versa (\code{\link{as.map}}).
}
\section{Map}{
 

The \pkg{R/qtl} \code{map} object - described in detail in Section 2.6 of 
Broman and Sen (2009) - comprises a nonempty list, each element of which 
contains two or more marker positions for a given chromosome. Each list 
element is named for its corresponding chromosome, while the names of each 
chromosome's marker position vector comprise the marker locus IDs.

A \code{map} must represent either a genetic or a physical map. These are 
distinguished by the \code{'map.unit'} attribute of an object, which can
contain either genetic or physical map units.
}

\section{Mapframe}{
 

A \code{mapframe} object is a class of \code{data.frame} that contains mapped
information, or a \code{data.frame} with equivalent properties. The leftmost 
two columns of every \code{mapframe} object must be as follows:

\itemize{
\item{\code{chr}:}   {Chromosome/sequence labels in the order defined for the current 
reference genome.}
\item{\code{pos}:}   {Map positions in increasing order within each 
chromosome/sequence. The map unit used in a \code{mapframe} 
position column must be indicated by setting its 
\code{'map.unit'} attribute.}
}

Any number of additional data columns can be included. If present, the row 
names indicate the locus IDs, which should be unique within the given
\code{mapframe}.
}

\section{Map Units}{
 
A \code{map} or \code{mapframe} must express positions in either genetic or 
physical map units. Genetic map positions must be in centiMorgans (cM), while 
physical map positions must be in base pairs (bp), kilobases (kb), or 
Megabases (Mb).
}

