% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapframe}
\alias{mapframe}
\title{Create a new \code{mapframe}.}
\usage{
mapframe(map.unit = NULL, ...)
}
\arguments{
\item{map.unit}{Map unit for the new \code{mapframe}.}

\item{...}{Further arguments. These are passed to the \code{data.frame}
constructor. All arguments should be passed as keyword arguments.}
}
\value{
A new \code{mapframe}.
}
\description{
Create a new \code{mapframe}.
}
\section{Mapframe}{
 

A \code{mapframe} object is a class of \code{data.frame} that contains mapped
information, or a \code{data.frame} with equivalent properties. The leftmost 
two columns of every \code{mapframe} object must be as follows:

\itemize{
\item{\code{chr}:}   {Chromosome/sequence labels in the order defined for the current 
reference genome.}
\item{\code{pos}:}   {Map positions in increasing order within each 
chromosome/sequence. The map unit used in a \code{mapframe} 
position column must be indicated by setting its 
\code{'map.unit'} attribute.}
}

Any number of additional data columns can be included. If present, the row 
names indicate the locus IDs, which should be unique within the given
\code{mapframe}.
}
\seealso{
Other map utilities: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{inMapOrder}}, \code{\link{inferMapStep}},
  \code{\link{orderMap}}, \code{\link{setMapUnit}}
}

