% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapkey}
\alias{mapkey}
\title{Make new \code{mapkey} rescaling function.}
\usage{
mapkey(...)
}
\arguments{
\item{...}{One or more component maps, all with comparable marker loci.}
}
\value{
A \code{mapkey} function similar to the \pkg{R/qtl} function
\code{interpPositions}, taking a locus \code{mapframe} and map unit
parameter, and returning a \code{mapframe} that has been rescaled in
terms of the given map unit.
}
\description{
This function takes as input one or more maps - one of each supported
map type - and returns a function that can be used to rescale a locus
\code{mapframe} of map positions. The input maps must be comparable,
having the same set of markers on the same chromosomes, although the
marker names may be different. As in \pkg{R/qtl}, map positions are
interpolated if they are within the bounds of the component maps, and
extrapolated otherwise.
}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})
}
\seealso{
Other mapkey functions: \code{\link{mapkeyOpt}}
}
