% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapkeyOpt}
\alias{mapkeyOpt}
\title{Set/get \pkg{shmootl} \code{mapkey} option.}
\usage{
mapkeyOpt(value)
}
\arguments{
\item{value}{New \code{mapkey} function.}
}
\value{
Current \code{mapkey} function.
}
\description{
This function returns the current \pkg{shmootl} \code{mapkey}. If a new
\code{mapkey} is specified, the \pkg{shmootl} \code{mapkey} option is
updated with the given \code{mapkey} function.
}
\seealso{
Other mapkey functions: \code{\link{mapkey}}

Other package option functions: \code{\link{genomeOpt}}
}

