% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mappingKeys}
\alias{mappingKeys}
\title{Get keys of a \code{mapping} object.}
\usage{
mappingKeys(x)
}
\arguments{
\item{x}{A \code{mapping} object.}
}
\value{
Character vector of keys in a \code{mapping} object.
}
\description{
Get keys of a \code{mapping} object.
}
\seealso{
Other mapping object functions: \code{\link{as.mapping}},
  \code{\link{is.mapping}}, \code{\link{mappingValues}},
  \code{\link{mapping}}
}
\keyword{internal}
