% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapsEqual}
\alias{mapsEqual}
\title{Test if two \pkg{R/qtl} \code{map} objects are essentially equal.}
\usage{
mapsEqual(map1, map2, tol = 1e-05)
}
\arguments{
\item{map1}{An \pkg{R/qtl} \code{map} object.}

\item{map2}{An \pkg{R/qtl} \code{map} object.}

\item{tol}{Numeric tolerance for comparing map positions.}
}
\value{
Returns \code{TRUE} if the two input \code{map} objects are identical
within the given tolerance. Otherwise, this function returns a character
vector listing observed differences between the two \code{map} objects.
}
\description{
This is based on the \pkg{R/qtl} function \code{comparecrosses},
and checks that two input \code{map} objects are essentially equal, having
identical reference sequences, marker names, map unit, and map positions
(within the given numeric tolerance).
}
\seealso{
\href{http://www.rqtl.org/manual/qtl-manual.pdf}{R/qtl manual}

Other map utility functions: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{hasMapCSV}}, \code{\link{inMapOrder}},
  \code{\link{inferMapStep}}, \code{\link{mapframe}},
  \code{\link{orderMap}}, \code{\link{readMapCSV}},
  \code{\link{readMapHDF5}}, \code{\link{readMapframeCSV}},
  \code{\link{setMapUnit}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapHDF5}},
  \code{\link{writeMapframeCSV}}
}

