% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{matchLoci}
\alias{matchLoci}
\title{Get object loci matching those specified.}
\usage{
matchLoci(x, loc)
}
\arguments{
\item{x}{Object containing map data.}

\item{loc}{Locus \code{mapframe} specifying map positions.}
}
\value{
Locus mapframe of those loci specified that were found in the target
object.
}
\description{
Get object loci matching those specified.
}
\keyword{internal}

