% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{matchLocusRowIndices}
\alias{matchLocusRowIndices}
\title{Find matching locus row indices.}
\usage{
matchLocusRowIndices(x, loc)
}
\arguments{
\item{x}{A \code{mapframe} or \code{scanone} object.}

\item{loc}{Locus \code{mapframe} specifying map positions.}
}
\value{
Vector of row indices with respect to the target \code{mapframe},
for which a matching locus exists in the locus mapframe.
}
\description{
Find matching locus row indices.
}
\keyword{internal}

