% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{matchSeqRowIndices}
\alias{matchSeqRowIndices}
\title{Find row indices of matching sequences.}
\usage{
matchSeqRowIndices(x, sequences, simplify = FALSE)
}
\arguments{
\item{x}{A \code{mapframe} object.}

\item{sequences}{Sequences for which row indices should be returned.}
}
\value{
Vector of row indices corresponding to the specified sequences. An
empty vector is returned if no sequences match the input \code{mapframe}.
}
\description{
Find row indices of matching sequences.
}
\keyword{internal}

