% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonebins.R
\name{mergeLODBinLabels}
\alias{mergeLODBinLabels}
\title{Merge multiple vectors of LOD bin labels.}
\usage{
mergeLODBinLabels(...)
}
\arguments{
\item{...}{One or more character vectors containing LOD bin labels.}
}
\value{
Merged bin label vector.
}
\description{
Merge multiple vectors of LOD bin labels.
}
\keyword{internal}

