% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScantwo.R
\name{nodePermScantwo}
\alias{nodePermScantwo}
\title{Run \code{qtl::scantwo} on a single permuted \code{cross} object.}
\usage{
nodePermScantwo(perm.id, cross, pheno.col = NULL, perm.pheno = TRUE,
  perm.geno = FALSE, ...)
}
\arguments{
\item{perm.id}{Permutation index.}

\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{pheno.col}{Phenotype columns for which QTL analysis should be run. 
If no phenotypes are specified, all are used.}

\item{perm.pheno}{Permute phenotype data.}

\item{perm.geno}{Permute genotype data.}

\item{...}{Additional keyword arguments passed to \code{scantwo}.}
}
\value{
A \code{scantwoperm} matrix containing the result of the QTL scan for
a single permutation.
}
\description{
Run \code{qtl::scantwo} on a single permuted \code{cross} object.
}
\seealso{
Other scan utilities: \code{\link{batchPermScanoneF}},
  \code{\link{batchPermScanone}},
  \code{\link{batchPermScantwoF}},
  \code{\link{batchPermScantwo}},
  \code{\link{batchPhenoScanone}},
  \code{\link{batchPhenoScantwo}}, \code{\link{batchScan}},
  \code{\link{nodePermScanoneF}},
  \code{\link{nodePermScanone}},
  \code{\link{nodePermScantwoF}},
  \code{\link{nodePhenoScanone}},
  \code{\link{nodePhenoScantwo}}
}

