% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScantwoF.R
\name{nodePermScantwoF}
\alias{nodePermScantwoF}
\title{Run \code{funqtl::scantwoF} on a single permutation of a \code{cross} object.}
\usage{
nodePermScantwoF(perm.id, cross, pheno.col = NULL, perm.pheno = TRUE,
  perm.geno = FALSE, ...)
}
\arguments{
\item{perm.id}{Permutation index.}

\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{pheno.col}{Phenotype columns for which QTL analysis should be 
performed. If specified, phenotypes should contain measurements for 
consecutive values of the parameter of the function-valued trait 
(e.g. times). If no phenotypes are specified, all are used.}

\item{perm.pheno}{Permute phenotype data.}

\item{perm.geno}{Permute genotype data.}

\item{...}{Additional keyword arguments passed to \code{scantwoF}.}
}
\value{
A \code{scantwoperm} list containing the results of \code{scantwoF}
for a single permutation.
}
\description{
Run \code{funqtl::scantwoF} on a single permutation of a \code{cross} object.
}
\seealso{
Other scan utilities: \code{\link{batchPermScanoneF}},
  \code{\link{batchPermScanone}},
  \code{\link{batchPermScantwoF}},
  \code{\link{batchPermScantwo}},
  \code{\link{batchPhenoScanone}},
  \code{\link{batchPhenoScantwo}}, \code{\link{batchScan}},
  \code{\link{nodePermScanoneF}},
  \code{\link{nodePermScanone}},
  \code{\link{nodePermScantwo}},
  \code{\link{nodePhenoScanone}},
  \code{\link{nodePhenoScantwo}}
}

