% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScanone.R
\name{nodePhenoScanone}
\alias{nodePhenoScanone}
\title{Run \code{qtl::scanone} on a single phenotype of a \code{cross} object.}
\usage{
nodePhenoScanone(pheno.col, cross, ...)
}
\arguments{
\item{pheno.col}{Phenotype column for which QTL analysis should be run.}

\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{...}{Additional keyword arguments passed to \code{scanone}.}
}
\value{
A \code{scanone} object containing the result of \code{scanone}
for a single phenotype.
}
\description{
Run \code{qtl::scanone} on a single phenotype of a \code{cross} object.
}
\seealso{
Other scan utilities: \code{\link{batchPermScanoneF}},
  \code{\link{batchPermScanone}},
  \code{\link{batchPermScantwoF}},
  \code{\link{batchPermScantwo}},
  \code{\link{batchPhenoScanone}},
  \code{\link{batchPhenoScantwo}}, \code{\link{batchScan}},
  \code{\link{nodePermScanoneF}},
  \code{\link{nodePermScanone}},
  \code{\link{nodePermScantwoF}},
  \code{\link{nodePermScantwo}},
  \code{\link{nodePhenoScantwo}}
}

