% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{normChr}
\alias{normChr}
\title{Normalise chromosome labels.}
\usage{
normChr(x)
}
\arguments{
\item{x}{Vector of chromosome labels.}
}
\value{
Vector of normalised chromosome labels.
}
\description{
Normalise chromosome labels.
}
\keyword{internal}

