% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{normSeq}
\alias{normSeq}
\alias{normSeq.character}
\alias{normSeq.data.frame}
\alias{normSeq.factor}
\alias{normSeq.integer}
\alias{normSeq.map}
\alias{normSeq.numeric}
\title{Normalise sequence labels.}
\usage{
normSeq(x)

\method{normSeq}{character}(x)

\method{normSeq}{data.frame}(x)

\method{normSeq}{factor}(x)

\method{normSeq}{integer}(x)

\method{normSeq}{map}(x)

\method{normSeq}{numeric}(x)
}
\arguments{
\item{x}{Object containing sequence labels.}
}
\value{
Input object with normalised sequence labels.
}
\description{
Normalise sequence labels.
}

