% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{orderChr}
\alias{orderChr}
\title{Order chromosome labels.}
\usage{
orderChr(x)
}
\arguments{
\item{x}{Vector of chromosome labels.}
}
\value{
Vector of indices for the input chromosome labels,
ordered with respect to their normalised form.
}
\description{
Order chromosome labels.
}
\seealso{
Other chromosome/sequence functions: \code{\link{formatChr}},
  \code{\link{formatSeq}}, \code{\link{isNormChr}},
  \code{\link{isNormSeq}}, \code{\link{normChr}},
  \code{\link{normSeq}}, \code{\link{orderSeq}},
  \code{\link{rankChr}}, \code{\link{rankSeq}},
  \code{\link{sortChr}}, \code{\link{sortSeq}}
}

