% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{orderMap}
\alias{orderMap}
\alias{orderMap.data.frame}
\alias{orderMap.map}
\title{Put object in map order.}
\usage{
orderMap(x)

\method{orderMap}{data.frame}(x)

\method{orderMap}{map}(x)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
Input object ordered by the normalised form of
the reference sequence, then numerically by map position.
}
\description{
Put object in map order.
}
\seealso{
Other map utility functions: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{hasMapCSV}}, \code{\link{inMapOrder}},
  \code{\link{inferMapStep}}, \code{\link{mapframe}},
  \code{\link{mapsEqual}}, \code{\link{readMapCSV}},
  \code{\link{readMapHDF5}}, \code{\link{readMapframeCSV}},
  \code{\link{setMapUnit}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapHDF5}},
  \code{\link{writeMapframeCSV}}
}

