% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{orderSeq}
\alias{orderSeq}
\title{Order sequence labels.}
\usage{
orderSeq(x)
}
\arguments{
\item{x}{Vector of sequence labels.}
}
\value{
Vector of ordered indices for the input sequence labels.
}
\description{
Order sequence labels.
}

