% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{otherattributes}
\alias{otherattributes}
\alias{otherattributes<-}
\title{Get non-reserved object attributes.}
\usage{
otherattributes(x)

otherattributes(x) <- value
}
\arguments{
\item{x}{R object.}

\item{value}{List of attributes to set.}

\item{x}{R object.}
}
\value{
List of non-reserved object attributes.

R object with non-reserved attributes set.
}
\description{
Get non-reserved object attributes.

Set non-reserved object attributes.
}
\details{
As with R base function \code{mostattributes}, this provides access
to non-reserved attributes, but with simpler behaviour.

As with R base function \code{mostattributes}, this provides access
to non-reserved attributes, but with simpler behaviour.
}
\keyword{internal}

