% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{padTimeSeries}
\alias{padTimeSeries}
\title{Pad gaps in a time-series.}
\usage{
padTimeSeries(cross, tol = 1e-05)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{tol}{Tolerance for time step equality.}
}
\value{
The input \code{cross} object is returned with all gaps in 
time-series phenotypes padded with NA values.
}
\description{
Pad gaps in a time-series.
}
\seealso{
Other cross utilities: \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}}, \code{\link{permCross}},
  \code{\link{pull.alleles}}, \code{\link{pull.chr}},
  \code{\link{pull.crosstype}}, \code{\link{pull.ind}}
}

