% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parseDefaultMarkerID}
\alias{parseDefaultMarkerID}
\title{Parse a default marker ID.}
\usage{
parseDefaultMarkerID(marker.id)
}
\arguments{
\item{marker.id}{Marker ID to parse.}
}
\value{
Physical \code{mapframe} with locus corresponding to the specified 
marker ID. Returns \code{NULL} if marker ID cannot be parsed as a default 
marker ID.
}
\description{
Parse a default marker ID.
}
\keyword{internal}

