% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parseDefaultMarkerIDs}
\alias{parseDefaultMarkerIDs}
\title{Parse default marker IDs.}
\usage{
parseDefaultMarkerIDs(marker.ids)
}
\arguments{
\item{marker.ids}{Vector of default marker IDs.}
}
\value{
Physical \code{mapframe} with loci corresponding to the specified
marker IDs. Raises an error if any of the input values cannot be parsed as
a default marker ID.
}
\description{
Parse default marker IDs.
}

