% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parseDefaultQTLName}
\alias{parseDefaultQTLName}
\title{Parse a default QTL name.}
\usage{
parseDefaultQTLName(qtl.name)
}
\arguments{
\item{qtl.name}{QTL name to parse.}
}
\value{
Genetic \code{mapframe} with locus corresponding to the specified QTL
name. Returns \code{NULL} if QTL name cannot be parsed as a default QTL name.
}
\description{
Parse a default QTL name.
}
\keyword{internal}

