% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parsePseudomarkerID}
\alias{parsePseudomarkerID}
\title{Parse a pseudomarker ID.}
\usage{
parsePseudomarkerID(loc.id)
}
\arguments{
\item{loc.id}{Locus ID to parse.}
}
\value{
Genetic \code{mapframe} with locus corresponding to the pseudomarker 
ID. Returns \code{NULL} if locus ID cannot be parsed as a pseudomarker ID.
}
\description{
Parse a pseudomarker ID.
}
\keyword{internal}

