% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parsePseudomarkerIDs}
\alias{parsePseudomarkerIDs}
\title{Parse pseudomarker IDs.}
\usage{
parsePseudomarkerIDs(loc.ids)
}
\arguments{
\item{loc.ids}{Vector of pseudomarker IDs.}
}
\value{
Genetic \code{mapframe} with loci corresponding to the specified
pseudomarker IDs. Raises an error if any input values cannot be parsed as
a pseudomarker ID.
}
\description{
Parse pseudomarker IDs.
}

