% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{parseWorksheetNames}
\alias{parseWorksheetNames}
\title{Parse \pkg{shmootl} output worksheet names.}
\usage{
parseWorksheetNames(worksheets)
}
\arguments{
\item{worksheets}{Vector of \pkg{shmootl} output worksheet names.}
}
\value{
A list of character vectors in which the names of the list are
analyses, and the elements of each character vector are results for the
given analysis.
}
\description{
Parse \pkg{shmootl} output worksheet names.
}
\keyword{internal}
