% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5Stack-class.R
\docType{methods}
\name{peek}
\alias{peek}
\alias{peek,H5Stack-method}
\title{Get HDF5 object ID at top of \code{H5Stack}.}
\usage{
peek(h5stack)

\S4method{peek}{H5Stack}(h5stack)
}
\arguments{
\item{h5stack}{A \code{H5Stack} object.}
}
\value{
HDF5 object ID at top of \code{H5Stack}. Returns \code{NULL} if stack
is empty.
}
\description{
Get HDF5 object ID at top of \code{H5Stack}.
}
\keyword{internal}
