% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{permCross}
\alias{permCross}
\title{Permute \code{cross} phenotype or genotype data.}
\usage{
permCross(cross, perm.indices = NULL, perm.pheno = TRUE,
  perm.geno = FALSE)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{perm.indices}{Permutation indices.}

\item{perm.pheno}{Permute phenotype data.}

\item{perm.geno}{Permute genotype data.}
}
\value{
Permuted \code{cross} object.
}
\description{
Permute \code{cross} phenotype or genotype data.
}
\details{
Given an input \code{cross} object, this function permutes either
the phenotype or genotype data (but not both), and returns the permuted 
\code{cross} object. Note that when permuting phenotypes, any covariates
must also be permuted in the same way.
}
\seealso{
Other cross utilities: \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{pull.alleles}},
  \code{\link{pull.chr}}, \code{\link{pull.ind}}
}

