% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{permIndices}
\alias{permIndices}
\title{Generate permutation indices for a \code{cross} object.}
\usage{
permIndices(cross)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}
}
\value{
Vector of sample indices that can be used to permute a 
\code{cross} object or associated data.
}
\description{
Generate permutation indices for a \code{cross} object.
}

