% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQTLIntervals.R
\name{plotQTLIntervals}
\alias{plotQTLIntervals}
\alias{plotQTLIntervals.qtlintervals}
\alias{plotQTLIntervals.scanone}
\title{Plot QTL intervals.}
\usage{
## Generic method.
plotQTLIntervals(x, ...)

\method{plotQTLIntervals}{qtlintervals}(x, lod.profile, chr = NULL,
  lodcolumn = NULL, phenotype = NULL, ...)

\method{plotQTLIntervals}{scanone}(x, chr = NULL, drop = 1.5,
  expandtomarkers = FALSE, threshold = NULL, alpha = NULL,
  lodcolumn = NULL, phenotype = NULL, ...)
}
\arguments{
\item{x}{A \code{scanone} or \code{qtlintervals} object.}

\item{...}{Further arguments (see below).}

\item{lod.profile}{For a \code{qtlintervals} object, the \code{scanone}
object from which the QTL intervals were obtained.}

\item{chr}{Vector indicating which sequences to plot. If no sequences are
specified, all sequences are plotted.}

\item{lodcolumn}{In a \code{scanone} or equivalent object, this parameter 
indicates which LOD column to consider. This must be either a column name 
or an index \emph{with respect to the set of LOD columns}. If no LOD column 
is specified and one such column is found, that column is used by default; 
otherwise a LOD column must be specified.}

\item{phenotype}{Name of the phenotype, to be used in plot title.}

\item{drop}{For a \code{scanone} object, LOD units that the LOD profile must
drop to form a QTL interval.}

\item{expandtomarkers}{For a \code{scanone} object, expand the LOD interval
to the nearest flanking markers, or to the respective terminal loci.}

\item{threshold}{For a \code{scanone} object, this indicates 
the LOD significance threshold for identifying QTL intervals.}

\item{alpha}{For a \code{scanone} object, this contains
the significance level of the given threshold.}
}
\description{
Plot QTL intervals.
}

