% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQTLScanone.R
\name{plotQTLScanone}
\alias{plotQTLScanone}
\title{Plot \pkg{R/qtl} \code{scanone} result.}
\usage{
plotQTLScanone(x, chr = NULL, lodcolumn = NULL, qtl.intervals = NULL,
  threshold = NULL, col = c("gray10", "gray60"), gap = 25,
  phenotype = NULL, type = NULL, ...)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{scanone} object.}

\item{chr}{Vector indicating which sequences to plot. If no sequences are
specified, all are plotted.}

\item{lodcolumn}{In a \code{scanone} or equivalent object, this
parameter indicates which LOD column to consider. In other objects
with LOD-column-associated elements, this parameter indicates which
LOD-column-associated element to consider. This must be either a LOD
column name or an index \emph{with respect to the set of LOD columns}.
If no LOD column is specified and one such column is found, that column
is used by default; otherwise a LOD column must be specified.}

\item{qtl.intervals}{A \code{qtlintervals} object created from the same data
as the \code{scanone} object. These intervals are added to the plot, unless
the median interval size is too small to be distinguishable. Threshold
information is taken from this object, if available.}

\item{threshold}{A single \code{numeric} LOD significance threshold, or an
object (e.g. \code{summary.scanoneperm}) containing one such threshold and
its associated significance level.}

\item{col}{Analogous to the standard \code{'col'} plotting parameter. This is
recycled to match the number of sequences being plotted. As in the package
\pkg{qqman}, this defaults to two alternating monochrome shades.}

\item{gap}{Gap (in cM) between chromosomes in multi-chromosome plots.}

\item{phenotype}{Name of the phenotype, to be used in plot title.}

\item{type}{Type of plot. Set to \code{'l'} to output a stanard LOD curve, or
to \code{'p'} for a Manhattan plot of individual LOD scores.}

\item{...}{Unused arguments.}
}
\description{
Plotting function based on \pkg{R/qtl} \code{plot.scanone} (Broman
\emph{et al.} 2003) and \pkg{qqman} \code{manhattan} (Turner 2014).
Plots a \code{scanone} result as a LOD curve (with \code{type} set
to \code{'l'}) or as a Manhattan plot (with \code{type} set to
\code{'p'}). If no plot type is specified, this is set automatically
based on marker density.
}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})

Turner SD (2014) qqman: an R package for visualizing GWAS results
using Q-Q and manhattan plots.
(\href{http://dx.doi.org/10.1101/005165}{bioRXiv})
}
\seealso{
Other QTL functions: \code{\link{estPhysicalPositions}},
  \code{\link{getQTLFeatures}},
  \code{\link{getQTLIntervals}}, \code{\link{getQTLPeaks}}
}

