% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQTLScantwo.R
\name{plotQTLScantwo}
\alias{plotQTLScantwo}
\title{Plot \pkg{R/qtl} \code{scantwo} result.}
\usage{
plotQTLScantwo(x, chr = NULL, incl.markers = FALSE, lodcolumn = NULL,
  lower = c("full", "add", "cond-int", "cond-add", "int"), upper = c("int",
  "cond-add", "cond-int", "add", "full"), zscale = TRUE,
  palette = c("viridis", "redblue", "cm", "gray", "heat", "terrain", "topo"),
  phenotype = NULL, ...)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{scantwo} object.}

\item{chr}{Vector indicating which sequences to plot.
If none are specified, all are plotted.}

\item{incl.markers}{When \code{scantwo} has been run with a fixed non-zero
step width, this indicates if markers (as distinct from pseudomarkers)
should be included in the plot.}

\item{lodcolumn}{In a \code{scanone} or equivalent object, this
parameter indicates which LOD column to consider. In other objects
with LOD-column-associated elements, this parameter indicates which
LOD-column-associated element to consider. This must be either a LOD
column name or an index \emph{with respect to the set of LOD columns}.
If no LOD column is specified and one such column is found, that column
is used by default; otherwise a LOD column must be specified.}

\item{lower}{The \code{scantwo} LOD type to
output in the lower triangle of the plot.}

\item{upper}{The \code{scantwo} LOD type to
output in the upper triangle of the plot.}

\item{zscale}{Display a z-scale in the right margin of the plot.}

\item{palette}{Palette to use in the plot. This is equivalent to the
\pkg{R/qtl} \code{plot.scantwo} parameter \code{'col.scheme'}.}

\item{phenotype}{Name of the phenotype, to be shown in plot information.}

\item{...}{Unused arguments.}
}
\description{
This function is a restricted version of \pkg{R/qtl} function
\code{plot.scantwo} (Broman \emph{et al.} 2003), plotting the
input \code{scantwo} result as a heatmap. The set of palettes
are the same as in \pkg{R/qtl} \code{plot.scantwo}, except for
\code{'viridis'} which uses the recently-developed \code{R}
package of the same name.
}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})
}
\seealso{
Function \code{plot.scantwo} in the
\href{http://www.rqtl.org/manual/qtl-manual.pdf}{R/qtl manual}.

Other plot functions: \code{\link{plotQTLScanone}}
}

