% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{plotReportTitlePagePDF}
\alias{plotReportTitlePagePDF}
\title{Plot a QTL analysis report title page.}
\usage{
plotReportTitlePagePDF(scanfile)
}
\arguments{
\item{scanfile}{File path of HDF5 scan file
from which the report is being generated.}
}
\description{
Plot a QTL analysis report title page.
}
\keyword{internal}

