% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5Stack-class.R
\docType{methods}
\name{pop}
\alias{pop}
\alias{pop,H5Stack-method}
\title{Remove HDF5 object ID at top of \code{H5Stack}.}
\usage{
pop(h5stack)

\S4method{pop}{H5Stack}(h5stack)
}
\arguments{
\item{h5stack}{A \code{H5Stack} object.}
}
\value{
Input \code{H5Stack} object with top element removed.
}
\description{
Remove HDF5 object ID at top of \code{H5Stack}.
}
\keyword{internal}
