% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{prepPipelineArgparser}
\alias{prepPipelineArgparser}
\title{Prep argument parser for \pkg{shmootl} pipeline.}
\usage{
prepPipelineArgparser(pipeline)
}
\arguments{
\item{pipeline}{Name of a \pkg{shmootl} pipeline.}
}
\value{
Argument parser for the specified \pkg{shmootl} pipeline.
}
\description{
Prep argument parser for \pkg{shmootl} pipeline.
}
\keyword{internal}

