% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlintervals.R
\name{print.qtlintervals}
\alias{print.qtlintervals}
\title{Print a \code{qtlintervals} object.}
\usage{
\method{print}{qtlintervals}(qtl.intervals)
}
\arguments{
\item{qtl.intervals}{A \code{qtlintervals} object.}
}
\description{
Print a \code{qtlintervals} object.
}
