% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonebins.R
\name{print.summary.scanonebins}
\alias{print.summary.scanonebins}
\title{Print \code{summary.scanonebins} object.}
\usage{
\method{print}{summary.scanonebins}(x, ...)
}
\arguments{
\item{x}{A \code{summary.scanonebins} object.}

\item{...}{Unused arguments.}
}
\description{
Print \code{summary.scanonebins} object.
}
\keyword{internal}
