% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{procPipelineArgs}
\alias{procPipelineArgs}
\title{Process arguments parsed by \pkg{argparser}.}
\usage{
procPipelineArgs(ap, args)
}
\arguments{
\item{ap}{An \pkg{argparser} parser object.}

\item{args}{Arguments to be processed, having been parsed by \pkg{argparser}.}
}
\value{
Processed pipeline arguments.
}
\description{
Process arguments parsed by \pkg{argparser}.
}
\keyword{internal}

