% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pull.alleles}
\alias{pull.alleles}
\alias{pull.alleles.cross}
\alias{pull.alleles.geno}
\title{Pull alleles from the given object.}
\usage{
pull.alleles(x)

\method{pull.alleles}{cross}(x)

\method{pull.alleles}{geno}(x)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{cross} object,
or other object with alleles.}
}
\value{
Vector of alleles in the given object.
}
\description{
Pull alleles from the given object.
}
\seealso{
Other cross object functions: \code{\link{crossesEqual}},
  \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{hasTimeSeriesPhenotypes}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{permIndices}}, \code{\link{pull.chr}},
  \code{\link{pull.crosstype}}, \code{\link{pull.ind}},
  \code{\link{readCrossCSV}}, \code{\link{readCrossHDF5}},
  \code{\link{writeCrossCSV}}, \code{\link{writeCrossHDF5}}
}
