% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{pull.chr}
\alias{pull.chr}
\title{Pull chromosomes/sequences from \pkg{R/qtl} \code{cross}.}
\usage{
pull.chr(cross)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}
}
\value{
Vector of chromosomes/sequences in the \code{cross} object.
}
\description{
Pull chromosomes/sequences from \pkg{R/qtl} \code{cross}.
}
\seealso{
Other cross object functions: \code{\link{crossesEqual}},
  \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{hasTimeSeriesPhenotypes}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{permIndices}}, \code{\link{pull.alleles}},
  \code{\link{pull.crosstype}}, \code{\link{pull.ind}},
  \code{\link{readCrossCSV}}, \code{\link{readCrossHDF5}},
  \code{\link{writeCrossCSV}}, \code{\link{writeCrossHDF5}}
}

