% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{pull.crosstype}
\alias{pull.crosstype}
\title{Pull cross type from \pkg{R/qtl} \code{cross}.}
\usage{
pull.crosstype(cross)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}
}
\value{
Cross type of the \code{cross} object.
}
\description{
Pull cross type from \pkg{R/qtl} \code{cross}.
}
\seealso{
Other cross utilities: \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{pull.alleles}}, \code{\link{pull.chr}},
  \code{\link{pull.ind}}
}

