% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pull.ind}
\alias{pull.ind}
\alias{pull.ind.cross}
\alias{pull.ind.geno}
\alias{pull.ind.pheno}
\title{Pull individual sample IDs from the given object.}
\usage{
pull.ind(x)

\method{pull.ind}{cross}(x)

\method{pull.ind}{geno}(x)

\method{pull.ind}{pheno}(x)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{cross} object,
or other object with sample ID information.}
}
\value{
Vector of individual sample IDs in the given object.
Returns \code{NULL} if no sample IDs are found.
}
\description{
Pull individual sample IDs from the given object.
}
\seealso{
Other cross object functions: \code{\link{crossesEqual}},
  \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{hasTimeSeriesPhenotypes}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{permIndices}}, \code{\link{pull.alleles}},
  \code{\link{pull.chr}}, \code{\link{pull.crosstype}},
  \code{\link{readCrossCSV}}, \code{\link{readCrossHDF5}},
  \code{\link{writeCrossCSV}}, \code{\link{writeCrossHDF5}}
}
