% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pullLoci}
\alias{pullLoci}
\title{Pull loci from object.}
\usage{
pullLoci(x)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
Locus mapframe of individual loci.
}
\description{
Pull loci from object.
}
\keyword{internal}
