% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pullLocusIDs}
\alias{pullLocusIDs}
\alias{pullLocusIDs.data.frame}
\alias{pullLocusIDs.list}
\alias{pullLocusIDs.map}
\title{Pull individual locus IDs.}
\usage{
pullLocusIDs(x)

\method{pullLocusIDs}{data.frame}(x)

\method{pullLocusIDs}{list}(x)

\method{pullLocusIDs}{map}(x)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
Character vector of individual locus IDs.
}
\description{
Pull individual locus IDs.
}
\keyword{internal}
