% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pullLocusPos}
\alias{pullLocusPos}
\alias{pullLocusPos.data.frame}
\alias{pullLocusPos.list}
\alias{pullLocusPos.map}
\alias{pullLocusPos.mapframe}
\title{Pull individual locus positions.}
\usage{
pullLocusPos(x)

\method{pullLocusPos}{data.frame}(x)

\method{pullLocusPos}{list}(x)

\method{pullLocusPos}{map}(x)

\method{pullLocusPos}{mapframe}(x)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
Numeric vector of individual locus positions.
}
\description{
Pull individual locus positions.
}
\keyword{internal}
