% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pullLocusSeq}
\alias{pullLocusSeq}
\alias{pullLocusSeq.data.frame}
\alias{pullLocusSeq.list}
\alias{pullLocusSeq.map}
\title{Pull sequence labels for individual loci.}
\usage{
pullLocusSeq(x)

\method{pullLocusSeq}{data.frame}(x)

\method{pullLocusSeq}{list}(x)

\method{pullLocusSeq}{map}(x)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
Character vector of sequence labels associated with individual loci.
}
\description{
Pull sequence labels for individual loci.
}
\keyword{internal}
