% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pullMap}
\alias{pullMap}
\alias{pullMap.cross}
\alias{pullMap.geno}
\title{Pull map from object.}
\usage{
pullMap(x)

\method{pullMap}{cross}(x)

\method{pullMap}{geno}(x)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
An \pkg{R/qtl} \code{map} extracted from the input object.
}
\description{
Pull map from object.
}
\keyword{internal}
