% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5Stack-class.R
\docType{methods}
\name{push}
\alias{push}
\alias{push,H5Stack-method}
\title{Push HDF5 object ID onto top of \code{H5Stack}.}
\usage{
push(h5stack, h5obj)

\S4method{push}{H5Stack}(h5stack, h5obj)
}
\arguments{
\item{h5stack}{A \code{H5Stack} object.}

\item{h5obj}{A HDF5 object ID.}
}
\value{
Input \code{H5Stack} object with new element added.
}
\description{
Push HDF5 object ID onto top of \code{H5Stack}.
}
\keyword{internal}
