% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pushLoci}
\alias{pushLoci}
\title{Replace loci of object.}
\usage{
pushLoci(x, loc)
}
\arguments{
\item{x}{Object containing map data.}
}
\value{
loc Locus mapframe of individual loci.

Input object with individual loci replaced.
}
\description{
Replace loci of object.
}
\keyword{internal}

