% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pushLocusIDs}
\alias{pushLocusIDs}
\alias{pushLocusIDs.data.frame}
\alias{pushLocusIDs.list}
\alias{pushLocusIDs.map}
\title{Replace individual locus IDs.}
\usage{
pushLocusIDs(x, value)

\method{pushLocusIDs}{data.frame}(x, value)

\method{pushLocusIDs}{list}(x, value)

\method{pushLocusIDs}{map}(x, value)
}
\arguments{
\item{x}{Object containing map data.}

\item{value}{Character vector of individual locus IDs.}
}
\value{
Input object with individual locus IDs replaced.
}
\description{
Replace individual locus IDs.
}
\keyword{internal}
