% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pushLocusPos}
\alias{pushLocusPos}
\alias{pushLocusPos.data.frame}
\alias{pushLocusPos.list}
\alias{pushLocusPos.map}
\title{Replace individual locus positions.}
\usage{
pushLocusPos(x, value)

\method{pushLocusPos}{data.frame}(x, value)

\method{pushLocusPos}{list}(x, value)

\method{pushLocusPos}{map}(x, value)
}
\arguments{
\item{x}{Object containing map data.}

\item{value}{Numeric vector of individual locus positions.}
}
\value{
Input object with individual locus positions replaced.
}
\description{
Replace individual locus positions.
}
\keyword{internal}

