% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pushLocusSeq}
\alias{pushLocusSeq}
\alias{pushLocusSeq.data.frame}
\alias{pushLocusSeq.list}
\alias{pushLocusSeq.map}
\title{Replace sequence labels of individual loci.}
\usage{
pushLocusSeq(x, value)

\method{pushLocusSeq}{data.frame}(x, value)

\method{pushLocusSeq}{list}(x, value)

\method{pushLocusSeq}{map}(x, value)
}
\arguments{
\item{x}{Object containing map data.}

\item{value}{Character vector of sequence labels for individual loci.}
}
\value{
Input object with individual locus sequence labels replaced.
}
\description{
Replace sequence labels of individual loci.
}
\keyword{internal}
