% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pushMap}
\alias{pushMap}
\alias{pushMap.geno}
\title{Push map into object.}
\usage{
pushMap(x, map)

\method{pushMap}{geno}(x, map)
}
\arguments{
\item{x}{Object that can contain map data.}

\item{map}{An \pkg{R/qtl} \code{map} object.}
}
\value{
Input object incorporating the given map data.
}
\description{
Push map into object.
}
\keyword{internal}
