% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlintervals.R
\name{qtlintervals}
\alias{qtlintervals}
\title{Create a \code{qtlintervals} object.}
\usage{
qtlintervals(threshold = NULL, drop = NULL, prob = NULL, ...)
}
\arguments{
\item{threshold}{A single \code{numeric} LOD significance threshold, or an
object (e.g. \code{summary.scanoneperm}) containing one such threshold and
its associated significance level.}

\item{drop}{LOD units that the LOD profile must drop to form a QTL interval.
This should only be specified if the QTL intervals are LOD support intervals,
and is incompatible with the \code{prob} parameter.}

\item{prob}{The probability coverage of the Bayesian credible interval.
This should only be specified if the QTL intervals are Bayesian credible
intervals, and is incompatible with the \code{drop} parameter.}

\item{...}{Unused arguments.}
}
\value{
An empty \code{qtlintervals} list with the given attributes.
}
\description{
This function creates a \code{qtlintervals} object. Currently it
can only create a \code{qtlintervals} object of length zero.
}
\author{
Thomas Walsh <tw164@protonmail.com>

Yue Hu
}
\keyword{internal}
