% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{rankChr}
\alias{rankChr}
\title{Rank chromosome labels.}
\usage{
rankChr(x)
}
\arguments{
\item{x}{Vector of chromosome labels.}
}
\value{
Vector of ranks for the normalised
form of the input chromosome labels.
}
\description{
Rank chromosome labels.
}
\seealso{
Other chromosome/sequence functions: \code{\link{formatChr}},
  \code{\link{formatSeq}}, \code{\link{isNormChr}},
  \code{\link{isNormSeq}}, \code{\link{normChr}},
  \code{\link{normSeq}}, \code{\link{orderChr}},
  \code{\link{orderSeq}}, \code{\link{rankSeq}},
  \code{\link{sortChr}}, \code{\link{sortSeq}}
}
