% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{rankSeq}
\alias{rankSeq}
\title{Rank sequence labels.}
\usage{
rankSeq(x)
}
\arguments{
\item{x}{Vector of sequence labels.}
}
\value{
Vector of ranks for the input sequence labels.
}
\description{
Rank sequence labels.
}

