% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readClassHDF5}
\alias{readClassHDF5}
\alias{readGroupMemberNamesHDF5}
\title{Read \code{'class'} attribute of a HDF5 object.}
\usage{
readClassHDF5(infile, h5name)

readGroupMemberNamesHDF5(infile, h5name)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{h5name}{HDF5 object name.}

\item{infile}{An input HDF5 file.}

\item{h5name}{HDF5 group name.}
}
\value{
HDF5 object class.

Vector of group member names. Returns \code{NULL} if group has no 
members.
}
\description{
Read \code{'class'} attribute of a HDF5 object.

Read HDF5 group member names.
}
\keyword{internal}

