% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{readCrossCSV}
\alias{readCrossCSV}
\title{Read yeast \code{cross} from a CSV file.}
\usage{
readCrossCSV(infile, missing.value = "-", error.prob = 1e-04,
  map.function = c("haldane", "kosambi", "c-f", "morgan"))
}
\arguments{
\item{infile}{Input CSV file path.}

\item{missing.value}{Missing data value. This can be any single character 
that is not a possible phenotype or genotype value.}

\item{error.prob}{Genotyping error rate (ignored unless estimating genetic map).}

\item{map.function}{Genetic map function (ignored unless estimating genetic map).}
}
\value{
An \pkg{R/qtl} \code{cross} object with an attribute \code{'info'} of 
type \code{CrossInfo}.
}
\description{
This function reads yeast cross data from an \pkg{R/qtl} CSV file and returns 
an \pkg{R/qtl} \code{cross} object with an attribute \code{'info'} of type 
\code{CrossInfo}.
}
\seealso{
Other csv utilities: \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeGenoCSV}},
  \code{\link{writeMapCSV}}, \code{\link{writeMapframeCSV}}
}

