% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readCrossHDF5}
\alias{readCrossHDF5}
\title{Read \pkg{R/qtl} \code{cross} object from an input HDF5 file.}
\usage{
readCrossHDF5(infile)
}
\arguments{
\item{infile}{An input HDF5 file.}
}
\value{
The \pkg{R/qtl} \code{cross} object in the given HDF5 file.
}
\description{
Read \pkg{R/qtl} \code{cross} object from an input HDF5 file.
}
\seealso{
Other cross object functions: \code{\link{crossesEqual}},
  \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{hasTimeSeriesPhenotypes}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{permIndices}}, \code{\link{pull.alleles}},
  \code{\link{pull.chr}}, \code{\link{pull.crosstype}},
  \code{\link{pull.ind}}, \code{\link{readCrossCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeCrossHDF5}}

Other HDF5 functions: \code{\link{readMapHDF5}},
  \code{\link{readOverviewHDF5}},
  \code{\link{readResultHDF5}},
  \code{\link{writeCrossHDF5}}, \code{\link{writeMapHDF5}},
  \code{\link{writeOverviewHDF5}},
  \code{\link{writeResultHDF5}}
}

