% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{readGenoCSV}
\alias{readGenoCSV}
\title{Read yeast genotype data from a CSV file.}
\usage{
readGenoCSV(infile, require.mapunit = TRUE)
}
\arguments{
\item{infile}{Input CSV file path.}

\item{require.mapunit}{Require map unit information in map positions.}
}
\description{
Read yeast genotype data from a CSV file.
}
\seealso{
Other csv utilities: \code{\link{readCovarCSV}},
  \code{\link{readCrossCSV}}, \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{readPhenoCSV}}, \code{\link{writeCrossCSV}},
  \code{\link{writeGenoCSV}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapframeCSV}},
  \code{\link{writePhenoCSV}}
}

