% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{readGenoVCF}
\alias{readGenoVCF}
\title{Read genotype data from a VCF file.}
\usage{
readGenoVCF(..., samples, founders = NULL, alleles = NULL)
}
\arguments{
\item{...}{Input VCF file paths.}

\item{samples}{Cross sample IDs.}

\item{founders}{Founder sample IDs.}

\item{alleles}{Founder allele symbols.}
}
\value{
An \pkg{R/qtl} \code{cross} \code{geno} object.
}
\description{
This function reads SNP genotype data from one or more VCF files, and returns
these as an \pkg{R/qtl} \code{cross} \code{geno} object.
}
\details{
If no founder samples are specified, this function assigns enumerated
genotypes at each locus according to the observed raw SNP genotype. So
for example, if the SNVs at a given locus are 'A' and 'C', samples are
assigned the genotypes '1' and '2', respectively. The enumeration of
genotypes is performed independently for each locus, so a given
enumerated genotype does not have the same meaning across loci.

If founder samples are specified, this function assigns a genotype symbol
from the \code{alleles} parameter, or otherwise from a default symbol. In
the case of founder genotypes, a given genotype represents the same founder
strain across markers.
}

