% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readMapHDF5}
\alias{readMapHDF5}
\title{Read \code{map} from a HDF5 file.}
\usage{
readMapHDF5(infile, name)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{name}{Map name.}
}
\value{
An \pkg{R/qtl} \code{map} object.
}
\description{
Read \code{map} from a HDF5 file.
}
\details{
A \code{mapframe} dataset is read from the 'Maps' group at the root
of the HDF5 file, and converted to an \pkg{R/qtl} \code{map} object.
}
\seealso{
Other hdf5 utilities: \code{\link{readDatasetHDF5}},
  \code{\link{readOverviewHDF5}},
  \code{\link{readResultHDF5}},
  \code{\link{writeDatasetHDF5}},
  \code{\link{writeMapHDF5}},
  \code{\link{writeOverviewHDF5}},
  \code{\link{writeResultHDF5}}
}

