% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readObjectAttributesHDF5}
\alias{readObjectAttributesHDF5}
\title{Read HDF5 object attributes.}
\usage{
readObjectAttributesHDF5(infile, h5name)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{h5name}{HDF5 object name.}
}
\value{
List of attributes of the named HDF5 object.
}
\description{
Read HDF5 object attributes.
}
\keyword{internal}
