% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readOverviewHDF5}
\alias{readOverviewHDF5}
\title{Read QTL analysis results overview from a HDF5 file.}
\usage{
readOverviewHDF5(infile)
}
\arguments{
\item{infile}{An input HDF5 file.}
}
\value{
A \code{data.frame} containing QTL analysis results overview.
}
\description{
Read QTL analysis results overview from a HDF5 file.
}
\seealso{
Other hdf5 utilities: \code{\link{readDatasetHDF5}},
  \code{\link{readMapHDF5}}, \code{\link{readResultHDF5}},
  \code{\link{writeDatasetHDF5}},
  \code{\link{writeMapHDF5}},
  \code{\link{writeOverviewHDF5}},
  \code{\link{writeResultHDF5}}
}

