% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{readPhenoCSV}
\alias{readPhenoCSV}
\title{Read yeast phenotype data from a CSV file.}
\usage{
readPhenoCSV(infile)
}
\arguments{
\item{infile}{Input CSV file path.}
}
\description{
Read yeast phenotype data from a CSV file.
}
\seealso{
Other csv utilities: \code{\link{readCrossCSV}},
  \code{\link{readGenoCSV}}, \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeGenoCSV}},
  \code{\link{writeMapCSV}},
  \code{\link{writeMapframeCSV}},
  \code{\link{writePhenoCSV}}
}

