% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readResultHDF5}
\alias{readResultHDF5}
\title{Read QTL analysis result from a HDF5 file.}
\usage{
readResultHDF5(infile, phenotype, name)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{phenotype}{Name of phenotype (or equivalent analysis unit).}

\item{name}{Name of QTL analysis result.}
}
\value{
R object containing QTL analysis result. Returns \code{NULL}
if the given phenotype has no result with the specified name.
}
\description{
Read QTL analysis result from a HDF5 file.
}
\keyword{internal}

