% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{recodeCSV}
\alias{recodeCSV}
\title{Recode \pkg{R/qtl} data in CSV file.}
\usage{
recodeCSV(infile, outfile, geno = NULL, enum.geno = FALSE)
}
\arguments{
\item{infile}{Input CSV file path.}

\item{outfile}{Output CSV file path.}

\item{geno}{A simple \code{\link{mapping}} of old to new genotype symbols,
with names containing existing genotype symbols, and elements containing
their replacement values (incompatible with \code{enum.geno}).}

\item{enum.geno}{Option indicating if genotype data should be recoded as
enumerated genotypes (incompatible with \code{geno}).}
}
\description{
This function recodes the data in an \pkg{R/qtl} cross or genotype CSV file.
Genotypes can be recoded by passing to the \code{geno} parameter a simple
\code{\link{mapping}} of old to new genotypes. Alternatively, genotype data
can be converted to enumerated genotypes by setting parameter
\code{enum.geno} to \code{TRUE}.
}
\details{
Any \code{\link{mapping}} of old to new symbols must be complete and
unambiguous. All existing values must be mapped to a new value, and
data with different input values cannot have the same output value.
}
\seealso{
Other CSV functions: \code{\link{hasMapCSV}},
  \code{\link{readCovarCSV}}, \code{\link{readCrossCSV}},
  \code{\link{readGenoCSV}}, \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{readPhenoCSV}}, \code{\link{sniffCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeGenoCSV}},
  \code{\link{writeMapCSV}},
  \code{\link{writeMapframeCSV}},
  \code{\link{writePhenoCSV}}
}
