% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{refreshCross}
\alias{refreshCross}
\title{Refresh derived data attributes of \pkg{R/qtl} \code{cross}.}
\usage{
refreshCross(cross, derived.attributes = NULL)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{derived.attributes}{Names of derived data attributes to refresh. If 
none are specified, all derived data attributes in the \code{cross} are 
refreshed.}
}
\value{
Input \code{cross} object with derived data attributes refreshed.
}
\description{
Refresh derived data attributes of \pkg{R/qtl} \code{cross}.
}
\keyword{internal}

