% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{requestPkgDataPath}
\alias{requestPkgDataPath}
\title{Request path of \pkg{shmootl} package data.}
\usage{
requestPkgDataPath(...)
}
\arguments{
\item{...}{Path elements. These should be specified relative to the package 
data root. For the source package, this is the \pkg{shmootl} \code{'inst'} 
directory, while for the binary package, this is the same as the package 
root.}
}
\value{
An absolute path to the requested \pkg{shmootl} data file/directory.
}
\description{
Request path of \pkg{shmootl} package data.
}
\keyword{internal}
