% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{resolveAnalysisTitle}
\alias{resolveAnalysisTitle}
\title{Resolve analysis title.}
\usage{
resolveAnalysisTitle(analyses = NULL)
}
\arguments{
\item{analyses}{Vector of \pkg{shmootl} analysis pipeline names or titles.
If no analysis pipelines are specified, all supported analysis titles are
returned.}
}
\value{
If the \code{analyses} parameter is specified, this is a character
vector in which each element contains the title of the \pkg{shmootl} analysis
pipeline referenced in the corresponding element of the input vector. If the
\code{analyses} parameter is \code{NULL}, this is a character vector of all
supported analysis titles.
}
\description{
Resolve analysis title.
}
\keyword{internal}

