% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{resolveH5ObjectName}
\alias{resolveH5ObjectName}
\title{Resolve HDF5 object name.}
\usage{
resolveH5ObjectName(h5name)
}
\arguments{
\item{components}{HDF5 name components.}
}
\value{
HDF5 object name.
}
\description{
Validates and (if necessary) corrects a HDF5 object name: ensuring that it 
has a leading slash and does not have a trailing slash.
}
\keyword{internal}

