% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{resolveMapNameHDF5}
\alias{resolveMapNameHDF5}
\title{Resolve map name in HDF5 file.}
\usage{
resolveMapNameHDF5(infile, name = NULL)
}
\arguments{
\item{infile}{Input HDF5 file.}

\item{name}{Map name.}
}
\value{
Resolved map name.
}
\description{
This function resolves the name of a map in the given HDF5 file: if the given
map is present in the HDF5 file, that map name is returned. If no map name is
specified and one map is found, the name of that map is returned by default;
otherwise a map name must be given.
}
\keyword{internal}

