% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{resolveQtlIndices}
\alias{resolveQtlIndices}
\title{Resolve indices of QTL object.}
\usage{
resolveQtlIndices(x, qtl.indices = NULL)
}
\arguments{
\item{x}{A \code{qtl} object.}

\item{qtl.indices}{QTL indices referring to the given QTL object.}
}
\value{
QTL indices resolved with respect to the given QTL object.
}
\description{
Resolve indices of QTL object.
}
\keyword{internal}

