% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rstripBlankCols}
\alias{rstripBlankCols}
\title{Strip blank columns from right of \code{data.frame}.}
\usage{
rstripBlankCols(x)
}
\arguments{
\item{x}{A \code{data.frame} with columns of type \code{character}.}
}
\value{
Input object in which rightmost blank columns (i.e. columns
containing no non-whitespace characters) have been stripped.
}
\description{
Strip blank columns from right of \code{data.frame}.
}
\keyword{internal}
