% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_digest.R
\name{run_digest}
\alias{run_digest}
\title{Create digest of QTL analysis results.}
\usage{
run_digest(h5list = character(), digest = NA_character_,
  scanfile.pattern = NA_character_)
}
\arguments{
\item{h5list}{list of scan result files}

\item{digest}{scan digest file}

\item{scanfile.pattern}{scan file name pattern}
}
\description{
Create digest of QTL analysis results.
}
\concept{
shmootl:pipelines
}

