% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_estimap.R
\name{run_estimap}
\alias{run_estimap}
\title{Estimate map from cross/geno data.}
\usage{
run_estimap(datafile, mapfile, n.cluster = 1L, error.prob = 1e-04,
  map.function = c("haldane", "kosambi", "c-f", "morgan"),
  jittermap = FALSE)
}
\arguments{
\item{datafile}{input cross/geno CSV file}

\item{mapfile}{output map CSV file}

\item{n.cluster}{number of threads}

\item{error.prob}{genotyping error rate}

\item{map.function}{genetic map function}

\item{jittermap}{jitter map positions}
}
\description{
This script reads an R/qtl cross or genotype file, and outputs a
map estimated from the input genotype data. Note that a genetic map cannot be
estimated from enumerated genotypes, as these are generated independently for
each marker.
}

