% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_makecross.R
\name{run_makecross}
\alias{run_makecross}
\title{Run run_makecross pipeline.}
\usage{
run_makecross(genfile, phefile, crossfile)
}
\arguments{
\item{genfile}{input genotype CSV file}

\item{phefile}{input phenotype CSV file}

\item{crossfile}{output cross CSV file}
}
\description{
This script combines separate genotype and 
phenotype files into a single cross data file.
}

