% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_makegeno.R
\name{run_makegeno}
\alias{run_makegeno}
\title{Make genotype data from VCF.}
\usage{
run_makegeno(datafile, genfile, fdrfile = NA_character_,
  alleles = character(), digits = NA_integer_)
}
\arguments{
\item{datafile}{sample VCF file}

\item{genfile}{output genotype CSV file}

\item{fdrfile}{optional founder VCF file}

\item{alleles}{founder allele symbols}

\item{digits}{numeric precision [default: unrounded]}
}
\description{
This script will read genotype data from the specified VCF 
files, then write that genotype data to an \pkg{R/qtl} genotype CSV file.
}
\concept{
shmootl:utilities
}

