% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prep.R
\name{run_prep}
\alias{run_prep}
\title{Prepare data in \pkg{R/qtl} CSV file.}
\usage{
run_prep(datafile)
}
\arguments{
\item{datafile}{\pkg{R/qtl} CSV file}
}
\description{
This script will take an \pkg{R/qtl} input data file, and
prepare it for input to \pkg{shmootl}. Marker sequence IDs are resolved,
genetic map units are appended to map positions, and empty cells are
replaced with a missing value symbol. It is the responsibility of the
user to ensure that these changes are valid for a given input file.
}

