% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pullmap.R
\name{run_pullmap}
\alias{run_pullmap}
\title{Pull map from cross/geno data.}
\usage{
run_pullmap(infile, mapfile, require.mapunit = TRUE, include.mapunit = TRUE)
}
\arguments{
\item{infile}{input cross/geno CSV file}

\item{mapfile}{output map CSV file}

\item{require.mapunit}{require map units in input}

\item{include.mapunit}{include map units in output}
}
\description{
This script pulls the map from an R/qtl cross or genotype file,
and writes it to a separate map file.
}

