% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pushmap.R
\name{run_pushmap}
\alias{run_pushmap}
\title{Push map into cross/geno data file.}
\usage{
run_pushmap(mapfile, datafile, require.mapunit = TRUE,
  include.mapunit = TRUE)
}
\arguments{
\item{mapfile}{input map CSV file}

\item{datafile}{cross/geno CSV file}

\item{require.mapunit}{require map units in input}

\item{include.mapunit}{include map units in output}
}
\description{
This script replaces the map of the specified data file with the
data from the given map file.
}

