% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_recode.R
\name{run_recode}
\alias{run_recode}
\title{Recode data in \pkg{R/qtl} CSV file.}
\usage{
run_recode(infile, outfile, geno = NA)
}
\arguments{
\item{infile}{input CSV file}

\item{outfile}{output CSV file}

\item{geno}{genotype recode mapping}
}
\description{
Recode data in \pkg{R/qtl} CSV file.
}

