% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Create report of \pkg{R/qtl} output.}
\usage{
run_report(resultfile, reportfile)
}
\arguments{
\item{resultfile}{HDF5 result file}

\item{reportfile}{PDF report file}
}
\description{
Create report of \pkg{R/qtl} output.
}

