% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scanone.R
\name{run_scanone}
\alias{run_scanone}
\title{Do single QTL scan.}
\usage{
run_scanone(infile = NA_character_, h5file = NA_character_,
  chr = character(), pheno = character(), model = c("normal", "binary",
  "2part", "np"), method = c("em", "imp", "hk", "ehk", "mr", "mr-imp",
  "mr-argmax"), n.perm = 1000L, n.cluster = 1L, alpha = NA_real_,
  fdr = NA_real_, step = 0, error.prob = 1e-04,
  map.function = c("haldane", "kosambi", "c-f", "morgan"),
  acovfile = NA_character_, icovfile = NA_character_)
}
\arguments{
\item{infile}{input cross file}

\item{h5file}{scan result file}

\item{chr}{sequences [default: all]}

\item{pheno}{phenotypes [default: all]}

\item{model}{phenotype model}

\item{method}{method of QTL analysis}

\item{n.perm}{number of permutations}

\item{n.cluster}{number of threads}

\item{alpha}{significance level for LOD threshold}

\item{fdr}{FDR for LOD threshold}

\item{step}{step size for genotype probabilities}

\item{error.prob}{genotyping error rate}

\item{map.function}{genetic map function}

\item{acovfile}{additive covariates file}

\item{icovfile}{interactive covariates file}
}
\description{
Read cross data from the specified cross input file, run a single QTL
analysis using \pkg{R/qtl} \code{scanone}, and write the results of
that scan to the specified output file. If the input cross contains
enumerated genotypes, marker regresion is performed regardless of
the value of the \code{method} parameter.
}
\details{
LOD threshold stringency can be set through either the significance
level (\code{alpha}), or the false-discovery rate (\code{fdr}), but
not both. If neither is specified, a significance level \code{alpha}
of \code{0.05} is used by default.
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_digest}}, \code{\link{run_estimap}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_pushmap}}, \code{\link{run_recode}},
  \code{\link{run_report}}
}
\concept{
shmootl:pipelines
}

