% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scantwo.R
\name{run_scantwo}
\alias{run_scantwo}
\title{Do 2-QTL scan.}
\usage{
run_scantwo(infile = NA_character_, h5file = NA_character_,
  chr = character(), pheno = character(), model = c("normal", "binary"),
  method = c("em", "imp", "hk", "mr", "mr-imp", "mr-argmax"),
  n.perm = 1000L, n.cluster = 1L, alpha = NA_real_,
  thresholds = mapping(), step = 0, map.function = c("haldane", "kosambi",
  "c-f", "morgan"), error.prob = 1e-04, acovfile = NA_character_,
  icovfile = NA_character_)
}
\arguments{
\item{infile}{input cross file}

\item{h5file}{scan result file}

\item{chr}{sequences [default: all]}

\item{pheno}{phenotypes [default: all]}

\item{model}{phenotype model}

\item{method}{method of QTL analysis}

\item{n.perm}{number of permutations}

\item{n.cluster}{number of threads}

\item{alpha}{significance level for LOD thresholds}

\item{thresholds}{fixed LOD thresholds}

\item{step}{step size for genotype probabilities}

\item{map.function}{genetic map function}

\item{error.prob}{genotyping error rate}

\item{acovfile}{additive covariates file}

\item{icovfile}{interactive covariates file}
}
\description{
Read cross data from the specified cross input file, run 2-dimensional
QTL analysis using \pkg{R/qtl} \code{scantwo} (Broman \emph{et al.} 
2003), and write the results of that scan to the specified output file.
}
\details{
If the input cross contains enumerated genotypes, marker regression
is performed regardless of the value of the \code{method} parameter.

In typical usage, LOD threshold stringency can be set by the significance
level (\code{alpha}), with a default value of \code{0.05}. The significance
level is used to estimate LOD thresholds from \code{scantwo} permutations.

This can be bypassed by setting six fixed LOD \code{thresholds}, along with
a nominal significance level (\code{alpha}), in which case permutations are
skipped and the fixed LOD thresholds are applied directly for assessing
significance. When calling this function from the command line using
\code{Rscript}, the \code{thresholds} parameter must be specified as a YAML
string mapping \code{scantwo} LOD type names to threshold values (e.g.
\code{"full: 8, fv1: 7, int: 6, add: 5, av1: 4, one: 3"}). When called from
within the \code{R} environment, the \code{thresholds} parameter must be
specified as a mapping object (e.g.
\code{mapping( c(full=8, fv1=7, int=6, add=5, av1=4, one=3) )}).
}
\author{
Yue Hu

Thomas Walsh <tw164@le.ac.uk>
}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})
}
\seealso{
\href{http://www.rqtl.org/manual/qtl-manual.pdf}{R/qtl manual}

Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_digest}}, \code{\link{run_estimap}},
  \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_pushmap}}, \code{\link{run_recode}},
  \code{\link{run_report}}, \code{\link{run_scanone}}
}
\concept{
shmootl:analysis
}

