% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\docType{package}
\name{Scan Utilities}
\alias{Scan Utilities}
\alias{Scan Utilities-package}
\title{QTL scan utilities.}
\description{
The most basic scan utility is \code{\link{batchScan}}, which 
performs a batch QTL analysis on an \pkg{R/qtl} \code{cross} with a 
user-specified scan function. In general, batch scan functions should 
call \code{batchScan}. Each batch scan function must pass a node scan 
function to \code{batchScan}: the node scan function should operate on 
a single element of the batch input vector, which might contain phenotypes 
or permutation indices. 

For example, \code{\link{batchPermScanone}} can be used to perform a batch of
permutation scans with \code{qtl::scanone}. This function generates a vector
of permutation indices from the specified number of permutations, and passes
node function \code{\link{nodePermScanone}} to \code{\link{batchScan}}, which 
applies the node function to the vector of permutation indices. The result is
a \code{scanoneperm} matrix with one row for each permutation.

As another example, \code{\link{batchPhenoScanone}} can be used to perform a 
multi-phenotype batch of \code{qtl::scanone} tasks. This function generates a 
vector of phenotype column indices, each of which is individually passed to
\code{\link{nodePermScanone}} by \code{\link{batchScan}}. Each individual 
phenotype is analysed in a separate batch task, and the output is a 
\code{scanone} object containing results for the given phenotypes.
}

