% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setAlleles}
\alias{setAlleles}
\alias{setAlleles,CrossInfo-method}
\title{Set allele symbols.}
\usage{
setAlleles(cross.info, alleles)

\S4method{setAlleles}{CrossInfo}(cross.info, alleles)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{alleles}{Vector of allele symbols.}
}
\value{
Input \code{CrossInfo} object with the given alleles.
}
\description{
Set allele symbols.
}
\keyword{internal}

