% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{setColumnFromRownames}
\alias{setColumnFromRownames}
\title{Move object row names to the specified column.}
\usage{
setColumnFromRownames(x, col.name = "rownames")
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}.}

\item{col.name}{Name of the column to which row names are to be set.}
}
\value{
Input object with row names inserted into the leftmost column, and 
original row names removed.
}
\description{
Move object row names to the specified column.
}
\keyword{internal}
