% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setCrossType}
\alias{setCrossType}
\alias{setCrossType,CrossInfo-method}
\title{Set cross type.}
\usage{
setCrossType(cross.info, crosstype)

\S4method{setCrossType}{CrossInfo}(cross.info, crosstype)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{crosstype}{Cross type name.}
}
\value{
Input \code{CrossInfo} object with the given cross type name.
}
\description{
Set cross type.
}
\keyword{internal}

