% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setGenotypes}
\alias{setGenotypes}
\alias{setGenotypes,CrossInfo-method}
\title{Set genotype symbols.}
\usage{
setGenotypes(cross.info, genotypes)

\S4method{setGenotypes}{CrossInfo}(cross.info, genotypes)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{alleles}{Vector of genotype symbols.}
}
\value{
Input \code{CrossInfo} object with the given genotypes.
}
\description{
Set genotype symbols.
}
\keyword{internal}

