% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setMarkers}
\alias{setMarkers}
\alias{setMarkers,CrossInfo-method}
\title{Set markers.}
\usage{
setMarkers(cross.info, markers)

\S4method{setMarkers}{CrossInfo}(cross.info, markers)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{markers}{Vector of marker IDs.}
}
\value{
Input \code{CrossInfo} object with the given markers.
}
\description{
Set markers.
}
\keyword{internal}

