% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setMarkerSeqs}
\alias{setMarkerSeqs}
\alias{setMarkerSeqs,CrossInfo-method}
\title{Set marker sequences.}
\usage{
setMarkerSeqs(cross.info, markers = NULL, sequences = NULL)

\S4method{setMarkerSeqs}{CrossInfo}(cross.info, markers = NULL,
  sequences = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{markers}{Vector containing markers to assign sequences.}

\item{sequences}{Vector containing sequences for the given markers.}
}
\value{
Input \code{CrossInfo} object with the given marker-sequence info.
}
\description{
Set marker sequences.
}
\keyword{internal}

