% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setPhenotypes}
\alias{setPhenotypes}
\alias{setPhenotypes,CrossInfo-method}
\title{Set phenotypes.}
\usage{
setPhenotypes(cross.info, phenotypes)

\S4method{setPhenotypes}{CrossInfo}(cross.info, phenotypes)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{phenotypes}{Vector of phenotype IDs.}
}
\value{
Input \code{CrossInfo} object with the given phenotypes.
}
\description{
Set phenotypes.
}
\keyword{internal}

