% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{setPosColDataMapUnit}
\alias{setPosColDataMapUnit}
\alias{setPosColDataMapUnit.character}
\alias{setPosColDataMapUnit.data.frame}
\alias{setPosColDataMapUnit.numeric}
\title{Set map unit of position column data.}
\usage{
setPosColDataMapUnit(x, map.unit)

\method{setPosColDataMapUnit}{data.frame}(x, map.unit)

\method{setPosColDataMapUnit}{character}(x, map.unit)

\method{setPosColDataMapUnit}{numeric}(x, map.unit)
}
\arguments{
\item{x}{A \code{mapframe} or equivalent \code{data.frame}, 
or a vector of map positions.}

\item{map.unit}{Map unit. Set to \code{NULL} to remove map unit information 
from position column data.}
}
\value{
Input object with the specified map unit appended to position column 
data.
}
\description{
Set map unit of position column data.
}
\section{Map Units}{
 
A \code{map} or \code{mapframe} must express positions in either genetic or 
physical map units. Genetic map positions must be in centiMorgans (cM), while 
physical map positions must be in base pairs (bp), kilobases (kb), or 
Megabases (Mb).
}
\keyword{internal}

