% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{setRownamesFromColumn}
\alias{setRownamesFromColumn}
\title{Move specified column to object row names.}
\usage{
setRownamesFromColumn(x, col.name = "rownames")
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}.}

\item{col.name}{Name of the column from which row names are to be set.}
}
\value{
Input object with the leftmost column removed, and with row names set
from its contents.
}
\description{
Move specified column to object row names.
}
\keyword{internal}

