% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setSequences}
\alias{setSequences}
\alias{setSequences,CrossInfo-method}
\title{Set sequences.}
\usage{
setSequences(cross.info, sequences)

\S4method{setSequences}{CrossInfo}(cross.info, sequences)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{sequences}{Vector of sequence labels.}
}
\value{
Input \code{CrossInfo} object with the given sequences.
}
\description{
Set sequences.
}
\keyword{internal}

