% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setTetradIndices}
\alias{setTetradIndices}
\alias{setTetradIndices,CrossInfo-method}
\title{Set sample tetrad indices.}
\usage{
setTetradIndices(cross.info, samples = NULL, tetrads = NULL)

\S4method{setTetradIndices}{CrossInfo}(cross.info, samples = NULL,
  tetrads = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{samples}{Vector of samples for which a tetrad index should be set.}

\item{tetrads}{Vector of tetrad indices.}
}
\value{
Input \code{CrossInfo} object with the given tetrad indices.
}
\description{
Set sample tetrad indices.
}
\keyword{internal}

