% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R, R/H5Stack-class.R
\docType{methods}
\name{setValidity}
\alias{setValidity}
\alias{setValidity,CrossInfo-method}
\alias{setValidity,H5Stack-method}
\title{Validate \code{\linkS4class{CrossInfo}} object.}
\arguments{
\item{object}{A \code{\linkS4class{CrossInfo}} object.}

\item{object}{A \code{H5Stack} object.}
}
\value{
\code{TRUE} if object is valid;
otherwise, a character vector of errors.

\code{TRUE} if object is valid;
otherwise, a character vector of errors.
}
\description{
Validate \code{\linkS4class{CrossInfo}} object.

Validate \code{H5Stack} object.
}
\keyword{internal}

