% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmootl.R
\docType{package}
\name{shmootl-package}
\alias{shmootl-package}
\alias{shmootl}
\title{QTL Analysis Utilities for Yeast}
\description{
The \pkg{shmootl} package contains pipelines and functions for QTL analysis
in yeast. This package follows many of the conventions on \pkg{R/qtl}, and
depends on that package for standard QTL analysis. It also depends on
\pkg{R/funqtl} for QTL analysis with time-series phenotypes. See
below for links to these packages, as well as key references.
}
\section{Item IDs}{


Various items in \pkg{shmootl} (e.g. chromosomes/sequences, markers, samples,
phenotypes) have an associated identifier (ID). A valid item ID in
\pkg{shmootl} is a string of nonzero length, containing any combination
of printable ASCII characters (including letters and numbers), except for a
single quote, double quotes, backtick, comma, forward slash, or backslash.
}

\section{Syntactic Names}{


A syntactically valid R name contains only letters, numbers, dots, and
underscores. The name should start with a letter, or a dot that is not
followed by a number. In addition, specific reserved words are not
permitted. Many R packages and functions automatically render item names
syntactically valid, so for example, a phenotype name may be changed when
added to an \pkg{R/qtl} \code{cross} object. For more information, see the
relevant \href{https://cran.r-project.org/manuals.html}{R documentation}.
See also the base R function \code{make.names}.
}

\section{Package Pipelines}{


Within \pkg{\link{shmootl}}, package pipelines are fully-documented functions that
follow specific conventions regarding function name and formal arguments.
Pipeline function names are of the form \code{'run_<pipeline>'} (where
\code{'<pipeline>'} is replaced by the pipeline name).

Package pipelines can be run from a command line with \code{Rscript} as follows:

\code{Rscript -e 'library(shmootl)' -e 'run()' <pipeline> [-h] [<args>]}

...where \code{\link{run}} is the general \pkg{shmootl} pipeline-running
function, \code{<pipeline>} is the name of the pipeline to run, and
\code{<args>} represents any arguments to be passed to the given pipeline.
To see the available options for a pipeline, input the help flag (\code{-h})
after the name of the pipeline. To list all available pipelines, input the
help flag (\code{-h}) without a pipeline name. (See usage examples below.)
}

\section{See Also}{
 

\itemize{
  \item{\href{https://github.com/gact/shmootl}{shmootl on GitHub}}
  \item{\href{https://github.com/ikwak2/funqtl}{R/funqtl on GitHub}}
  \item{\href{http://www.rqtl.org/}{R/qtl}}
}
}

\examples{
\dontrun{
# To list available pipelines from the command line.
Rscript -e 'library(shmootl)' -e 'run()' -h

# To run the 'scanone' pipeline from the command line.
Rscript -e 'library(shmootl)' -e 'run()' scanone \\
    --infile input.csv --h5file output.hdf5

# To run the 'scanone' pipeline from within R.
run_scanone(infile='input.csv', h5file='output.hdf5')
}
}
\references{
Arends D, Prins P, Jansen RC, Broman KW (2010) R/qtl:
high-throughput multiple QTL mapping. \emph{Bioinformatics}
\bold{26}(26):2990-2.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/20966004}{PubMed})

Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})

Kwak IY, Moore CR, Spalding EP, Broman KW (2014) A simple
regression-based method to map quantitative trait loci underlying
function-valued phenotypes. \emph{Genetics} \bold{197}(4):1409-16.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/24931408}{PubMed})
}
