% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5Stack-class.R
\docType{methods}
\name{length,H5Stack-method}
\alias{length,H5Stack-method}
\alias{show,H5Stack-method}
\alias{show.H5Stack}
\title{Get number of HDF5 objects in \code{H5Stack}.}
\usage{
\S4method{length}{H5Stack}(x)

show.H5Stack(object)
}
\arguments{
\item{object}{\code{H5Stack} object.}

\item{object}{A \code{H5Stack} object.}
}
\value{
Number of HDF5 objects in \code{H5Stack}.
}
\description{
Get number of HDF5 objects in \code{H5Stack}.

Display contents of \code{H5Stack} object.
}
\keyword{internal}

