% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{sniffCSV}
\alias{sniffCSV}
\title{Identify type of \pkg{R/qtl} data in CSV file.}
\usage{
sniffCSV(infile)
}
\arguments{
\item{infile}{Input CSV file path.}
}
\value{
A string describing the type of data in the input CSV file. This can
be \code{'cross'}, \code{'geno'}, \code{'pheno'}, or \code{'map'}. Returns
\code{'unknown'} if the data could not be identified.
}
\description{
Identify type of \pkg{R/qtl} data in CSV file.
}
\keyword{internal}

