% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{sortChr}
\alias{sortChr}
\title{Sort chromosome labels.}
\usage{
sortChr(x)
}
\arguments{
\item{x}{Vector of chromosome labels.}
}
\value{
Input vector of chromosome labels, sorted
with respect to their normalised form.
}
\description{
Sort chromosome labels.
}
\seealso{
Other chromosome/sequence functions: \code{\link{formatChr}},
  \code{\link{formatSeq}}, \code{\link{isNormChr}},
  \code{\link{isNormSeq}}, \code{\link{normChr}},
  \code{\link{normSeq}}, \code{\link{orderChr}},
  \code{\link{orderSeq}}, \code{\link{rankChr}},
  \code{\link{rankSeq}}, \code{\link{sortSeq}}
}

