% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{sortSeq}
\alias{sortSeq}
\title{Sort sequence labels.}
\usage{
sortSeq(x)
}
\arguments{
\item{x}{Vector of sequence labels.}
}
\value{
Input vector of sequence labels, sorted
with respect to their normalised form.
}
\description{
Sort sequence labels.
}
\seealso{
Other chromosome/sequence functions: \code{\link{formatChr}},
  \code{\link{formatSeq}}, \code{\link{isNormChr}},
  \code{\link{isNormSeq}}, \code{\link{normChr}},
  \code{\link{normSeq}}, \code{\link{orderChr}},
  \code{\link{orderSeq}}, \code{\link{rankChr}},
  \code{\link{rankSeq}}, \code{\link{sortChr}}
}

