% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{splitCSL}
\alias{splitCSL}
\title{Split a comma-separated list.}
\usage{
splitCSL(x)
}
\arguments{
\item{x}{Character vector of length one, containing items as a 
comma-separated list (CSL).}
}
\value{
Character vector with each element containing one item of the input 
CSL.
}
\description{
Split a comma-separated list.
}
\keyword{internal}

