% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{splitH5ObjectName}
\alias{splitH5ObjectName}
\title{Split HDF5 object name into component parts.}
\usage{
splitH5ObjectName(h5name)
}
\arguments{
\item{h5name}{HDF5 object name.}
}
\value{
HDF5 name components.
}
\description{
Split HDF5 object name into component parts.
}
\keyword{internal}

