% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{splitSeqLabels}
\alias{splitSeqLabels}
\title{Split sequence labels.}
\usage{
splitSeqLabels(x)
}
\arguments{
\item{x}{Vector of sequence labels.}
}
\value{
A \code{data.frame} representing split sequence labels.
}
\description{
Split sequence labels.
}
\keyword{internal}

