% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stopif}
\alias{stopif}
\title{Stop if the expression is \code{TRUE}.}
\usage{
stopif(expression)
}
\arguments{
\item{expression}{R expression to be tested.}
}
\description{
Stop if the expression is \code{TRUE}.
}
\details{
This function is based on the R base function \code{stopifnot}, but  
is defined only for the simple case of a single expression. It should be used
to avoid confusing double negatives.
}
\keyword{internal}

