% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stripWhite}
\alias{stripWhite}
\title{Strip leading/trailing whitespace from elements of a character vector.}
\usage{
stripWhite(x)
}
\arguments{
\item{x}{Character vector whose elements are to be stripped.}
}
\value{
Whitespace-stripped copy of the input character vector.
}
\description{
Strip leading/trailing whitespace from elements of a character vector.
}
\keyword{internal}
