% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonebins.R
\name{subset.summary.scanonebins}
\alias{subset.summary.scanonebins}
\title{Subset a \code{summary.scanonebins} object.}
\usage{
\method{subset}{summary.scanonebins}(x, fdrs = NULL, ...)
}
\arguments{
\item{x}{A \code{summary.scanonebins} object.}

\item{fdrs}{FDR levels by which to subset the input object.}

\item{...}{Unused arguments.}
}
\value{
The input \code{summary.scanonebins} object,
subsetted with respect to the specified constraints.
}
\description{
Subset a \code{summary.scanonebins} object.
}
\keyword{internal}

