% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.summary.scanoneperm.R
\name{subset.summary.scanoneperm}
\alias{subset.summary.scanoneperm}
\title{Subset a \code{summary.scanoneperm} object.}
\usage{
\method{subset}{summary.scanoneperm}(x, alphas = NULL, lodcolumns = NULL,
  ...)
}
\arguments{
\item{x}{A \code{summary.scanoneperm} object.}

\item{alphas}{Significance levels by which to subset the input object.}

\item{lodcolumns}{In a \code{scanone} or equivalent object, this
parameter indicates which LOD columns to consider. In other objects
with LOD-column-associated elements, this parameter indicates which
of those LOD-column-associated elements to consider. The specified
LOD columns must be a character vector of LOD column names, a logical
vector with length equal to the number of LOD columns, or a numeric
vector of indices \emph{with respect to the set of LOD columns}. If
no LOD columns are specified, all are used.}

\item{...}{Unused arguments.}
}
\value{
The input \code{summary.scanoneperm} object,
subsetted with respect to the specified constraints.
}
\description{
Subset a \code{summary.scanoneperm} object.
}
\keyword{internal}
