% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByLoci.R
\docType{methods}
\name{subsetByLoci}
\alias{subsetByLoci}
\alias{subsetByLoci,DNAStringSet-method}
\alias{subsetByLoci,QualityScaledDNAStringSet-method}
\alias{subsetByLoci.DNAStringSet}
\alias{subsetByLoci.QualityScaledDNAStringSet}
\alias{subsetByLoci.map}
\alias{subsetByLoci.mapframe}
\title{Subset object by the specified loci.}
\usage{
subsetByLoci(x, loc = NULL)

subsetByLoci.mapframe(x, loc = NULL)

subsetByLoci.map(x, loc = NULL)

subsetByLoci.DNAStringSet(x, loc = NULL)

subsetByLoci.QualityScaledDNAStringSet(x, loc = NULL)

subsetByLoci(x, loc = NULL)

\S4method{subsetByLoci}{DNAStringSet}(x, loc = NULL)

\S4method{subsetByLoci}{QualityScaledDNAStringSet}(x, loc = NULL)
}
\arguments{
\item{x}{Object subsettable by loci.}

\item{loc}{Locus \code{mapframe} specifying map loci.}
}
\value{
Input object subsetted by the specified loci.
}
\description{
Subset object by the specified loci.
}
\keyword{internal}

