% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByLocusID.R
\name{subsetByLocusID}
\alias{subsetByLocusID}
\alias{subsetByLocusID.data.frame}
\alias{subsetByLocusID.map}
\title{Subset object by locus ID.}
\usage{
subsetByLocusID(x, predicate)

\method{subsetByLocusID}{data.frame}(x, predicate)

\method{subsetByLocusID}{map}(x, predicate)
}
\arguments{
\item{x}{Object with locus IDs.}

\item{predicate}{Function applied to locus IDs that determines which loci are
retained in the output object.}
}
\value{
Input object containing only loci whose locus IDs are evaluated as 
true by the predicate function.
}
\description{
Subset object by locus ID.
}
\keyword{internal}
