% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetBySeq.R
\docType{methods}
\name{subsetBySeq}
\alias{subsetBySeq}
\alias{subsetBySeq,DNAStringSet-method}
\alias{subsetBySeq,QualityScaledDNAStringSet-method}
\alias{subsetBySeq.DNAStringSet}
\alias{subsetBySeq.QualityScaledDNAStringSet}
\alias{subsetBySeq.character}
\alias{subsetBySeq.data.frame}
\alias{subsetBySeq.map}
\title{Subset object by the specified sequences.}
\usage{
subsetBySeq(x, sequences = NULL)

subsetBySeq.character(x, sequences = NULL)

subsetBySeq.data.frame(x, sequences = NULL)

subsetBySeq.map(x, sequences = NULL)

subsetBySeq.DNAStringSet(x, sequences = NULL)

subsetBySeq.QualityScaledDNAStringSet(x, sequences = NULL)

subsetBySeq(x, sequences = NULL)

\S4method{subsetBySeq}{DNAStringSet}(x, sequences = NULL)

\S4method{subsetBySeq}{QualityScaledDNAStringSet}(x, sequences = NULL)
}
\arguments{
\item{x}{Object subsettable by sequence.}

\item{sequences}{Sequences with which to subset object.}
}
\value{
Input object subsetted by the specified sequences.
}
\description{
Subset object by the specified sequences.
}
\keyword{internal}

