% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{subsetMap}
\alias{subsetMap}
\title{Subset \pkg{shmootl} \code{map} object.}
\usage{
subsetMap(x, ...)
}
\arguments{
\item{x}{A \code{map} object.}

\item{...}{Arguments passed to \code{subset}.}
}
\value{
Subsetted \code{map} object.
}
\description{
\pkg{R/qtl} implements a \code{subset.map} method, but this method strips
some attributes from the \code{map} object. This function transfers any
stripped attributes to the subsetted \code{map} object.
}
\keyword{internal}

