% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5Stack-class.R
\name{summary.H5Stack}
\alias{summary,H5Stack-method}
\alias{summary.H5Stack}
\title{Display summary of \code{H5Stack} object.}
\usage{
summary.H5Stack(object)
}
\arguments{
\item{object}{A \code{H5Stack} object.}
}
\description{
Display summary of \code{H5Stack} object.
}
\keyword{internal}

