% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonebins.R
\name{summary.scanonebins}
\alias{summary.scanonebins}
\title{Estimate FDR LOD thresholds from a \code{scanonebins} object.}
\usage{
\method{summary}{scanonebins}(object, scanone.result, lodcolumns = NULL,
  fdr = 0.01, ...)
}
\arguments{
\item{object}{A \code{scanonebins} object.}

\item{scanone.result}{A \code{scanone} object created from the same data that
was used to create the \code{scanonebins} object.}

\item{lodcolumns}{In a \code{scanone} or equivalent object, this parameter 
indicates which LOD columns to consider. These must be either a character
vector of column names or a numeric vector of indices \emph{with respect
to the set of LOD columns}. If no LOD columns are specified, all are used.}

\item{fdr}{False discovery rates for which an approximate LOD thresholds
should be estimated.}

\item{...}{Unused arguments.}
}
\value{
A \code{summary.scanonebins} matrix containing LOD thresholds that
correspond approximately to the specified false discovery rates.
}
\description{
This function is based on the \pkg{R/qtl} function \code{mqmscanfdr} (see
reference below), but uses a precalculated \code{scanonebins} object
instead of running permutations directly.
}
\details{
Thresholds are set from the bin edges of the \code{scanonebins} object, and
empirical false-discovery rates (FDRs) are calculated at these thresholds.
For each false-discovery rate specified in the \code{fdr} parameter, the
threshold with the next lowest FDR estimate is returned where possible, or
next highest FDR estimate otherwise. Summary results take \code{NA} values
in cases where no sensible FDR can be returned (e.g. fewer loci above a
threshold in the main scanone result than in the corresponding permutation
results).
}
\references{
Arends D, Prins P, Jansen RC, Broman KW (2010) R/qtl:
high-throughput multiple QTL mapping. Bioinformatics 26(23):2990-2.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/20966004}{PubMed})
}
\keyword{internal}

