% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{updateResultsOverview}
\alias{updateResultsOverview}
\title{Update QTL analysis results overview.}
\usage{
updateResultsOverview(overview, analysis, results = NULL)
}
\arguments{
\item{overview}{A \code{data.frame} containing
a QTL analysis results overview.}

\item{analysis}{Name of analysis from which results were generated.}

\item{results}{Named character vector summarising results of analysis.
Each result is matched to its corresponding phenotype by name.}
}
\value{
A results overview \code{data.frame}, with two or more columns:
\code{'Phenotype'}, listing phenotypes for which analyses were done; and one
or more additional columns, each summarising the results of a QTL analysis.
}
\description{
Update QTL analysis results overview.
}
\keyword{internal}

