% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{validateAlleles}
\alias{validateAlleles}
\alias{validateAlleles,CrossInfo-method}
\title{Validate allele information.}
\usage{
validateAlleles(cross.info)

\S4method{validateAlleles}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
\code{TRUE} if alleles are valid; otherwise, returns first error.
}
\description{
Validate allele information.
}
\keyword{internal}
