% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{validateAlleleSet}
\alias{validateAlleleSet}
\title{Validate a set of alleles.}
\usage{
validateAlleleSet(x)
}
\arguments{
\item{x}{Vector of alleles.}
}
\value{
\code{TRUE} if allele set is valid; otherwise raises first error.
}
\description{
Validate a set of alleles.
}
\keyword{internal}
