% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{validateCrossType}
\alias{validateCrossType}
\alias{validateCrossType,CrossInfo-method}
\title{Validate cross type information.}
\usage{
validateCrossType(cross.info)

\S4method{validateCrossType}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
TRUE if cross type is valid; otherwise, returns first error.
}
\description{
Validate cross type information.
}
\keyword{internal}

