% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{validateGenotypes}
\alias{validateGenotypes}
\alias{validateGenotypes,CrossInfo-method}
\title{Validate genotype information.}
\usage{
validateGenotypes(cross.info)

\S4method{validateGenotypes}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
\code{TRUE} if genotypes are valid; otherwise, returns first error.
}
\description{
Validate genotype information.
}
\keyword{internal}
