% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{validateMap}
\alias{validateMap}
\alias{validateMap.list}
\alias{validateMap.map}
\title{Validate \code{map} object.}
\usage{
validateMap(x, ...)

\method{validateMap}{list}(x)

\method{validateMap}{map}(x, package = c("qtl", "shmootl"))
}
\arguments{
\item{x}{A \code{map} or equivalent \code{list}.}
}
\value{
TRUE if object is a valid \code{map} or equivalent \code{list}; 
otherwise, returns first error.
}
\description{
Validate \code{map} object.
}
\section{Map}{
 

The \pkg{R/qtl} \code{map} object - described in detail in Section 2.6 of 
Broman and Sen (2009) - comprises a nonempty list, each element of which 
contains two or more marker positions for a given chromosome. Each list 
element is named for its corresponding chromosome, while the names of each 
chromosome's marker position vector comprise the marker locus IDs.

A \code{map} must represent either a genetic or a physical map. These are 
distinguished by the \code{'map.unit'} attribute of an object, which can
contain either genetic or physical map units.
}
\keyword{internal}

