% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{validateMapframe}
\alias{validateMapframe}
\alias{validateMapframe.data.frame}
\alias{validateMapframe.mapframe}
\title{Validate \code{mapframe} object.}
\usage{
validateMapframe(x, ...)

\method{validateMapframe}{data.frame}(x)

\method{validateMapframe}{mapframe}(x)
}
\arguments{
\item{x}{A \code{mapframe} or equivalent \code{data.frame}.}
}
\value{
TRUE if object is a valid \code{mapframe} or equivalent 
\code{data.frame}; otherwise, returns first error.
}
\description{
Validate \code{mapframe} object.
}
\section{Mapframe}{
 

A \code{mapframe} object is a class of \code{data.frame} that contains mapped
information, or a \code{data.frame} with equivalent properties. The leftmost 
two columns of every \code{mapframe} object must be as follows:

\itemize{
\item{\code{chr}:}   {Chromosome/sequence labels in the order defined for the current 
reference genome.}
\item{\code{pos}:}   {Map positions in increasing order within each 
chromosome/sequence. The map unit used in a \code{mapframe} 
position column must be indicated by setting its 
\code{'map.unit'} attribute.}
}

Any number of additional data columns can be included. If present, the row 
names indicate the locus IDs, which should be unique within the given
\code{mapframe}.
}
\keyword{internal}

