% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{validateMapUnit}
\alias{validateMapUnit}
\alias{validateMapUnit.character}
\alias{validateMapUnit.default}
\title{Validate map unit.}
\usage{
validateMapUnit(x, map.type = NULL)

\method{validateMapUnit}{character}(x, map.type = NULL)

\method{validateMapUnit}{default}(x, map.type = NULL)
}
\arguments{
\item{x}{Map unit, or object with a \code{'map.unit'} attribute.}

\item{map.type}{Map type for which map unit should be validated. If not
specified, map unit is validated for all map types.}
}
\value{
TRUE if the specified map unit is a valid map unit for the given
map type(s); otherwise, returns error.
}
\description{
Validate map unit.
}
\section{Map Units}{
 
A \code{map} or \code{mapframe} must express positions in either genetic or 
physical map units. Genetic map positions must be in centiMorgans (cM), while 
physical map positions must be in base pairs (bp), kilobases (kb), or 
Megabases (Mb).
}
\keyword{internal}

