% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{validateMarkers}
\alias{validateMarkers}
\alias{validateMarkers,CrossInfo-method}
\title{Validate marker information.}
\usage{
validateMarkers(cross.info)

\S4method{validateMarkers}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
\code{TRUE} if marker information is valid;
otherwise, returns first error.
}
\description{
Validate marker information.
}
\keyword{internal}

