% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{validatePhenotypes}
\alias{validatePhenotypes}
\alias{validatePhenotypes,CrossInfo-method}
\title{Validate phenotype information.}
\usage{
validatePhenotypes(cross.info)

\S4method{validatePhenotypes}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
\code{TRUE} if phenotypes are valid;
otherwise, returns first error.
}
\description{
Validate phenotype information.
}
\keyword{internal}

