% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{validatePhysicalMapUnit}
\alias{validatePhysicalMapUnit}
\title{Validate physical map unit.}
\usage{
validatePhysicalMapUnit(x)
}
\arguments{
\item{x}{Map unit, or object with a \code{'map.unit'} attribute.}
}
\value{
TRUE if map unit is a valid physical map unit;
otherwise, returns error.
}
\description{
Validate physical map unit.
}
\section{Map Units}{
 
A \code{map} or \code{mapframe} must express positions in either genetic or 
physical map units. Genetic map positions must be in centiMorgans (cM), while 
physical map positions must be in base pairs (bp), kilobases (kb), or 
Megabases (Mb).
}
\keyword{internal}

