% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{validateRawGenoArray}
\alias{validateRawGenoArray}
\title{Validate \code{array} as containing raw genotype data.}
\usage{
validateRawGenoArray(x)
}
\arguments{
\item{x}{Raw genotype \code{array}.}
}
\value{
\code{TRUE} if object is an \code{array} containing raw
genotype data in expected form; otherwise, raises first error.
}
\description{
Validate \code{array} as containing raw genotype data.
}
\keyword{internal}
