% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{validateResultsOverview}
\alias{validateResultsOverview}
\title{Validate a QTL analysis results overview.}
\usage{
validateResultsOverview(overview)
}
\arguments{
\item{overview}{Results overview \code{data.frame}.}
}
\value{
\code{TRUE} if results overview is valid;
otherwise raises first error.
}
\description{
Validate a QTL analysis results overview.
}
\keyword{internal}

