% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{validateSamples}
\alias{validateSamples}
\alias{validateSamples,CrossInfo-method}
\title{Validate sample information.}
\usage{
validateSamples(cross.info)

\S4method{validateSamples}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
\code{TRUE} if sample information is valid;
otherwise, returns first error.
}
\description{
Validate sample information.
}
\keyword{internal}

