% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{validateSequences}
\alias{validateSequences}
\alias{validateSequences,CrossInfo-method}
\title{Validate sequence information.}
\usage{
validateSequences(cross.info)

\S4method{validateSequences}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
TRUE if sequences are valid; otherwise, returns first error.
}
\description{
Validate sequence information.
}
\keyword{internal}

