% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{values}
\alias{values}
\title{Get values of vector object.}
\usage{
values(x)
}
\arguments{
\item{x}{R vector object.}
}
\value{
Unnamed list of values in R vector object.
}
\description{
Get values of vector object.
}
\keyword{internal}

