% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{writeCrossCSV}
\alias{writeCrossCSV}
\title{Write yeast \code{cross} to a CSV file.}
\usage{
writeCrossCSV(cross, outfile, chr = NULL, digits = NULL,
  include.mapunit = TRUE)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{outfile}{Output CSV file path.}

\item{chr}{Vector of sequences for which genotype data should be included in 
the output file. If none are specified, genotype data are output for all 
sequences.}

\item{digits}{If specified, round genetic map positions and numeric phenotype 
values to the specified number of digits.}

\item{include.mapunit}{Include map unit information in map positions.}
}
\description{
Function \code{writeCrossCSV} writes a yeast \code{cross} to a 
CSV file. Phenotype, genotype, and sample IDs are taken from the \code{'info'}
attribute of the \code{cross}, if present.
}
\seealso{
Other csv utilities: \code{\link{readCrossCSV}},
  \code{\link{readGenoCSV}}, \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{readPhenoCSV}}, \code{\link{writeGenoCSV}},
  \code{\link{writeMapCSV}},
  \code{\link{writeMapframeCSV}},
  \code{\link{writePhenoCSV}}
}

