% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{writeCrossHDF5}
\alias{writeCrossHDF5}
\title{Write \pkg{R/qtl} \code{cross} object to an output HDF5 file.}
\usage{
writeCrossHDF5(cross, outfile, overwrite = FALSE)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{outfile}{An output HDF5 file.}

\item{overwrite}{Option indicating whether to force overwrite of an existing
\code{cross}. By default, existing \code{cross} objects cannot be overwritten.}
}
\description{
Write \pkg{R/qtl} \code{cross} object to an output HDF5 file.
}
\seealso{
Other cross object functions: \code{\link{crossesEqual}},
  \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{hasTimeSeriesPhenotypes}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{permIndices}}, \code{\link{pull.alleles}},
  \code{\link{pull.chr}}, \code{\link{pull.crosstype}},
  \code{\link{pull.ind}}, \code{\link{readCrossCSV}},
  \code{\link{readCrossHDF5}}, \code{\link{writeCrossCSV}}

Other HDF5 functions: \code{\link{readCrossHDF5}},
  \code{\link{readMapHDF5}},
  \code{\link{readOverviewHDF5}},
  \code{\link{readResultHDF5}}, \code{\link{writeMapHDF5}},
  \code{\link{writeOverviewHDF5}},
  \code{\link{writeResultHDF5}}
}

