% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{writeGenoCSV}
\alias{writeGenoCSV}
\title{Write yeast genotype data to a CSV file.}
\usage{
writeGenoCSV(geno, outfile, chr = NULL, digits = NULL,
  missing.value = "-", include.mapunit = TRUE)
}
\arguments{
\item{geno}{An \pkg{R/qtl} \code{cross} \code{geno} object.}

\item{outfile}{Output CSV file path.}

\item{chr}{Vector of sequences for which genotype data should be included in 
the output file. If none are specified, a genotype file is output for all 
available sequences.}

\item{digits}{If specified, round genetic map positions to the specified 
number of digits.}

\item{missing.value}{Missing data value. This can be any single character 
that is not a possible genotype value.}

\item{include.mapunit}{Include map unit information in map positions.}
}
\description{
Write yeast genotype data to a CSV file.
}
\seealso{
Other csv utilities: \code{\link{readCrossCSV}},
  \code{\link{readGenoCSV}}, \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapframeCSV}}
}

