% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{writeMapCSV}
\alias{writeMapCSV}
\title{Write \code{map} to a CSV file.}
\usage{
writeMapCSV(map, outfile, include.mapunit = TRUE)
}
\arguments{
\item{map}{An \pkg{R/qtl} \code{map} object.}

\item{outfile}{Output CSV file path.}

\item{include.mapunit}{Include map unit information in map positions.}
}
\description{
Write \code{map} to a CSV file.
}
\seealso{
Other csv utilities: \code{\link{readCrossCSV}},
  \code{\link{readGenoCSV}}, \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeGenoCSV}},
  \code{\link{writeMapframeCSV}}
}

