% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{writeMapframeCSV}
\alias{writeMapframeCSV}
\title{Write \code{mapframe} to a CSV file.}
\usage{
writeMapframeCSV(x, outfile, include.mapunit = TRUE)
}
\arguments{
\item{x}{A \code{mapframe} object.}

\item{outfile}{Output CSV file path.}

\item{include.mapunit}{Include map unit information in map positions.}
}
\description{
Write \code{mapframe} to a CSV file.
}
\seealso{
Other csv utilities: \code{\link{readCrossCSV}},
  \code{\link{readGenoCSV}}, \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeGenoCSV}},
  \code{\link{writeMapCSV}}
}

