% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{writeObjectAttributesHDF5}
\alias{writeObjectAttributesHDF5}
\title{Write attributes to a HDF5 object.}
\usage{
writeObjectAttributesHDF5(x, outfile, h5name)
}
\arguments{
\item{x}{List of attributes, or an R object whose attributes will be written.}

\item{outfile}{An output HDF5 file.}

\item{h5name}{HDF5 object name.}
}
\description{
Write attributes to a HDF5 object.
}
\keyword{internal}

