% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{writePhenoCSV}
\alias{writePhenoCSV}
\title{Write yeast pheno data to a CSV file.}
\usage{
writePhenoCSV(pheno, outfile, digits = NULL)
}
\arguments{
\item{pheno}{An \pkg{R/qtl} \code{cross} \code{pheno} object.}

\item{outfile}{Output CSV file path.}

\item{digits}{If specified, round numeric phenotype
values to the specified number of digits.}
}
\description{
This function writes a yeast \code{pheno} object to an \pkg{R/qtl} CSV file.
The \code{pheno} object must have an attribute \code{'info'} of type
\code{\linkS4class{CrossInfo}}, from which phenotype and sample IDs are taken.
}
\seealso{
Other CSV functions: \code{\link{hasMapCSV}},
  \code{\link{readCovarCSV}}, \code{\link{readCrossCSV}},
  \code{\link{readGenoCSV}}, \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{readPhenoCSV}}, \code{\link{recodeCSV}},
  \code{\link{sniffCSV}}, \code{\link{writeCrossCSV}},
  \code{\link{writeGenoCSV}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapframeCSV}}
}
