% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{writeReportExcel}
\alias{writeReportExcel}
\title{Write an Excel report of QTL scan results.}
\usage{
writeReportExcel(scanfile, report, phenotypes = NULL, analyses = NULL,
  worksheets = NULL, scanfile.pattern = NULL)
}
\arguments{
\item{scanfile}{A QTL scan result HDF5 file.}

\item{report}{Path of output report Excel file.}

\item{phenotypes}{Phenotypes for which results should be included in the
output file. If none are specified, results are output for all phenotypes.}

\item{analyses}{Analyses for which results should be included in the output
file. If none are specified, results are output for all available analyses.}

\item{worksheets}{Worksheets to include in the output file. If none are
specified, default worksheets are output. To view available worksheets,
call the function \code{\link{listWorksheets}}.}

\item{scanfile.pattern}{Pattern for extracting experiment info from scan file
names. This must be a valid Perl regex with named capture groups. Neither the
capture groups nor the pattern itself are required to match any given scan
file, but all capture groups must have a name, and that name cannot clash
with other names that might be used alongside the extracted information.}
}
\description{
Write an Excel report of QTL scan results.
}
\author{
Yue Hu

Thomas Walsh <tw164@le.ac.uk>
}
\seealso{
Other Excel functions: \code{\link{listWorksheets}},
  \code{\link{writeDigestExcel}}
}

