% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{writeResultHDF5}
\alias{writeResultHDF5}
\title{Write QTL analysis result to a HDF5 file.}
\usage{
writeResultHDF5(result, outfile, phenotype, name = NULL)
}
\arguments{
\item{result}{R object containing a QTL analysis result.}

\item{outfile}{An output HDF5 file.}

\item{phenotype}{Name of phenotype (or equivalent analysis unit).}

\item{name}{Name of QTL analysis result.}
}
\description{
Write QTL analysis result to a HDF5 file.
}
\keyword{internal}

