% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{writeWorkbookExcel}
\alias{writeWorkbookExcel}
\title{Write an Excel workbook of QTL analysis results.}
\usage{
writeWorkbookExcel(scanfiles, workbook, phenotypes = NULL, analyses = NULL,
  worksheets = NULL, scanfile.pattern = NULL)
}
\arguments{
\item{scanfiles}{One or more QTL scan result HDF5 files.}

\item{workbook}{Path of output Excel file.}

\item{phenotypes}{Phenotypes for which results should be included in the
output file. If none are specified, results are output for all phenotypes.}

\item{analyses}{Analyses for which results should be included in the output
file. If none are specified, results are output for all available analyses.}

\item{worksheets}{Worksheets to include in the output file. If none are
specified, default worksheets are output. To view available worksheets,
call the function \code{\link{listWorksheets}}.}

\item{scanfile.pattern}{Pattern for extracting experiment info from scan file
names. This must be a valid Perl regex with named capture groups. Neither the
capture groups nor the pattern itself are required to match any given scan
file, but all capture groups must have a name, and that name cannot clash
with other names that might be used alongside the extracted information.}
}
\description{
Write an Excel workbook of QTL analysis results.
}
\keyword{internal}

