% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{allNA}
\alias{allNA}
\title{Test if all elements are NA values.}
\usage{
allNA(x)
}
\arguments{
\item{x}{Test vector.}
}
\value{
TRUE if vector is of length zero or contains only NA values;
FALSE otherwise.
}
\description{
Test if all elements are NA values.
}
\keyword{internal}

