% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{allNA}
\alias{allNA}
\title{Test if all elements are \code{NA} values.}
\usage{
allNA(x)
}
\arguments{
\item{x}{Test vector.}
}
\value{
\code{TRUE} if vector is of length zero or
contains only \code{NA} values; \code{FALSE} otherwise.
}
\description{
Test if all elements are \code{NA} values.
}
\author{
Thomas Walsh <tw164@le.ac.uk>

Yue Hu
}
\keyword{internal}
