% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{as.data.frame.mapframe}
\alias{as.data.frame.mapframe}
\title{Convert \code{mapframe} to \code{data.frame}.}
\usage{
\method{as.data.frame}{mapframe}(x, ...)
}
\arguments{
\item{x}{A \code{mapframe} object.}

\item{...}{Unused arguments.}
}
\value{
A \code{data.frame} corresponding to the input \code{mapframe} object.
}
\description{
Convert \code{mapframe} to \code{data.frame}.
}
\seealso{
Other map utility functions: \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{hasMapCSV}}, \code{\link{inMapOrder}},
  \code{\link{inferMapStep}}, \code{\link{mapframe}},
  \code{\link{orderMap}}, \code{\link{readMapCSV}},
  \code{\link{readMapHDF5}}, \code{\link{readMapframeCSV}},
  \code{\link{setMapUnit}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapHDF5}},
  \code{\link{writeMapframeCSV}}
}

