% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{as.geno}
\alias{as.geno}
\alias{as.geno.data.frame}
\title{Coerce to \code{geno} object.}
\usage{
as.geno(from)

\method{as.geno}{data.frame}(from)
}
\arguments{
\item{from}{Object containing genotype data.}
}
\value{
A \code{geno} object corresponding to the input object.
}
\description{
Coerce to \code{geno} object.
}
\keyword{internal}

