% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{as.geno}
\alias{as.geno}
\alias{as.geno.data.frame}
\title{Coerce to \code{geno} object.}
\usage{
as.geno(from, missing.value = "-")

\method{as.geno}{data.frame}(from, missing.value = "-")
}
\arguments{
\item{from}{Object containing genotype data.}

\item{missing.value}{Missing data value. This can be any single character
that is not a possible phenotype or genotype value.}
}
\value{
An \code{geno} object corresponding to the input object.
}
\description{
Coerce to \code{geno} object.
}
\keyword{internal}

