% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{as.mapframe}
\alias{as.mapframe}
\alias{as.mapframe.data.frame}
\alias{as.mapframe.list}
\alias{as.mapframe.map}
\alias{as.mapframe.scanone}
\title{Coerce object to \code{mapframe}.}
\usage{
as.mapframe(from, ...)

\method{as.mapframe}{data.frame}(from, map.unit = NULL)

\method{as.mapframe}{list}(from, map.unit = NULL)

\method{as.mapframe}{map}(from, map.unit = NULL)

\method{as.mapframe}{scanone}(from)
}
\arguments{
\item{from}{Object containing map data.}

\item{...}{Further arguments (see below).}

\item{map.unit}{Explicitly sets \code{'map.unit'}. This must be specified if 
an object does not have an existing \code{'map.unit'} attribute.}
}
\value{
A \code{mapframe} corresponding to the input object.
}
\description{
Coerce object to \code{mapframe}.
}
\seealso{
Other map utilities: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{inMapOrder}}, \code{\link{inferMapStep}},
  \code{\link{mapframe}}, \code{\link{orderMap}},
  \code{\link{setMapUnit}}
}

