% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScanone.R
\name{batchPhenoScanone}
\alias{batchPhenoScanone}
\title{Run \code{qtl::scanone} on a batch of phenotypes.}
\usage{
batchPhenoScanone(cross, pheno.col = NULL, n.cluster = 1, iseed = NULL,
  ...)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{pheno.col}{Phenotype columns for which QTL analysis should be run.
If no phenotypes are specified, all are used.}

\item{n.cluster}{Number of nodes to request for parallel execution. Any
number of processes can be requested with the argument \code{n.cluster},
but the number actually returned can be limited to the number of processes
available. On a local computer, this limit is the number of physical cores
in the computer. In a PBS batch system, the maximum number of processes
will be set from the PBS node file.}

\item{iseed}{Seed for random number generator.}

\item{...}{Additional keyword arguments passed to \code{scanone}.}
}
\value{
A \code{scanone} object containing the results of \code{scanone} 
for the given phenotypes.
}
\description{
Run \code{qtl::scanone} on a batch of phenotypes.
}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})
}
\seealso{
Other scan utility functions: \code{\link{batchPermScanoneF}},
  \code{\link{batchPermScanone}},
  \code{\link{batchPermScantwoF}},
  \code{\link{batchPermScantwo}},
  \code{\link{batchPhenoScantwo}}, \code{\link{batchScan}},
  \code{\link{nodePermScanoneF}},
  \code{\link{nodePermScanone}},
  \code{\link{nodePermScantwoF}},
  \code{\link{nodePermScantwo}},
  \code{\link{nodePhenoScanone}},
  \code{\link{nodePhenoScantwo}}
}
