% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCrossInfo.R
\name{compareCrossInfo}
\alias{compareCrossInfo}
\alias{compareCrossInfo.cross}
\alias{compareCrossInfo.geno}
\title{Compare object with associated \code{CrossInfo}.}
\usage{
compareCrossInfo(x, cross.info = NULL)

\method{compareCrossInfo}{cross}(x, cross.info = NULL)

\method{compareCrossInfo}{geno}(x, cross.info = NULL)
}
\arguments{
\item{x}{Test object to be compared.}

\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
TRUE if \code{CrossInfo} and test object are concordant; otherwise
returns a character vector of mismatch errors.
}
\description{
Test concordance of an object with its associated 
\code{CrossInfo} object. If no \code{CrossInfo} object is specified,
it will be extracted from the test object.
}
\seealso{
\code{\linkS4class{CrossInfo}}
}
\keyword{internal}

