% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{crossesEqual}
\alias{crossesEqual}
\title{Test if two \pkg{R/qtl} \code{cross} objects are essentially equal.}
\usage{
crossesEqual(cross1, cross2, tol = 1e-05)
}
\arguments{
\item{cross1}{An \pkg{R/qtl} \code{cross} object.}

\item{cross2}{An \pkg{R/qtl} \code{cross} object.}

\item{tol}{Tolerance for comparing numeric values.}
}
\value{
Returns \code{TRUE} if the two input \code{cross} objects are
essentially equal. Otherwise, this function returns a character vector
describing an observed difference between the two \code{cross} objects.
}
\description{
This uses the \pkg{R/qtl} function \code{comparecrosses} to check if two
input \code{cross} objects are essentially equal, with key elements being
identical, and numeric values being equal (within the given tolerance).
}
\seealso{
\href{http://www.rqtl.org/manual/qtl-manual.pdf}{R/qtl manual}
}

