% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeGenotypes.R
\name{decodeGenotypes}
\alias{decodeGenotypes}
\alias{decodeGenotypes.data.frame}
\alias{decodeGenotypes.integer}
\alias{decodeGenotypes.matrix}
\title{Decode genotype integers to genotype strings.}
\usage{
decodeGenotypes(x, genotypes)

\method{decodeGenotypes}{integer}(x, genotypes)

\method{decodeGenotypes}{data.frame}(x, genotypes)

\method{decodeGenotypes}{matrix}(x, genotypes)
}
\arguments{
\item{x}{Object containing genotype integers.}

\item{genotypes}{Sorted character vector of genotype symbols.
Input genotype integers should be valid indices of this vector.}
}
\value{
Object of same class and dimensions as input object,
but with genotype integers decoded to genotype strings.
}
\description{
Decode genotype integers to genotype strings.
}
\keyword{internal}

