% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{emptyArgs}
\alias{emptyArgs}
\title{Test if ellipsis arguments are empty.}
\usage{
emptyArgs(...)
}
\arguments{
\item{...}{Ellipsis arguments.}
}
\value{
\code{TRUE} if there are no ellipsis arguments;
\code{FALSE} otherwise.
}
\description{
Test if ellipsis arguments are empty.
}
\keyword{internal}

