% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{formatSeq}
\alias{formatSeq}
\title{Format sequence labels.}
\usage{
formatSeq(x, prefix = c("", "c", "chr"), use.roman = TRUE)
}
\arguments{
\item{x}{Vector of sequence labels.}

\item{prefix}{Chromosome prefix.}

\item{use.roman}{Use the Roman numeral form of the chromosome.}
}
\value{
Vector of formatted sequence labels.
}
\description{
Format sequence labels.
}
\seealso{
Other chromosome/sequence functions: \code{\link{formatChr}},
  \code{\link{isNormChr}}, \code{\link{isNormSeq}},
  \code{\link{normChr}}, \code{\link{normSeq}},
  \code{\link{orderChr}}, \code{\link{orderSeq}},
  \code{\link{rankChr}}, \code{\link{rankSeq}},
  \code{\link{sortChr}}, \code{\link{sortSeq}}
}

