% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getColIndices}
\alias{getColIndices}
\title{Get column indices of object.}
\usage{
getColIndices(x)
}
\arguments{
\item{x}{An object with columns.}
}
\value{
Integer vector of all column indices. Returns an empty integer vector
if the object has zero columns.
}
\description{
Get column indices of object.
}
\keyword{internal}

