% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getCrossType}
\alias{getCrossType}
\alias{getCrossType,CrossInfo-method}
\title{Get cross type.}
\usage{
getCrossType(cross.info)

\S4method{getCrossType}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
Cross type name.
}
\description{
Get cross type.
}
\keyword{internal}

