% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getGenotypes}
\alias{getGenotypes}
\alias{getGenotypes,CrossInfo-method}
\title{Get genotype symbols.}
\usage{
getGenotypes(cross.info)

\S4method{getGenotypes}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
Character vector of genotype symbols.
}
\description{
Get genotype symbols.
}
\seealso{
Other CrossInfo methods: \code{\link{getAlleles}},
  \code{\link{getCrosstype}},
  \code{\link{getMarkerIndices}},
  \code{\link{getMarkerNames}},
  \code{\link{getMarkerSeqs}}, \code{\link{getMarkers}},
  \code{\link{getNumMarkers}},
  \code{\link{getNumPhenotypes}},
  \code{\link{getNumSamples}}, \code{\link{getNumSeqs}},
  \code{\link{getPhenotypeIndices}},
  \code{\link{getPhenotypeNames}},
  \code{\link{getPhenotypes}},
  \code{\link{getSampleIndices}},
  \code{\link{getSampleNames}}, \code{\link{getSamples}},
  \code{\link{getSeqIndices}}, \code{\link{getSeqMarkers}},
  \code{\link{getSeqNames}}, \code{\link{getSequences}},
  \code{\link{getStrainIndices}},
  \code{\link{getTetradIndices}},
  \code{\link{hasMarkerSeqs}}, \code{\link{hasSampleIDs}},
  \code{\link{hasStrainIndices}},
  \code{\link{hasTetradIndices}}, \code{\link{setAlleles}},
  \code{\link{setCrosstype}}, \code{\link{setGenotypes}},
  \code{\link{setMarkerSeqs}}, \code{\link{setMarkers}},
  \code{\link{setPhenotypes}}, \code{\link{setSamples}},
  \code{\link{setSequences}},
  \code{\link{setStrainIndices}},
  \code{\link{setTetradIndices}}
}
\keyword{internal}

