% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getIndices}
\alias{getIndices}
\alias{getIndices.default}
\alias{getIndices.qtl}
\title{Get indices of object elements.}
\usage{
getIndices(x, requested = NULL, strict = FALSE)

\method{getIndices}{default}(x, requested = NULL, strict = FALSE)

\method{getIndices}{qtl}(x, requested = NULL, strict = FALSE)
}
\arguments{
\item{x}{An object with elements that are accessible by an index.}

\item{requested}{A character vector of names, a logical vector of the same
length as the input object, or a numeric vector containing indices of the
input object. If this parameter is not specified, all indices are returned.}

\item{strict}{Option indicating that \code{requested}, if specified, must
be in the same order as the corresponding elements of the input object.}
}
\value{
Integer vector of object indices.
}
\description{
Get the indices of an object, as constrained by the \code{requested}
parameter. If using this function without \code{requested} constraints,
consider using the faster primitive R function \code{seq_along} instead.
}
\keyword{internal}
