% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLODProfile.R
\name{getLODProfile}
\alias{getLODProfile}
\alias{getLODProfile.data.frame}
\alias{getLODProfile.qtl}
\alias{getLODProfile.scantwo}
\title{Get LOD profile.}
\usage{
## Generic method.
getLODProfile(x, ...)

\method{getLODProfile}{data.frame}(x, lodcolumn = NULL, ...)

\method{getLODProfile}{qtl}(x, qtl.index = NULL, ...)

\method{getLODProfile}{scantwo}(x, lodcolumn = NULL, ...)
}
\arguments{
\item{x}{Object containing single-QTL LOD profile data.}

\item{...}{Further arguments (see below).}

\item{lodcolumn}{In a \code{scanone} or equivalent object, this
parameter indicates which LOD column to consider. In other objects
with LOD-column-associated elements, this parameter indicates which
LOD-column-associated element to consider. This must be either a LOD
column name or an index \emph{with respect to the set of LOD columns}.
If no LOD column is specified and one such column is found, that column
is used by default; otherwise a LOD column must be specified.}

\item{qtl.index}{In a \code{qtl} object, this indicates
the QTL for which a LOD profile should be returned.}
}
\value{
A \code{scanone} object containing
a LOD profile for a single phenotype.
}
\description{
Get LOD profile.
}

