% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{getMetadataCSV}
\alias{getMetadataCSV}
\title{Get parameters of \pkg{R/qtl} input data.}
\usage{
getMetadataCSV(x)
}
\arguments{
\item{x}{An \pkg{R/qtl} input data frame as read from a CSV file.}
}
\value{
A list containing parameters of \pkg{R/qtl} input data.
}
\description{
Given a data frame as loaded from an \pkg{R/qtl} input CSV file,
this function determines the parameters of that data (e.g. phenotype columns).
}
\details{
The returned list includes information on the following parameters:

\itemize{
 \item{\code{pheno.cols}}{Indices of phenotype columns, or NULL if none found.}
 \item{\code{id.col}}{Index of ID column, or NA if not found.}
 \item{\code{geno.cols}}{Indices of genotype columns, or NULL if none found.}
 \item{\code{dat.rows}}{Row indices of data.}
 \item{\code{map.present}}{TRUE if the data contains a map; FALSE otherwise.}
 \item{\code{class}}{Expected class of the input data.}
}

This information can then guide further processing of the input data.
}
\keyword{internal}

