% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getNumPhenotypes}
\alias{getNumPhenotypes}
\alias{getNumPhenotypes,CrossInfo-method}
\title{Get the number of phenotypes.}
\usage{
getNumPhenotypes(cross.info)

\S4method{getNumPhenotypes}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
Number of phenotypes.
}
\description{
Get the number of phenotypes.
}
\keyword{internal}

