% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getPhenoColIndices}
\alias{getPhenoColIndices}
\alias{getPhenoColIndices.cross}
\alias{getPhenoColIndices.data.frame}
\title{Get phenotype column indices.}
\usage{
getPhenoColIndices(x, pheno.col = NULL)

\method{getPhenoColIndices}{cross}(x, pheno.col = NULL)

\method{getPhenoColIndices}{data.frame}(x, pheno.col = NULL)
}
\arguments{
\item{x}{A \code{data.frame} containing phenotype data, or an \pkg{R/qtl}
\code{cross} object containing such a \code{data.frame}.}

\item{pheno.col}{Vector indicating the phenotype indices to return. This can
be an integer vector of phenotype column indices with respect to the columns
of the \code{data.frame} (which may or may not be within a \code{cross}
object), or a character vector that contains phenotype IDs or their
syntactically valid names. If none are specified, all phenotype column
indices are returned.}
}
\value{
Phenotype column indices.
}
\description{
Get the indices of the phenotype columns in the given object. In a
\code{data.frame}, phenotype columns are those columns that do not have
reserved phenotype headings (e.g. \code{'ID'}). In a \code{cross} object,
these are the corresponding columns of the phenotype \code{data.frame}.
}
\seealso{
Other cross object functions: \code{\link{crossesEqual}},
  \code{\link{getIdColIndex}},
  \code{\link{hasTimeSeriesPhenotypes}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{permIndices}}, \code{\link{pull.alleles}},
  \code{\link{pull.chr}}, \code{\link{pull.crosstype}},
  \code{\link{pull.ind}}, \code{\link{readCrossCSV}},
  \code{\link{readCrossHDF5}}, \code{\link{writeCrossCSV}},
  \code{\link{writeCrossHDF5}}
}
