% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getPhenotypeNames}
\alias{getPhenotypeNames}
\alias{getPhenotypeNames,CrossInfo-method}
\title{Get phenotype names.}
\usage{
getPhenotypeNames(cross.info, phenotypes = NULL)

\S4method{getPhenotypeNames}{CrossInfo}(cross.info, phenotypes = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{phenotypes}{Vector indicating phenotypes to consider. This may be a 
logical vector of phenotypes, an integer vector of phenotype indices, or 
a character vector that contains phenotype IDs or their equivalent 
syntactically valid name. If none are specified, results are returned 
for all phenotypes.}
}
\value{
Character vector of phenotype names.
}
\description{
Get phenotype names.
}
\keyword{internal}

