% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{getPkgPipelineNames}
\alias{getPkgPipelineNames}
\title{Get \pkg{shmootl} pipeline names.}
\usage{
getPkgPipelineNames(group = NULL)
}
\arguments{
\item{group}{Pipeline group for which pipeline names should be retrieved.
If no pipeline group is specified, all pipeline names are returned.}
}
\value{
Character vector of \pkg{shmootl} pipeline names.
}
\description{
Package pipeline names are taken from the loaded \pkg{shmootl} namespace, if
available; otherwise they are taken from installed \pkg{shmootl} package.
}

