% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQTLFeatures.R
\name{getQTLFeatures}
\alias{getQTLFeatures}
\title{Get annotation of features within QTL intervals.}
\usage{
getQTLFeatures(qtl.intervals, features)
}
\arguments{
\item{qtl.intervals}{A \code{qtlintervals} object.}

\item{features}{A \code{data.frame} of primary genome features, as returned
by \code{\link{readFeaturesGFF}}. For each feature, this \code{data.frame}
must include the reference sequence (\code{'chr'}), 1-offset base-pair
positions of the start and end of the feature (\code{'start'} and
\code{'end'} respectively), and the feature \code{'ID'}.}
}
\value{
An object of class \code{qtlfeatures}, which is essentially
a list of \code{data.frame} objects, each containing the set of genome
features in a given QTL interval.
}
\description{
Get annotation of features within QTL intervals.
}

