% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQTLIntervals.R
\name{getQTLIntervals}
\alias{getQTLIntervals}
\alias{getQTLIntervals.mapframe}
\alias{getQTLIntervals.qtl}
\alias{getQTLIntervals.scanone}
\title{Get QTL intervals.}
\usage{
## Generic method.
getQTLIntervals(x, chr=NULL, drop=1.5, expandtomarkers=FALSE, ...)

\method{getQTLIntervals}{mapframe}(x, chr = NULL, drop = 1.5,
  expandtomarkers = FALSE, threshold = NULL, alpha = NULL, fdr = NULL,
  lodcolumn = NULL, qtl.peaks = NULL, ...)

\method{getQTLIntervals}{qtl}(x, chr = NULL, drop = 1.5,
  expandtomarkers = FALSE, qtl.indices = NULL, ...)

\method{getQTLIntervals}{scanone}(x, chr = NULL, drop = 1.5,
  expandtomarkers = FALSE, threshold = NULL, alpha = NULL, fdr = NULL,
  lodcolumn = NULL, qtl.peaks = NULL, ...)
}
\arguments{
\item{x}{A \code{scanone} or \code{qtl} object.}

\item{chr}{Vector indicating which sequences to consider. If no sequences
are specified, QTL intervals are returned for all available sequences.}

\item{drop}{LOD units that the LOD profile must drop to form the interval.}

\item{expandtomarkers}{Expand the LOD interval to the nearest flanking 
markers, or to the respective terminal loci.}

\item{...}{Further arguments (see below).}

\item{threshold}{In a \code{scanone} or equivalent object, this indicates
the LOD significance threshold for QTL intervals.}

\item{alpha}{In a \code{scanone} or equivalent object, this contains the
the significance level of the given threshold. (Mutually exclusive
with \code{fdr}.)}

\item{fdr}{In a \code{scanone} or equivalent object, this contains
the false discovery rate (FDR) of the given threshold. (Mutually exclusive
with \code{alpha}.)}

\item{lodcolumn}{In a \code{scanone} or equivalent object, this parameter 
indicates which LOD column to consider. This must be either a column name 
or an index \emph{with respect to the set of LOD columns}. If no LOD column 
is specified and one such column is found, that column is used by default; 
otherwise a LOD column must be specified.}

\item{qtl.peaks}{Locus \code{mapframe} indicating the location of the QTL 
peaks in a \code{scanone} result. If not specified, these are found using 
the LOD profile of the \code{scanone} object.}

\item{qtl.indices}{In a \code{qtl} object, this option indicates the QTLs 
for which a QTL interval should be returned.}
}
\value{
An object of class \code{qtlintervals}, which is essentially a list
of \code{mapframe} objects, each containing three rows of information about
the lower QTL interval limit, QTL peak, and upper QTL interval limit, 
respectively. Returns an empty \code{qtlintervals} object if there are no
significant QTLs.
}
\description{
Get QTL intervals.
}

