% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getSampleIndices}
\alias{getSampleIndices}
\alias{getSampleIndices,CrossInfo-method}
\title{Get sample indices.}
\usage{
getSampleIndices(cross.info, samples = NULL, strains = NULL,
  tetrads = NULL, simplify = FALSE)

\S4method{getSampleIndices}{CrossInfo}(cross.info, samples = NULL,
  strains = NULL, tetrads = NULL, simplify = FALSE)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{samples}{Vector indicating samples to consider. This may be a 
logical vector of samples, an integer vector of sample indices, or 
a character vector that contains sample IDs or their equivalent 
syntactically valid name.}

\item{strains}{Vector indicating strains to consider. This may be a 
logical vector of strains, an integer vector of strain indices, or 
a character vector that contains strain IDs or their equivalent 
syntactically valid name.}

\item{tetrads}{Vector indicating tetrads to consider. This may be a 
logical vector of tetrads, or an integer vector of tetrad indices.}

\item{simplify}{Simplify list return value to a vector.}
}
\value{
If samples are specified by strain or tetrad, and if the option 
\code{simplify} is not \code{TRUE}, this method returns a list of integer
vectors, each containing the sample indices corresponding to a given
strain/tetrad. Otherwise, a vector of sample indices is returned.
}
\description{
Get indices of the specified samples. Samples can be specified at one of the 
following levels: sample, strain, or tetrad. If none are specified, all 
sample indices are returned.
}
\seealso{
Other CrossInfo methods: \code{\link{getAlleles}},
  \code{\link{getCrosstype}}, \code{\link{getGenotypes}},
  \code{\link{getMarkerIndices}},
  \code{\link{getMarkerNames}},
  \code{\link{getMarkerSeqs}}, \code{\link{getMarkers}},
  \code{\link{getNumMarkers}},
  \code{\link{getNumPhenotypes}},
  \code{\link{getNumSamples}}, \code{\link{getNumSeqs}},
  \code{\link{getPhenotypeIndices}},
  \code{\link{getPhenotypeNames}},
  \code{\link{getPhenotypes}},
  \code{\link{getSampleNames}}, \code{\link{getSamples}},
  \code{\link{getSeqIndices}}, \code{\link{getSeqMarkers}},
  \code{\link{getSeqNames}}, \code{\link{getSequences}},
  \code{\link{getStrainIndices}},
  \code{\link{getTetradIndices}},
  \code{\link{hasMarkerSeqs}}, \code{\link{hasSampleIDs}},
  \code{\link{hasStrainIndices}},
  \code{\link{hasTetradIndices}}, \code{\link{setAlleles}},
  \code{\link{setCrosstype}}, \code{\link{setGenotypes}},
  \code{\link{setMarkerSeqs}}, \code{\link{setMarkers}},
  \code{\link{setPhenotypes}}, \code{\link{setSamples}},
  \code{\link{setSequences}},
  \code{\link{setStrainIndices}},
  \code{\link{setTetradIndices}}
}
\keyword{internal}

