% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getSeqMarkers}
\alias{getSeqMarkers}
\alias{getSeqMarkers,CrossInfo-method}
\title{Get markers by sequence.}
\usage{
getSeqMarkers(cross.info, sequences = NULL, simplify = FALSE)

\S4method{getSeqMarkers}{CrossInfo}(cross.info, sequences = NULL,
  simplify = FALSE)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{sequences}{Vector indicating sequences to consider. Sequences may 
alternatively be specified by a logical vector. If none are specified, 
results are returned for all sequences.}

\item{simplify}{Simplify list return value to a vector.}
}
\value{
List of character vectors, each containing the marker IDs
corresponding to a given sequence. If option \code{simplify} is \code{TRUE},
this is simplified to a vector. Returns \code{NULL} if no sequence-marker 
information is available.
}
\description{
Get markers by sequence.
}
\seealso{
Other CrossInfo methods: \code{\link{getAlleles}},
  \code{\link{getCrosstype}}, \code{\link{getGenotypes}},
  \code{\link{getMarkerIndices}},
  \code{\link{getMarkerNames}},
  \code{\link{getMarkerSeqs}}, \code{\link{getMarkers}},
  \code{\link{getNumMarkers}},
  \code{\link{getNumPhenotypes}},
  \code{\link{getNumSamples}}, \code{\link{getNumSeqs}},
  \code{\link{getPhenotypeIndices}},
  \code{\link{getPhenotypeNames}},
  \code{\link{getPhenotypes}},
  \code{\link{getSampleIndices}},
  \code{\link{getSampleNames}}, \code{\link{getSamples}},
  \code{\link{getSeqIndices}}, \code{\link{getSeqNames}},
  \code{\link{getSequences}},
  \code{\link{getStrainIndices}},
  \code{\link{getTetradIndices}},
  \code{\link{hasMarkerSeqs}}, \code{\link{hasSampleIDs}},
  \code{\link{hasStrainIndices}},
  \code{\link{hasTetradIndices}}, \code{\link{setAlleles}},
  \code{\link{setCrosstype}}, \code{\link{setGenotypes}},
  \code{\link{setMarkerSeqs}}, \code{\link{setMarkers}},
  \code{\link{setPhenotypes}}, \code{\link{setSamples}},
  \code{\link{setSequences}},
  \code{\link{setStrainIndices}},
  \code{\link{setTetradIndices}}
}
\keyword{internal}

