% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{getSeqNames}
\alias{getSeqNames}
\alias{getSeqNames,CrossInfo-method}
\title{Get vector of sequence names.}
\usage{
getSeqNames(cross.info, sequences = NULL)

\S4method{getSeqNames}{CrossInfo}(cross.info, sequences = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{sequences}{Vector indicating sequences to consider. Sequences may 
alternatively be specified by a logical vector. If none are specified, 
results are returned for all sequences.}
}
\value{
Vector of sequence names.
}
\description{
Get vector of sequence names.
}
\keyword{internal}

