% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{h5writeAttributes}
\alias{h5writeAttributes}
\title{Write attributes to a HDF5 object.}
\usage{
h5writeAttributes(x, h5obj)
}
\arguments{
\item{x}{List of attributes, or an R object whose attributes will be written.}

\item{h5obj}{HDF5 object to which attributes will be assigned.}
}
\description{
Write attributes to a HDF5 object.
}
\keyword{internal}

