% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasEnumGenotypes.R
\name{hasEnumGenotypes}
\alias{hasEnumGenotypes}
\alias{hasEnumGenotypes.cross}
\title{Test if object contains enumerated genotypes.}
\usage{
hasEnumGenotypes(x)

\method{hasEnumGenotypes}{cross}(x)
}
\arguments{
\item{x}{Object containing genotype data.}
}
\value{
TRUE if object contains enumerated genotypes; FALSE otherwise.
}
\description{
Test if object contains enumerated genotypes.
}
\keyword{internal}

