% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{hasMapCSV}
\alias{hasMapCSV}
\title{Test if CSV file contains map data.}
\usage{
hasMapCSV(infile)
}
\arguments{
\item{infile}{Input CSV file path.}
}
\value{
TRUE if CSV file contains map data in a known format;
FALSE otherwise.
}
\description{
Test if CSV file contains map data.
}
\keyword{internal}

