% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hasNames}
\alias{hasNames}
\title{Test if object has names.}
\usage{
hasNames(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if object has nonempty names with no NA values; FALSE otherwise.
}
\description{
Test if object has names.
}
\keyword{internal}

