% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{interpTimeSeries}
\alias{interpTimeSeries}
\title{Interpolate gaps in a time-series.}
\usage{
interpTimeSeries(cross, tol = 1e-05)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{tol}{Tolerance for time step equality.}
}
\value{
The input \code{cross} object is returned with gaps in time-series
phenotypes filled with values interpolated from the gap endpoints.
}
\description{
Interpolate values for gaps in time-series
phenotypes of a \code{cross} object.
}
\section{Time-Series Phenotypes}{


A set of phenotypes can be designated as a time-series by naming each
phenotype with the time point at which phenotype observations were made
(e.g. \code{'0.0'}, \code{'1.0'}, \code{'2.0'}). Time points can be in
any unit, but must be non-negative, monotonically increasing, and have
a consistent time step. If some time points are missing, the resulting
gap in time must be a multiple of the time step.
}
\seealso{
Other cross object functions: \code{\link{crossesEqual}},
  \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{hasTimeSeriesPhenotypes}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}}, \code{\link{padTimeSeries}},
  \code{\link{permCross}}, \code{\link{permIndices}},
  \code{\link{pull.alleles}}, \code{\link{pull.chr}},
  \code{\link{pull.crosstype}}, \code{\link{pull.ind}},
  \code{\link{readCrossCSV}}, \code{\link{readCrossHDF5}},
  \code{\link{writeCrossCSV}}, \code{\link{writeCrossHDF5}}

Other time-series functions: \code{\link{hasTimeSeriesPhenotypes}},
  \code{\link{inferTimeStep}}, \code{\link{padTimeSeries}}
}

