% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isDefaultMarkerID}
\alias{isDefaultMarkerID}
\title{Test for default marker IDs.}
\usage{
isDefaultMarkerID(ids)
}
\arguments{
\item{ids}{Vector of locus IDs.}
}
\value{
Logical vector indicating which elements are valid
IDs that follow the pattern of a default marker ID.
}
\description{
Test for default marker IDs.
}
\section{Default Marker IDs}{


Default marker IDs are assigned to loci of a physical map if no other marker
ID is available. These are of the form \code{'c04:0108900'}, where
\code{'04'} is a zero-padded two-digit chromosome number and \code{'0108900'}
is a zero-padded seven-digit number giving the physical map position of the
given locus (i.e. the nucleotide position of the locus in the reference
sequence).
}

