% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isDefaultQTLName}
\alias{isDefaultQTLName}
\title{Test for default QTL names.}
\usage{
isDefaultQTLName(ids)
}
\arguments{
\item{ids}{Vector of item IDs.}
}
\value{
Logical vector indicating which elements are default QTL names.
}
\description{
Test for default QTL names.
}

