% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isFounderGenotype}
\alias{isFounderGenotype}
\title{Test if symbol is a valid founder genotype.}
\usage{
isFounderGenotype(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if symbol is a valid founder genotype; FALSE otherwise.
}
\description{
Test if symbol is a valid founder genotype.
}
\keyword{internal}

