% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isMarkerID}
\alias{isMarkerID}
\title{Test for marker IDs.}
\usage{
isMarkerID(loc.ids)
}
\arguments{
\item{loc.ids}{Vector of locus IDs.}
}
\value{
Logical vector indicating which elements are marker IDs.
}
\description{
Test for marker IDs.
}

