% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{isNormChr}
\alias{isNormChr}
\title{Test for normalised chromosome labels.}
\usage{
isNormChr(x)
}
\arguments{
\item{x}{Vector of chromosome labels.}
}
\value{
Logical vector indicating which chromosome labels are normalised.
}
\description{
Test for normalised chromosome labels.
}
\keyword{internal}

