% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isPseudomarkerID}
\alias{isPseudomarkerID}
\title{Test for pseudomarker IDs.}
\usage{
isPseudomarkerID(loc.ids)
}
\arguments{
\item{loc.ids}{Vector of locus IDs.}
}
\value{
Logical vector indicating which elements are pseudomarker IDs.
}
\description{
Test for pseudomarker IDs.
}

