% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isRawAllele}
\alias{isRawAllele}
\title{Test if symbol is a valid raw SNP allele.}
\usage{
isRawAllele(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if symbol is a valid raw SNP allele;
\code{FALSE} otherwise.
}
\description{
Test if symbol is a valid raw SNP allele.
}
\keyword{internal}

