% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isRawGenotype}
\alias{isRawGenotype}
\title{Test if symbol is a valid raw genotype.}
\usage{
isRawGenotype(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if symbol is a valid raw genotype; FALSE otherwise.
}
\description{
Test if symbol is a valid raw genotype.
}
\keyword{internal}

