% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isRawGenotype}
\alias{isRawGenotype}
\title{Test if symbol is a valid raw SNP genotype.}
\usage{
isRawGenotype(x, strict = FALSE)
}
\arguments{
\item{x}{Test object.}

\item{strict}{Return TRUE only for complete genotypes.}
}
\value{
TRUE if symbol is a valid raw SNP genotype; FALSE otherwise.
}
\description{
Test if symbol is a valid raw SNP genotype.
}
\keyword{internal}

