% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isSingleFiniteNumber}
\alias{isSingleFiniteNumber}
\title{Test for a single finite number.}
\usage{
isSingleFiniteNumber(n)
}
\arguments{
\item{n}{Test object.}
}
\value{
\code{TRUE} if the object is a single finite number;
\code{FALSE} otherwise.
}
\description{
Test for a single finite number.
}
\keyword{internal}
