% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isValidGenotype}
\alias{isValidGenotype}
\title{Test if symbol is a valid genotype.}
\usage{
isValidGenotype(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
TRUE if symbol is a valid genotype; FALSE otherwise.
}
\description{
Test if symbol is a valid genotype.
}
\keyword{internal}

