% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isValidID}
\alias{isValidID}
\title{Test identifier validity.}
\usage{
isValidID(x)
}
\arguments{
\item{x}{Test vector.}
}
\value{
Logical vector indicating which elements are valid identifiers.
}
\description{
Test identifier validity.
}

