% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isValidName}
\alias{isValidName}
\title{Test for syntactically valid names.}
\usage{
isValidName(x)
}
\arguments{
\item{x}{Test vector.}
}
\value{
Logical vector indicating which elements are syntactically valid 
names.
}
\description{
Test for syntactically valid names.
}

