% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makeDefaultQTLNames}
\alias{makeDefaultQTLNames}
\title{Make default QTL names for the specified loci.}
\usage{
makeDefaultQTLNames(loc, step = 0)
}
\arguments{
\item{loc}{Locus \code{data.frame} specifying genetic map positions.}

\item{step}{Map step size.}
}
\value{
Character vector of default QTL names.
}
\description{
Make default QTL names for the specified loci.
}

