% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\docType{methods}
\name{makeFounderGenoMatrix}
\alias{makeFounderGenoMatrix}
\alias{makeFounderGenoMatrix,DNAStringSet-method}
\alias{makeFounderGenoMatrix,QualityScaledDNAStringSet-method}
\alias{makeFounderGenoMatrix.DNAStringSet}
\alias{makeFounderGenoMatrix.QualityScaledDNAStringSet}
\title{Make founder genotype matrix from genotype data.}
\usage{
makeFounderGenoMatrix(x, founder.geno)

makeFounderGenoMatrix.DNAStringSet(x, founder.geno)

makeFounderGenoMatrix.QualityScaledDNAStringSet(x, founder.geno)

makeFounderGenoMatrix(x, founder.geno)

\S4method{makeFounderGenoMatrix}{DNAStringSet}(x, founder.geno)

\S4method{makeFounderGenoMatrix}{QualityScaledDNAStringSet}(x, founder.geno)
}
\arguments{
\item{x}{Sample genotype data.}

\item{founder.geno}{Founder genotype data.}
}
\value{
A genotype matrix, with genotypes encoded as integers and their
corresponding allele symbols in the attribute \code{'alleles'}.
}
\description{
Given input genotype data for cross segregant samples and their founder 
strains, this function assigns a symbol to each locus according to the 
inferred founder allele.
}
\keyword{internal}

