% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\docType{methods}
\name{makeGeno}
\alias{makeGeno}
\alias{makeGeno,DNAStringSet-method}
\alias{makeGeno,QualityScaledDNAStringSet-method}
\alias{makeGeno.DNAStringSet}
\alias{makeGeno.QualityScaledDNAStringSet}
\title{Make an \pkg{R/qtl} \code{cross} \code{geno} object.}
\usage{
makeGeno(sample.geno, founder.geno = NULL)

makeGeno.DNAStringSet(sample.geno, founder.geno = NULL)

makeGeno.QualityScaledDNAStringSet(sample.geno, founder.geno = NULL)

makeGeno(sample.geno, founder.geno = NULL)

\S4method{makeGeno}{DNAStringSet}(sample.geno, founder.geno = NULL)

\S4method{makeGeno}{QualityScaledDNAStringSet}(sample.geno,
  founder.geno = NULL)
}
\arguments{
\item{sample.geno}{Sample genotype data.}

\item{founder.geno}{Founder genotype data.}
}
\value{
A \code{geno} object.
}
\description{
Make an \pkg{R/qtl} \code{cross} \code{geno} object.
}
\keyword{internal}

