% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makePseudomarkerIDs}
\alias{makePseudomarkerIDs}
\title{Make pseudomarker IDs for the specified loci.}
\usage{
makePseudomarkerIDs(loc)
}
\arguments{
\item{loc}{Locus \code{data.frame} specifying map positions.}
}
\value{
Character vector of pseudomarker IDs.
}
\description{
Make pseudomarker IDs for the specified loci.
}

