% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScanone.R
\name{nodePermScanone}
\alias{nodePermScanone}
\title{Run \code{qtl::scanone} on a single permuted \code{cross} object.}
\usage{
nodePermScanone(perm.id, cross, pheno.col = NULL, perm.pheno = TRUE,
  perm.geno = FALSE, perm.type = c("max", "bins"), ...)
}
\arguments{
\item{perm.id}{Permutation index.}

\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{pheno.col}{Phenotype columns for which QTL analysis should be run. 
If no phenotypes are specified, all are used.}

\item{perm.pheno}{Permute phenotype data.}

\item{perm.geno}{Permute genotype data.}

\item{perm.type}{Type of permutation data to return (see below).}

\item{...}{Additional keyword arguments passed to \code{scanone}.}
}
\value{
If \code{perm.type} is set to \code{'max'}, a regular
\code{scanoneperm} object is returned, containing the maximum
LOD value from the given permutation. If \code{perm.type} is
set to \code{'bins'}, a \code{scanonebins} array is returned.
The single row of this array corresponds to the permutation,
each column corresponds to a bin spanning an interval of LOD
values, and each slice corresponds to a LOD column. Each array
element contains the number of loci in the given bin interval
for that LOD column in that permutation.
}
\description{
Run \code{qtl::scanone} on a single permuted \code{cross} object.
}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})
}
\seealso{
Other scan utility functions: \code{\link{batchPermScanoneF}},
  \code{\link{batchPermScanone}},
  \code{\link{batchPermScantwoF}},
  \code{\link{batchPermScantwo}},
  \code{\link{batchPhenoScanone}},
  \code{\link{batchPhenoScantwo}}, \code{\link{batchScan}},
  \code{\link{nodePermScanoneF}},
  \code{\link{nodePermScantwoF}},
  \code{\link{nodePermScantwo}},
  \code{\link{nodePhenoScanone}},
  \code{\link{nodePhenoScantwo}}
}

