% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScantwo.R
\name{nodePhenoScantwo}
\alias{nodePhenoScantwo}
\title{Run \code{qtl::scantwo} on a single phenotype of a \code{cross} object.}
\usage{
nodePhenoScantwo(pheno.col, cross, ...)
}
\arguments{
\item{pheno.col}{Phenotype column for which QTL analysis should be run.}

\item{cross}{An \pkg{R/qtl} \code{cross} object.}

\item{...}{Additional keyword arguments passed to \code{scantwo}.}
}
\value{
A \code{scantwo} object containing the result of the QTL scan for 
a single phenotype.
}
\description{
Run \code{qtl::scantwo} on a single phenotype of a \code{cross} object.
}
\seealso{
Other scan utilities: \code{\link{batchPermScanoneF}},
  \code{\link{batchPermScanone}},
  \code{\link{batchPermScantwoF}},
  \code{\link{batchPermScantwo}},
  \code{\link{batchPhenoScanone}},
  \code{\link{batchPhenoScantwo}}, \code{\link{batchScan}},
  \code{\link{nodePermScanoneF}},
  \code{\link{nodePermScanone}},
  \code{\link{nodePermScantwoF}},
  \code{\link{nodePermScantwo}},
  \code{\link{nodePhenoScanone}}
}

