% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{orderChr}
\alias{orderChr}
\title{Order chromosome labels.}
\usage{
orderChr(x)
}
\arguments{
\item{x}{Vector of chromosome labels.}
}
\value{
Vector of ordered indices for the input chromosome labels.
}
\description{
Order chromosome labels.
}
\keyword{internal}

