% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{orderSeq}
\alias{orderSeq}
\title{Order sequence labels.}
\usage{
orderSeq(x)
}
\arguments{
\item{x}{Vector of sequence labels.}
}
\value{
Vector of indices for the input sequence labels,
ordered with respect to their normalised form.
}
\description{
Order sequence labels.
}
\seealso{
Other chromosome/sequence functions: \code{\link{formatChr}},
  \code{\link{formatSeq}}, \code{\link{isNormChr}},
  \code{\link{isNormSeq}}, \code{\link{normChr}},
  \code{\link{normSeq}}, \code{\link{orderChr}},
  \code{\link{rankChr}}, \code{\link{rankSeq}},
  \code{\link{sortChr}}, \code{\link{sortSeq}}
}

