% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\docType{package}
\name{Package Constants}
\alias{Package Constants}
\alias{Package Constants-package}
\title{Package constants for \pkg{shmootl}.}
\description{
Package constants for \pkg{shmootl}.
}
\keyword{internal}

