% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonebins.R
\name{padBins}
\alias{padBins}
\title{Pad \code{scanonebins} object to given number of bins.}
\usage{
padBins(x, num.bins)
}
\arguments{
\item{x}{A \code{scanonebins} object.}

\item{num.bins}{Number of bins to which \code{scanonebins} object
should be padded.}
}
\value{
Input \code{scanonebins} object, padded with zeroes to the
specified size.
}
\description{
Pad \code{scanonebins} object to given number of bins.
}
\keyword{internal}

