% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parseDefaultQTLNames}
\alias{parseDefaultQTLNames}
\title{Parse default QTL names.}
\usage{
parseDefaultQTLNames(qtl.names)
}
\arguments{
\item{qtl.names}{Vector of default QTL names.}
}
\value{
Genetic \code{mapframe} with loci corresponding to the specified QTL
names. Raises an error if any of the input values cannot be parsed as a
default QTL name.
}
\description{
Parse default QTL names.
}

