% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonebins.R
\name{parseLODBinLabels}
\alias{parseLODBinLabels}
\title{Parse bin starts from LOD bin labels.}
\usage{
parseLODBinLabels(bin.labels)
}
\arguments{
\item{bin.labels}{Character vector of LOD bin labels to a set of LOD bins.}
}
\value{
Numeric vector of bin starts.
}
\description{
Parse bin starts from LOD bin labels.
}
\keyword{internal}
