% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parsePseudomarkerIDs}
\alias{parsePseudomarkerIDs}
\title{Parse pseudomarker IDs.}
\usage{
parsePseudomarkerIDs(loc.ids)
}
\arguments{
\item{loc.ids}{Vector of pseudomarker IDs.}
}
\value{
Genetic \code{mapframe} with loci
corresponding to the input pseudomarker IDs.
}
\description{
This function parses an input vector of pseudomarker IDs, and returns a
genetic \code{mapframe} with the locus in each row derived from the
corresponding pseudomarker ID in the input vector. An error is raised
if any of the input values cannot be parsed as a pseudomarker ID.
}
\section{Pseudomarker IDs}{


Pseudomarker IDs are used by \pkg{R/qtl} for inter-marker loci. They indicate
the reference sequence and genetic map position of the locus (e.g.
\code{'c04.loc33'} for a locus at position 33cM on chromosome IV).
}

