% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{pull.alleles}
\alias{pull.alleles}
\title{Pull alleles from \pkg{R/qtl} \code{cross}.}
\usage{
pull.alleles(cross)
}
\arguments{
\item{cross}{An \pkg{R/qtl} \code{cross} object.}
}
\value{
Vector of alleles in the \code{cross} object.
}
\description{
Pull alleles from \pkg{R/qtl} \code{cross}.
}
\seealso{
Other cross utilities: \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{pull.chr}}, \code{\link{pull.crosstype}},
  \code{\link{pull.ind}}
}

