% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pull.chr}
\alias{pull.chr}
\alias{pull.chr.cross}
\alias{pull.chr.data.frame}
\alias{pull.chr.geno}
\alias{pull.chr.list}
\alias{pull.chr.map}
\title{Pull chromosomes/sequences from the given object.}
\usage{
pull.chr(x)

\method{pull.chr}{cross}(x)

\method{pull.chr}{data.frame}(x)

\method{pull.chr}{geno}(x)

\method{pull.chr}{list}(x)

\method{pull.chr}{map}(x)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{cross} object,
or other object with chromosomes/sequences.}
}
\value{
Vector of chromosomes/sequences in the given object.
}
\description{
Pull chromosomes/sequences from the given object.
}
\seealso{
Other cross object functions: \code{\link{crossesEqual}},
  \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{hasTimeSeriesPhenotypes}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{permIndices}}, \code{\link{pull.alleles}},
  \code{\link{pull.crosstype}}, \code{\link{pull.ind}},
  \code{\link{readCrossCSV}}, \code{\link{readCrossHDF5}},
  \code{\link{writeCrossCSV}}, \code{\link{writeCrossHDF5}}
}
