% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pull.crosstype}
\alias{pull.crosstype}
\alias{pull.crosstype.cross}
\alias{pull.crosstype.geno}
\title{Pull cross type from the given object.}
\usage{
pull.crosstype(x)

\method{pull.crosstype}{cross}(x)

\method{pull.crosstype}{geno}(x)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{cross} object,
or other object with cross type information.}
}
\value{
Cross type of the given object.
}
\description{
Pull cross type from the given object.
}
\seealso{
Other cross object functions: \code{\link{crossesEqual}},
  \code{\link{getIdColIndex}},
  \code{\link{getPhenoColIndices}},
  \code{\link{hasTimeSeriesPhenotypes}},
  \code{\link{inferStrainIndices}},
  \code{\link{inferTetradIndices}},
  \code{\link{inferTimeStep}},
  \code{\link{interpTimeSeries}},
  \code{\link{padTimeSeries}}, \code{\link{permCross}},
  \code{\link{permIndices}}, \code{\link{pull.alleles}},
  \code{\link{pull.chr}}, \code{\link{pull.ind}},
  \code{\link{readCrossCSV}}, \code{\link{readCrossHDF5}},
  \code{\link{writeCrossCSV}}, \code{\link{writeCrossHDF5}}
}

