% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlintervals.R
\name{qtlintervals}
\alias{qtlintervals}
\title{Create an empty \code{qtlintervals} object.}
\usage{
qtlintervals(drop = NULL, threshold = NULL, alpha = NULL, fdr = NULL,
  ...)
}
\arguments{
\item{drop}{LOD units that the LOD profile must drop to form a QTL interval.}

\item{threshold}{LOD significance threshold for QTL intervals.}

\item{alpha}{Significance level of QTL intervals threshold. (Mutually
exclusive with \code{fdr}.)}

\item{fdr}{False discovery rate (FDR) of QTL intervals threshold. (Mutually
exclusive with \code{alpha}.)}

\item{...}{Unused arguments.}
}
\value{
An empty \code{qtlintervals} list with the given attributes.
}
\description{
Create an empty \code{qtlintervals} object.
}
\keyword{internal}

