% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{rankChr}
\alias{rankChr}
\title{Rank chromosome labels.}
\usage{
rankChr(x)
}
\arguments{
\item{x}{Vector of chromosome labels.}
}
\value{
Vector of ranks for the input chromosome labels.
}
\description{
Rank chromosome labels.
}
\keyword{internal}

