% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{rankSeq}
\alias{rankSeq}
\title{Rank sequence labels.}
\usage{
rankSeq(x)
}
\arguments{
\item{x}{Vector of sequence labels.}
}
\value{
Vector of ranks for the normalised
form of the input sequence labels.
}
\description{
Rank sequence labels.
}
\seealso{
Other chromosome/sequence functions: \code{\link{formatChr}},
  \code{\link{formatSeq}}, \code{\link{isNormChr}},
  \code{\link{isNormSeq}}, \code{\link{normChr}},
  \code{\link{normSeq}}, \code{\link{orderChr}},
  \code{\link{orderSeq}}, \code{\link{rankChr}},
  \code{\link{sortChr}}, \code{\link{sortSeq}}
}

