% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{readGenoCSV}
\alias{readGenoCSV}
\title{Read yeast genotype data from a CSV file.}
\usage{
readGenoCSV(infile, missing.value = "-")
}
\arguments{
\item{infile}{Input CSV file path.}

\item{missing.value}{Missing data value. This can be any single character 
that is not a possible genotype value.}
}
\description{
Read yeast genotype data from a CSV file.
}
\seealso{
Other csv utilities: \code{\link{readCrossCSV}},
  \code{\link{readMapCSV}}, \code{\link{readMapframeCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeGenoCSV}},
  \code{\link{writeMapCSV}}, \code{\link{writeMapframeCSV}}
}

