% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{readMapCSV}
\alias{readMapCSV}
\title{Read \code{map} from a CSV file.}
\usage{
readMapCSV(infile)
}
\arguments{
\item{infile}{Input CSV file path.}
}
\value{
An \pkg{R/qtl} \code{map} object.
}
\description{
Read \code{map} from a CSV file.
}
\seealso{
Other csv utilities: \code{\link{readCrossCSV}},
  \code{\link{readGenoCSV}}, \code{\link{readMapframeCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeGenoCSV}},
  \code{\link{writeMapCSV}}, \code{\link{writeMapframeCSV}}
}

