% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{readMapframeCSV}
\alias{readMapframeCSV}
\title{Read \code{mapframe} from a CSV file.}
\usage{
readMapframeCSV(infile, require.mapunit = TRUE)
}
\arguments{
\item{infile}{Input CSV file path.}

\item{require.mapunit}{Require map unit information.}
}
\value{
A \code{mapframe} object.
}
\description{
Read \code{mapframe} from a CSV file.
}
\seealso{
Other csv utilities: \code{\link{readCovarCSV}},
  \code{\link{readCrossCSV}}, \code{\link{readGenoCSV}},
  \code{\link{readMapCSV}}, \code{\link{readPhenoCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeGenoCSV}},
  \code{\link{writeMapCSV}},
  \code{\link{writeMapframeCSV}},
  \code{\link{writePhenoCSV}}
}

