% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readMapHDF5}
\alias{readMapHDF5}
\title{Read \code{map} from a HDF5 file.}
\usage{
readMapHDF5(infile, name)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{name}{Map name.}
}
\value{
An \pkg{R/qtl} \code{map} object. Returns \code{NULL}
if no map is found with the specified name.
}
\description{
Read \code{map} from a HDF5 file.
}
\details{
A \code{mapframe} dataset is read from the 'Maps' group at the root
of the HDF5 file, and converted to an \pkg{R/qtl} \code{map} object.
}
\keyword{internal}

