% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readMapHDF5}
\alias{readMapHDF5}
\title{Read \code{map} from an input HDF5 file.}
\usage{
readMapHDF5(infile, name)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{name}{Map name.}
}
\value{
An \pkg{R/qtl} \code{map} object. Returns \code{NULL}
if no map is found with the specified name.
}
\description{
Read \code{map} from an input HDF5 file.
}
\details{
A \code{mapframe} dataset is read from the \code{'Maps'} group at
the root of the HDF5 file, and converted to an \pkg{R/qtl} \code{map} object.
}
\seealso{
Other HDF5 functions: \code{\link{readCrossHDF5}},
  \code{\link{readResultHDF5}},
  \code{\link{readResultsOverviewHDF5}},
  \code{\link{writeCrossHDF5}}, \code{\link{writeMapHDF5}},
  \code{\link{writeResultHDF5}},
  \code{\link{writeResultsOverviewHDF5}}

Other map utility functions: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{hasMapCSV}}, \code{\link{inMapOrder}},
  \code{\link{inferMapStep}}, \code{\link{mapframe}},
  \code{\link{mapsEqual}}, \code{\link{orderMap}},
  \code{\link{readMapCSV}}, \code{\link{readMapframeCSV}},
  \code{\link{setMapUnit}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapHDF5}},
  \code{\link{writeMapframeCSV}}
}

