% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readOverviewHDF5}
\alias{readOverviewHDF5}
\title{Read QTL analysis results overview from a HDF5 file.}
\usage{
readOverviewHDF5(infile)
}
\arguments{
\item{infile}{An input HDF5 file.}
}
\value{
A \code{data.frame} containing QTL analysis results overview.
}
\description{
Read QTL analysis results overview from a HDF5 file.
}
\keyword{internal}

