% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readResultHDF5}
\alias{readResultHDF5}
\title{Read QTL analysis result from a HDF5 file.}
\usage{
readResultHDF5(infile, phenotype, name)
}
\arguments{
\item{infile}{An input HDF5 file.}

\item{phenotype}{Name of phenotype (or equivalent analysis unit).}

\item{name}{Name of QTL analysis result.}
}
\value{
R object containing QTL analysis result.
}
\description{
Read QTL analysis result from a HDF5 file.
}
\seealso{
Other hdf5 utilities: \code{\link{readDatasetHDF5}},
  \code{\link{readMapHDF5}},
  \code{\link{readOverviewHDF5}},
  \code{\link{writeDatasetHDF5}},
  \code{\link{writeMapHDF5}},
  \code{\link{writeOverviewHDF5}},
  \code{\link{writeResultHDF5}}
}

