% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{recodeCSV}
\alias{recodeCSV}
\title{Recode \pkg{R/qtl} data in CSV file.}
\usage{
recodeCSV(infile, outfile, geno = NULL, enum.geno = FALSE)
}
\arguments{
\item{infile}{Input CSV file path.}

\item{outfile}{Output CSV file path.}

\item{geno}{Named vector of genotype symbols, with vector names containing
existing genotype symbols, and vector elements containing their replacement
values (incompatible with \code{enum.geno}).}

\item{enum.geno}{Option indicating if genotype data should be recoded as
enumerated genotypes (incompatible with \code{geno}).}
}
\description{
Recode \pkg{R/qtl} data in CSV file.
}
\keyword{internal}

