% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{removeColsNA}
\alias{removeColsNA}
\title{Remove columns containing only \code{NA} values.}
\usage{
removeColsNA(x)
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}.}
}
\value{
Input object without those columns that contain only \code{NA} values.
}
\description{
Remove columns containing only \code{NA} values.
}
\keyword{internal}

