% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{removeRowsNA}
\alias{removeRowsNA}
\title{Remove rows containing only \code{NA} values.}
\usage{
removeRowsNA(x)
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}.}
}
\value{
Input object without those rows that contain only \code{NA} values.
}
\description{
Remove rows containing only \code{NA} values.
}
\keyword{internal}

