% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_annoqtl.R
\name{run_annoqtl}
\alias{run_annoqtl}
\title{Annotate QTL intervals.}
\usage{
run_annoqtl(infile = NA_character_, annofile = NA_character_,
  outfile = NA_character_)
}
\arguments{
\item{infile}{scan result file}

\item{annofile}{GFF annotation file}

\item{outfile}{annotated scan file}
}
\description{
QTL intervals in a scan result file are annotated with features
from the GFF annotation file, and a copy of the scan result file is output,
in which QTL feature annotation has been added. Physical positions of QTL
intervals must be present, either directly in the QTL intervals themselves,
or indirectly by pairing coordinates in genetic and physical maps for the
cross data.
}
\concept{
shmootl:pipelines
}

