% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_annoqtl.R
\name{run_annoqtl}
\alias{run_annoqtl}
\title{Annotate QTL intervals.}
\usage{
run_annoqtl(infile = NA_character_, annofile = NA_character_,
  outfile = NA_character_, pheno = character(), scans = character())
}
\arguments{
\item{infile}{input HDF5 scan file [required]}

\item{annofile}{GFF annotation file [required]}

\item{outfile}{annotated HDF5 scan file [required]}

\item{pheno}{phenotypes [default: all]}

\item{scans}{analyses [default: all]}
}
\description{
QTL intervals in a HDF5 scan file are annotated with features from the GFF
annotation file, and a copy of the HDF5 scan file is output, in which QTL
feature annotation has been added. The GFF annotation file must relate to
the same reference genome as was used in QTL analysis. In addition, physical
positions of QTL intervals must be present, either directly in the QTL
intervals themselves, or indirectly by pairing coordinates in the cross
genetic map to a corresponding physical map. QTL intervals can be annotated
for specific phenotypes and analyses with the \code{pheno} and \code{scans}
parameters, respectively.
}
\seealso{
Other pipeline functions: \code{\link{run_digest}},
  \code{\link{run_estimap}}, \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_pushmap}}, \code{\link{run_recode}},
  \code{\link{run_report}}, \code{\link{run_scanone}}
}
\concept{
shmootl:processing
}

