% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_digest.R
\name{run_digest}
\alias{run_digest}
\title{Create digest of QTL analysis results.}
\usage{
run_digest(h5list = character(), digest = NA_character_,
  pheno = character(), scans = character(), sheets = character(),
  scanfile.pattern = NA_character_)
}
\arguments{
\item{h5list}{list of HDF5 scan files [required]}

\item{digest}{scan digest file [required]}

\item{pheno}{phenotypes to output [default: all]}

\item{scans}{analyses to output [default: all]}

\item{sheets}{worksheets to output}

\item{scanfile.pattern}{scan file name pattern}
}
\description{
Given one or more scan result HDF5 files, create a digest of the results of
the QTL analyses. Results can be output for specific phenotypes, analyses,
and worksheets. These output constraints are applied with parameters
\code{pheno}, \code{scans}, and \code{sheets}, respectively.
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_estimap}}, \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_pushmap}}, \code{\link{run_recode}},
  \code{\link{run_report}}, \code{\link{run_scanone}},
  \code{\link{run_scantwo}}
}
\concept{
shmootl:processing
}
