% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_digest.R
\name{run_digest}
\alias{run_digest}
\title{Create digest of QTL analysis results.}
\usage{
run_digest(h5list = character(), digest = NA_character_,
  scanfile.pattern = NA_character_)
}
\arguments{
\item{h5list}{list of HDF5 scan files [required]}

\item{digest}{scan digest file [required]}

\item{scanfile.pattern}{scan file name pattern}
}
\description{
Given one or more scan result HDF5 files, create
a digest of the results of the QTL analyses.
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_estimap}}, \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_pushmap}}, \code{\link{run_recode}},
  \code{\link{run_report}}, \code{\link{run_scanone}}
}
\concept{
shmootl:processing
}

