% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_estimap.R
\name{run_estimap}
\alias{run_estimap}
\title{Run estimap pipeline.}
\usage{
run_estimap(infile, mapfile, n.cluster = 1L, error.prob = 1e-04,
  map.function = c("haldane", "kosambi", "c-f", "morgan"))
}
\arguments{
\item{infile}{input cross/geno CSV file}

\item{mapfile}{output map CSV file}

\item{n.cluster}{number of threads}

\item{error.prob}{genotyping error rate}

\item{map.function}{genetic map function}
}
\description{
This script reads an R/qtl cross or genotype file, and outputs a
map estimated from the input genotype data. Note that a genetic map cannot be
estimated from enumerated genotypes, as these are generated independently for
each marker.
}

