% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_estimap.R
\name{run_estimap}
\alias{run_estimap}
\title{Estimate map from cross/genotype data.}
\usage{
run_estimap(datafile = NA_character_, mapfile = NA_character_,
  n.cluster = 1L, error.prob = 1e-04, map.function = c("haldane",
  "kosambi", "c-f", "morgan"), jittermap = FALSE)
}
\arguments{
\item{datafile}{input cross/geno CSV file [required]}

\item{mapfile}{output map CSV file [required]}

\item{n.cluster}{number of threads}

\item{error.prob}{genotyping error rate}

\item{map.function}{genetic map function}

\item{jittermap}{jitter map positions}
}
\description{
This script reads an \pkg{R/qtl} cross or genotype file, and outputs
a map estimated from the input genotype data. Note that a genetic map
cannot be estimated from enumerated genotypes, as these are generated
independently for each marker.
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_digest}}, \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_pushmap}}, \code{\link{run_recode}},
  \code{\link{run_report}}, \code{\link{run_scanone}},
  \code{\link{run_scantwo}}
}
\concept{
shmootl:utilities
}
