% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_makecross.R
\name{run_makecross}
\alias{run_makecross}
\title{Make cross CSV file from genotype and phenotype data.}
\usage{
run_makecross(genfile = NA_character_, phefile = NA_character_,
  crossfile = NA_character_)
}
\arguments{
\item{genfile}{input genotype CSV file [required]}

\item{phefile}{input phenotype CSV file [required]}

\item{crossfile}{output cross CSV file [required]}
}
\description{
Read separate genotype and phenotype data files,
and combine these into a single cross data file.
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_digest}}, \code{\link{run_estimap}},
  \code{\link{run_interptimes}},
  \code{\link{run_makegeno}}, \code{\link{run_prep}},
  \code{\link{run_pullmap}}, \code{\link{run_pushmap}},
  \code{\link{run_recode}}, \code{\link{run_report}},
  \code{\link{run_scanone}}, \code{\link{run_scantwo}}
}
\concept{
shmootl:preparation
}

