% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_makegeno.R
\name{run_makegeno}
\alias{run_makegeno}
\title{Run makegeno pipeline.}
\usage{
run_makegeno(samples, genfile, founders = NA, digits = NA)
}
\arguments{
\item{samples}{sample VCF file}

\item{genfile}{output genotype CSV file}

\item{founders}{optional founder VCF file}

\item{digits}{numeric precision [default: unrounded]}
}
\description{
This script will read genotype data from the specified VCF 
files, then write that genotype data to an \pkg{R/qtl} genotype CSV file.
}

