% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pullmap.R
\name{run_pullmap}
\alias{run_pullmap}
\title{Pull map from a data file.}
\usage{
run_pullmap(datafile = NA_character_, mapfile = NA_character_,
  mapname = NA_character_, require.mapunit = TRUE, include.mapunit = TRUE)
}
\arguments{
\item{datafile}{file containing map data [required]}

\item{mapfile}{output map CSV file [required]}

\item{mapname}{name of map (if applicable)}

\item{require.mapunit}{require map units in input}

\item{include.mapunit}{include map units in output}
}
\description{
Pull the map from a CSV or HDF5 file, and write it to a separate map CSV
file. If a HDF5 data file contains multiple maps, the \code{mapname}
parameter must specify which map to pull.
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_digest}}, \code{\link{run_estimap}},
  \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pushmap}},
  \code{\link{run_recode}}, \code{\link{run_report}},
  \code{\link{run_scanone}}, \code{\link{run_scantwo}}
}
\concept{
shmootl:utilities
}

