% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pushmap.R
\name{run_pushmap}
\alias{run_pushmap}
\title{Push map into cross/geno data file.}
\usage{
run_pushmap(infile, mapfile, outfile, require.mapunit = TRUE,
  include.mapunit = TRUE)
}
\arguments{
\item{infile}{input cross/geno CSV file}

\item{mapfile}{input map CSV file}

\item{outfile}{output cross/geno CSV file}

\item{require.mapunit}{require map units in input}

\item{include.mapunit}{include map units in output}
}
\description{
This script reads cross or genotype data from the specified
input file, replaces its map with that of the given map file, and writes
the result to the specified output file.
}

