% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pushmap.R
\name{run_pushmap}
\alias{run_pushmap}
\title{Push map into a data file.}
\usage{
run_pushmap(mapfile = NA_character_, datafile = NA_character_,
  mapname = NA_character_, require.mapunit = TRUE, include.mapunit = TRUE)
}
\arguments{
\item{mapfile}{input map CSV file [required]}

\item{datafile}{file with map data [required]}

\item{mapname}{name of map (if applicable)}

\item{require.mapunit}{require map units in input}

\item{include.mapunit}{include map units in output}
}
\description{
Add or replace the map of the specified CSV or HDF5 file with the
data from the given map file. If the \code{mapname} parameter is
given for a HDF5 data file, the map is assigned that name and when
being written to the HDF5 file. Otherwise, a default map name
(e.g. \code{'Genetic Map'}) is used.
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_digest}}, \code{\link{run_estimap}},
  \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_recode}}, \code{\link{run_report}},
  \code{\link{run_scanone}}, \code{\link{run_scantwo}}
}
\concept{
shmootl:utilities
}
