% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_recode.R
\name{run_recode}
\alias{run_recode}
\title{Recode data in \pkg{R/qtl} CSV file.}
\usage{
run_recode(datafile, geno = mapping(), enum.geno = FALSE)
}
\arguments{
\item{datafile}{cross/geno CSV file}

\item{geno}{recode genotypes from mapping}

\item{enum.geno}{recode to enumerated genotypes}
}
\description{
Recode data in \pkg{R/qtl} CSV file.
}
\concept{
shmootl:utilities
}

