% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_recode.R
\name{run_recode}
\alias{run_recode}
\title{Recode data in \pkg{R/qtl} CSV file.}
\usage{
run_recode(datafile = NA_character_, geno = mapping(), enum.geno = FALSE)
}
\arguments{
\item{datafile}{cross/geno CSV file [required]}

\item{geno}{recode genotypes from mapping}

\item{enum.geno}{recode to enumerated genotypes}
}
\description{
Recode data in an \pkg{R/qtl} cross or genotype CSV file. Genotypes can be
recoded by passing to the \code{geno} parameter a mapping of old to new
genotypes. Alternatively, genotype data can be converted to enumerated
genotypes with the \code{enum.geno} parameter.
}
\details{
When calling this function from within the \code{R} environment, the
\code{geno} parameter must be specified as a mapping object (e.g.
\code{mapping( c(A = 'W', B = 'S') )}). When called from the command line
using \code{Rscript}, the \code{geno} parameter must be specified as a a
YAML string (or YAML file) mapping old to new genotypes
(e.g. \code{"A: W, B: S"}).
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_digest}}, \code{\link{run_estimap}},
  \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_pushmap}}, \code{\link{run_report}},
  \code{\link{run_scanone}}
}
\concept{
shmootl:utilities
}

