% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Create report of \pkg{R/qtl} analysis results.}
\usage{
run_report(h5file = NA_character_, report = NA_character_)
}
\arguments{
\item{h5file}{scan result file}

\item{report}{scan report file}
}
\description{
Given a scan result HDF5 file, create a
report of the results of the QTL analysis.
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_digest}}, \code{\link{run_estimap}},
  \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_pushmap}}, \code{\link{run_recode}},
  \code{\link{run_scanone}}
}
\concept{
shmootl:pipelines
}

