% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Create report of QTL analysis results.}
\usage{
run_report(h5file = NA_character_, report = NA_character_,
  pheno = character(), scans = character())
}
\arguments{
\item{h5file}{HDF5 scan file [required]}

\item{report}{scan report file [required]}

\item{pheno}{phenotypes to output [default: all]}

\item{scans}{analyses to output [default: all]}
}
\description{
Given a scan result HDF5 file, create a report of the results of the QTL
analysis. Results can be output for specific phenotypes and analyses.
These output constraints are applied with parameters \code{pheno} and
\code{scans}, respectively.
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_digest}}, \code{\link{run_estimap}},
  \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_pushmap}}, \code{\link{run_recode}},
  \code{\link{run_scanone}}, \code{\link{run_scantwo}}
}
\concept{
shmootl:processing
}

