% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Create report of \pkg{R/qtl} analysis results.}
\usage{
run_report(scanfile, report)
}
\arguments{
\item{scanfile}{scan result HDF5 file}

\item{report}{scan report file}
}
\description{
Create report of \pkg{R/qtl} analysis results.
}

