% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Create report of QTL analysis results.}
\usage{
run_report(h5file = NA_character_, report = NA_character_,
  analyses = character())
}
\arguments{
\item{h5file}{HDF5 scan file [required]}

\item{report}{scan report file [required]}

\item{analyses}{analyses to output [default: all]}
}
\description{
Given a scan result HDF5 file, create a
report of the results of the QTL analysis.
}
\seealso{
Other pipeline functions: \code{\link{run_annoqtl}},
  \code{\link{run_digest}}, \code{\link{run_estimap}},
  \code{\link{run_interptimes}},
  \code{\link{run_makecross}}, \code{\link{run_makegeno}},
  \code{\link{run_prep}}, \code{\link{run_pullmap}},
  \code{\link{run_pushmap}}, \code{\link{run_recode}},
  \code{\link{run_scanone}}
}
\concept{
shmootl:processing
}

