% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Create report of \pkg{R/qtl} analysis results.}
\usage{
run_report(h5file = NA_character_, report = NA_character_)
}
\arguments{
\item{h5file}{scan result file}

\item{report}{scan report file}
}
\description{
Create report of \pkg{R/qtl} analysis results.
}
\concept{
shmootl:pipelines
}

