% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scanone.R
\name{run_scanone}
\alias{run_scanone}
\title{Do single QTL scan.}
\usage{
run_scanone(infile, outfile, chr = NA, pheno = NA, model = c("normal",
  "binary", "2part", "np"), method = c("em", "imp", "hk", "ehk", "mr",
  "mr-imp", "mr-argmax"), n.perm = 1000L, n.cluster = 1L, alpha = 0.05,
  step = 0, error.prob = 1e-04, map.function = c("haldane", "kosambi",
  "c-f", "morgan"))
}
\arguments{
\item{infile}{input cross file}

\item{outfile}{output result file}

\item{chr}{sequences [default: all]}

\item{pheno}{phenotypes [default: all]}

\item{model}{phenotype model}

\item{method}{method of QTL analysis}

\item{n.perm}{number of permutations}

\item{n.cluster}{number of threads}

\item{alpha}{significance level for LOD threshold}

\item{step}{step size for genotype probabilities}

\item{error.prob}{genotyping error rate}

\item{map.function}{genetic map function}
}
\description{
This script will read cross data from the specified input file, 
run a single QTL analysis using \pkg{R/qtl} \code{scanone}, and write the
results to the specified output file.
}

