% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setGenotypes}
\alias{setGenotypes}
\alias{setGenotypes,CrossInfo-method}
\title{Set genotype symbols.}
\usage{
setGenotypes(cross.info, genotypes)

\S4method{setGenotypes}{CrossInfo}(cross.info, genotypes)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{genotypes}{Vector of genotype symbols.}
}
\value{
Input \code{\linkS4class{CrossInfo}}
object with the given genotypes.
}
\description{
Set genotype symbols.
}
\seealso{
Other CrossInfo methods: \code{\link{getAlleles}},
  \code{\link{getCrosstype}}, \code{\link{getGenotypes}},
  \code{\link{getMarkerIndices}},
  \code{\link{getMarkerNames}},
  \code{\link{getMarkerSeqs}}, \code{\link{getMarkers}},
  \code{\link{getNumMarkers}},
  \code{\link{getNumPhenotypes}},
  \code{\link{getNumSamples}}, \code{\link{getNumSeqs}},
  \code{\link{getPhenotypeIndices}},
  \code{\link{getPhenotypeNames}},
  \code{\link{getPhenotypes}},
  \code{\link{getSampleIndices}},
  \code{\link{getSampleNames}}, \code{\link{getSamples}},
  \code{\link{getSeqIndices}}, \code{\link{getSeqMarkers}},
  \code{\link{getSeqNames}}, \code{\link{getSequences}},
  \code{\link{getStrainIndices}},
  \code{\link{getTetradIndices}},
  \code{\link{hasMarkerSeqs}}, \code{\link{hasSampleIDs}},
  \code{\link{hasStrainIndices}},
  \code{\link{hasTetradIndices}}, \code{\link{setAlleles}},
  \code{\link{setCrosstype}}, \code{\link{setMarkerSeqs}},
  \code{\link{setMarkers}}, \code{\link{setPhenotypes}},
  \code{\link{setSamples}}, \code{\link{setSequences}},
  \code{\link{setStrainIndices}},
  \code{\link{setTetradIndices}}
}
\keyword{internal}
