% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setSamples}
\alias{setSamples}
\alias{setSamples,CrossInfo-method}
\title{Set samples by index or sample ID.}
\usage{
setSamples(cross.info, samples)

\S4method{setSamples}{CrossInfo}(cross.info, samples)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{samples}{Integer vector of sample indices or character vector of 
sample IDs.}
}
\value{
Input \code{CrossInfo} object with the given samples.
}
\description{
Set samples by index or sample ID.
}
\keyword{internal}

