% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setStrainIndices}
\alias{setStrainIndices}
\alias{setStrainIndices,CrossInfo-method}
\title{Set strain indices.}
\usage{
setStrainIndices(cross.info, samples = NULL, strains = NULL)

\S4method{setStrainIndices}{CrossInfo}(cross.info, samples = NULL,
  strains = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{samples}{Vector of samples to assign strain indices.}

\item{strains}{Vector of strain indices.}
}
\value{
Input \code{CrossInfo} object with the given strain indices.
}
\description{
Set strain indices.
}
\keyword{internal}

