% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{setTetradIndices}
\alias{setTetradIndices}
\alias{setTetradIndices,CrossInfo-method}
\title{Set sample tetrad indices.}
\usage{
setTetradIndices(cross.info, samples = NULL, tetrads = NULL)

\S4method{setTetradIndices}{CrossInfo}(cross.info, samples = NULL,
  tetrads = NULL)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}

\item{samples}{Vector of samples for which a tetrad index should be set.}

\item{tetrads}{Vector of tetrad indices.}
}
\value{
Input \code{\linkS4class{CrossInfo}}
object with the given tetrad indices.
}
\description{
Set sample tetrad indices.
}
\seealso{
Other CrossInfo methods: \code{\link{getAlleles}},
  \code{\link{getCrosstype}}, \code{\link{getGenotypes}},
  \code{\link{getMarkerIndices}},
  \code{\link{getMarkerNames}},
  \code{\link{getMarkerSeqs}}, \code{\link{getMarkers}},
  \code{\link{getNumMarkers}},
  \code{\link{getNumPhenotypes}},
  \code{\link{getNumSamples}}, \code{\link{getNumSeqs}},
  \code{\link{getPhenotypeIndices}},
  \code{\link{getPhenotypeNames}},
  \code{\link{getPhenotypes}},
  \code{\link{getSampleIndices}},
  \code{\link{getSampleNames}}, \code{\link{getSamples}},
  \code{\link{getSeqIndices}}, \code{\link{getSeqMarkers}},
  \code{\link{getSeqNames}}, \code{\link{getSequences}},
  \code{\link{getStrainIndices}},
  \code{\link{getTetradIndices}},
  \code{\link{hasMarkerSeqs}}, \code{\link{hasSampleIDs}},
  \code{\link{hasStrainIndices}},
  \code{\link{hasTetradIndices}}, \code{\link{setAlleles}},
  \code{\link{setCrosstype}}, \code{\link{setGenotypes}},
  \code{\link{setMarkerSeqs}}, \code{\link{setMarkers}},
  \code{\link{setPhenotypes}}, \code{\link{setSamples}},
  \code{\link{setSequences}},
  \code{\link{setStrainIndices}}
}
\keyword{internal}

