% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R, R/H5Stack-class.R
\docType{methods}
\name{setValidity}
\alias{setValidity}
\alias{setValidity,CrossInfo-method}
\alias{setValidity,H5Stack-method}
\title{Validate \code{CrossInfo} object.}
\arguments{
\item{object}{A \code{CrossInfo} object.}

\item{object}{A \code{H5Stack} object.}
}
\value{
TRUE if object is valid; otherwise, a character vector of errors.

TRUE if object is valid; otherwise, a character vector of errors.
}
\description{
Validate \code{CrossInfo} object.

Validate \code{H5Stack} object.
}
\keyword{internal}

