% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{sniffCSV}
\alias{sniffCSV}
\title{Identify type of \pkg{R/qtl} data in CSV file.}
\usage{
sniffCSV(infile)
}
\arguments{
\item{infile}{Input CSV file path.}
}
\value{
A string describing the type of data in the input CSV file.
}
\description{
This function identifies the type of data in the input CSV file. This can
be \code{'cross'}, \code{'geno'}, \code{'pheno'}, or \code{'map'}.
Alternatively, it can be \code{'unknown'} if the data cannot be identified.
}
\details{
Note that this function should be used with some discretion, as some
\pkg{R/qtl} input file types cannot be distinguished. For example, a
covariate data file containing a sample ID column will be misidentified
as containing phenotype data.
}
\seealso{
Other CSV functions: \code{\link{hasMapCSV}},
  \code{\link{readCovarCSV}}, \code{\link{readCrossCSV}},
  \code{\link{readGenoCSV}}, \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{readPhenoCSV}}, \code{\link{recodeCSV}},
  \code{\link{writeCrossCSV}}, \code{\link{writeGenoCSV}},
  \code{\link{writeMapCSV}},
  \code{\link{writeMapframeCSV}},
  \code{\link{writePhenoCSV}}
}

