% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{sortChr}
\alias{sortChr}
\title{Sort chromosome labels.}
\usage{
sortChr(x)
}
\arguments{
\item{x}{Vector of chromosome labels.}
}
\value{
Sorted vector of chromosome labels.
}
\description{
Sort chromosome labels.
}
\keyword{internal}

