% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{sortSeq}
\alias{sortSeq}
\title{Sort sequence labels.}
\usage{
sortSeq(x)
}
\arguments{
\item{x}{Vector of sequence labels.}
}
\value{
Sorted vector of sequence labels.
}
\description{
Sort sequence labels.
}
\keyword{internal}

