% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByLoci.R
\name{subsetByLoci}
\alias{subsetByLoci}
\alias{subsetByLoci.map}
\alias{subsetByLoci.mapframe}
\title{Subset object by the specified loci.}
\usage{
subsetByLoci(x, loc = NULL)

\method{subsetByLoci}{map}(x, loc = NULL)

\method{subsetByLoci}{mapframe}(x, loc = NULL)
}
\arguments{
\item{x}{Object subsettable by loci.}

\item{loc}{Locus \code{mapframe} specifying map loci.}
}
\value{
Input object subsetted by the specified loci.
}
\description{
Subset object by the specified loci.
}
\keyword{internal}
