% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByLocusID.R
\docType{methods}
\name{subsetByLocusID}
\alias{subsetByLocusID}
\alias{subsetByLocusID,DNAStringSet-method}
\alias{subsetByLocusID,QualityScaledDNAStringSet-method}
\alias{subsetByLocusID.DNAStringSet}
\alias{subsetByLocusID.QualityScaledDNAStringSet}
\alias{subsetByLocusID.data.frame}
\alias{subsetByLocusID.map}
\title{Subset object by locus ID.}
\usage{
subsetByLocusID(x, predicate)

subsetByLocusID.data.frame(x, predicate)

subsetByLocusID.map(x, predicate)

subsetByLocusID.DNAStringSet(x, predicate)

subsetByLocusID.QualityScaledDNAStringSet(x, predicate)

subsetByLocusID(x, predicate)

\S4method{subsetByLocusID}{DNAStringSet}(x, predicate)

\S4method{subsetByLocusID}{QualityScaledDNAStringSet}(x, predicate)
}
\arguments{
\item{x}{Object with locus IDs.}

\item{predicate}{Function applied to locus IDs that determines which loci are
retained in the output object.}
}
\value{
Input object containing only loci whose locus IDs are evaluated as 
true by the predicate function.
}
\description{
Subset object by locus ID.
}
\keyword{internal}

