% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetBySeq.R
\name{subsetBySeq}
\alias{subsetBySeq}
\alias{subsetBySeq.character}
\alias{subsetBySeq.data.frame}
\alias{subsetBySeq.map}
\alias{subsetBySeq.qtlintervals}
\title{Subset object by the specified sequences.}
\usage{
subsetBySeq(x, sequences = NULL)

\method{subsetBySeq}{character}(x, sequences = NULL)

\method{subsetBySeq}{data.frame}(x, sequences = NULL)

\method{subsetBySeq}{map}(x, sequences = NULL)

\method{subsetBySeq}{qtlintervals}(x, sequences = NULL)
}
\arguments{
\item{x}{Object subsettable by sequence.}

\item{sequences}{Sequences with which to subset object.}
}
\value{
Input object subsetted by the specified sequences.
}
\description{
Subset object by the specified sequences.
}
\keyword{internal}

