% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossInfo-class.R
\docType{methods}
\name{validateCrosstype}
\alias{validateCrosstype}
\alias{validateCrosstype,CrossInfo-method}
\title{Validate cross type information.}
\usage{
validateCrosstype(cross.info)

\S4method{validateCrosstype}{CrossInfo}(cross.info)
}
\arguments{
\item{cross.info}{A \code{\link{CrossInfo}} object.}
}
\value{
\code{TRUE} if cross type is valid; otherwise, returns first error.
}
\description{
Validate cross type information.
}
\keyword{internal}
