% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{writeDatasetHDF5}
\alias{writeDatasetHDF5}
\alias{writeDatasetHDF5.array}
\alias{writeDatasetHDF5.data.frame}
\alias{writeDatasetHDF5.default}
\alias{writeDatasetHDF5.list}
\alias{writeDatasetHDF5.map}
\alias{writeDatasetHDF5.mapframe}
\alias{writeDatasetHDF5.matrix}
\alias{writeDatasetHDF5.scanone}
\alias{writeDatasetHDF5.scanonebins}
\alias{writeDatasetHDF5.scanoneperm}
\alias{writeDatasetHDF5.scantwo}
\alias{writeDatasetHDF5.scantwoperm}
\title{Write R object to a HDF5 dataset.
  
The given R object is written to the named HDF5 dataset using the method
for the given dataset type, as determined by the class of the R object.}
\usage{
writeDatasetHDF5(dataset, outfile, h5name, ...)

\method{writeDatasetHDF5}{array}(dataset, outfile, h5name, ...)

\method{writeDatasetHDF5}{data.frame}(dataset, outfile, h5name,
  rownames.column = "rownames", ...)

\method{writeDatasetHDF5}{default}(dataset, outfile, h5name,
  overwrite = FALSE, ...)

\method{writeDatasetHDF5}{list}(dataset, outfile, h5name, ...)

\method{writeDatasetHDF5}{map}(dataset, outfile, h5name, ...)

\method{writeDatasetHDF5}{mapframe}(dataset, outfile, h5name, ...)

\method{writeDatasetHDF5}{matrix}(dataset, outfile, h5name, ...)

\method{writeDatasetHDF5}{scanone}(dataset, outfile, h5name, ...)

\method{writeDatasetHDF5}{scanonebins}(dataset, outfile, h5name, ...)

\method{writeDatasetHDF5}{scanoneperm}(dataset, outfile, h5name, ...)

\method{writeDatasetHDF5}{scantwo}(dataset, outfile, h5name, ...)

\method{writeDatasetHDF5}{scantwoperm}(dataset, outfile, h5name, ...)
}
\arguments{
\item{dataset}{R object.}

\item{outfile}{An output HDF5 file.}

\item{h5name}{HDF5 dataset name.}

\item{...}{Further arguments (see below).}

\item{rownames.column}{For a \code{data.frame}, any rownames are removed,
and then inserted into a column of the main table, with its column name
taken from this parameter.}

\item{overwrite}{Option indicating whether to force overwrite of an
existing dataset. By default, existing datasets cannot be overwritten.}
}
\description{
Write R object to a HDF5 dataset.
  
The given R object is written to the named HDF5 dataset using the method
for the given dataset type, as determined by the class of the R object.
}
\seealso{
Other HDF5 functions: \code{\link{readDatasetHDF5}},
  \code{\link{readMapHDF5}},
  \code{\link{readOverviewHDF5}},
  \code{\link{readResultHDF5}}, \code{\link{writeMapHDF5}},
  \code{\link{writeOverviewHDF5}},
  \code{\link{writeResultHDF5}}
}

