% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{writeDigestExcel}
\alias{writeDigestExcel}
\title{Write an Excel digest of QTL scan results.}
\usage{
writeDigestExcel(scanfiles, digest)
}
\arguments{
\item{scanfiles}{One or more QTL scan result HDF5 files.}

\item{digest}{Path of output digest Excel file.}
}
\description{
Write an Excel digest of QTL scan results.
}

