% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{writeGenoCSV}
\alias{writeGenoCSV}
\title{Write yeast genotype data to a CSV file.}
\usage{
writeGenoCSV(geno, outfile, chr = NULL, digits = NULL,
  include.mapunit = TRUE)
}
\arguments{
\item{geno}{An \pkg{R/qtl} \code{cross} \code{geno} object.}

\item{outfile}{Output CSV file path.}

\item{chr}{Vector of sequences for which genotype data should be included in 
the output file. If none are specified, a genotype file is output for all 
available sequences.}

\item{digits}{If specified, round genetic map positions to the specified 
number of digits.}

\item{include.mapunit}{Include map unit information in map positions.}
}
\description{
This function writes a yeast \code{geno} object to an
\pkg{R/qtl} CSV file. The \code{geno} object must have an attribute
\code{'info'} of type \code{\linkS4class{CrossInfo}}, from which
genotype and sample IDs are taken.
}
\seealso{
Other CSV functions: \code{\link{hasMapCSV}},
  \code{\link{readCovarCSV}}, \code{\link{readCrossCSV}},
  \code{\link{readGenoCSV}}, \code{\link{readMapCSV}},
  \code{\link{readMapframeCSV}},
  \code{\link{readPhenoCSV}}, \code{\link{recodeCSV}},
  \code{\link{sniffCSV}}, \code{\link{writeCrossCSV}},
  \code{\link{writeMapCSV}},
  \code{\link{writeMapframeCSV}},
  \code{\link{writePhenoCSV}}
}

