% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{writeMapHDF5}
\alias{writeMapHDF5}
\title{Write \code{map} to an output HDF5 file.}
\usage{
writeMapHDF5(map, outfile, name = NULL, overwrite = FALSE)
}
\arguments{
\item{map}{An \pkg{R/qtl} \code{map} object.}

\item{outfile}{An output HDF5 file.}

\item{name}{Map name.}

\item{overwrite}{Option indicating whether to force overwrite of an
existing map. By default, existing maps cannot be overwritten.}
}
\description{
Write \code{map} to an output HDF5 file.
}
\details{
An \pkg{R/qtl} \code{map} object is written to the \code{'Maps'}
group at the root of the given HDF5 file. If no map name is specified, a
default will be assigned based on whether it is a genetic or physical map.
}
\seealso{
Other HDF5 functions: \code{\link{readCrossHDF5}},
  \code{\link{readMapHDF5}}, \code{\link{readResultHDF5}},
  \code{\link{readResultsOverviewHDF5}},
  \code{\link{writeCrossHDF5}},
  \code{\link{writeResultHDF5}},
  \code{\link{writeResultsOverviewHDF5}}

Other map utility functions: \code{\link{as.data.frame.mapframe}},
  \code{\link{as.data.frame.map}},
  \code{\link{as.mapframe}}, \code{\link{as.map}},
  \code{\link{extractMarkers}},
  \code{\link{extractPseudomarkers}},
  \code{\link{findFlanking}}, \code{\link{findLoci}},
  \code{\link{findMarkers}}, \code{\link{getMapSteps}},
  \code{\link{getMapUnit}}, \code{\link{gmapframe}},
  \code{\link{hasMapCSV}}, \code{\link{inMapOrder}},
  \code{\link{inferMapStep}}, \code{\link{mapframe}},
  \code{\link{mapsEqual}}, \code{\link{orderMap}},
  \code{\link{readMapCSV}}, \code{\link{readMapHDF5}},
  \code{\link{readMapframeCSV}}, \code{\link{setMapUnit}},
  \code{\link{writeMapCSV}}, \code{\link{writeMapframeCSV}}
}
