% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{writeOverviewHDF5}
\alias{writeOverviewHDF5}
\title{Write QTL analysis results overview to an output HDF5 file.}
\usage{
writeOverviewHDF5(overview, outfile, overwrite = FALSE)
}
\arguments{
\item{overview}{Object containing a QTL analysis results overview.}

\item{outfile}{An output HDF5 file.}

\item{overwrite}{Option indicating whether to force overwrite of an
existing overview. By default, existing overviews cannot be overwritten.}
}
\description{
Write QTL analysis results overview to an output HDF5 file.
}
\seealso{
Other HDF5 functions: \code{\link{readCrossHDF5}},
  \code{\link{readMapHDF5}},
  \code{\link{readOverviewHDF5}},
  \code{\link{readResultHDF5}},
  \code{\link{writeCrossHDF5}}, \code{\link{writeMapHDF5}},
  \code{\link{writeResultHDF5}}
}

