% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{writeOverviewHDF5}
\alias{writeOverviewHDF5}
\title{Write QTL analysis results overview to a HDF5 file.}
\usage{
writeOverviewHDF5(overview, outfile)
}
\arguments{
\item{overview}{Object containing a QTL analysis results overview.}

\item{outfile}{An output HDF5 file.}
}
\description{
Write QTL analysis results overview to a HDF5 file.
}
\seealso{
Other hdf5 utilities: \code{\link{readDatasetHDF5}},
  \code{\link{readMapHDF5}},
  \code{\link{readOverviewHDF5}},
  \code{\link{readResultHDF5}},
  \code{\link{writeDatasetHDF5}},
  \code{\link{writeMapHDF5}}, \code{\link{writeResultHDF5}}
}

