% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{writePhenoCSV}
\alias{writePhenoCSV}
\title{Write yeast pheno data to a CSV file.}
\usage{
writePhenoCSV(pheno, outfile, digits = NULL)
}
\arguments{
\item{pheno}{An \pkg{R/qtl} \code{cross} \code{pheno} object.}

\item{outfile}{Output CSV file path.}

\item{digits}{If specified, round numeric phenotype values to the specified
number of digits.}
}
\description{
Write yeast pheno data to a CSV file.
}
\seealso{
Other csv utilities: \code{\link{readCovarCSV}},
  \code{\link{readCrossCSV}}, \code{\link{readGenoCSV}},
  \code{\link{readMapCSV}}, \code{\link{readMapframeCSV}},
  \code{\link{readPhenoCSV}}, \code{\link{writeCrossCSV}},
  \code{\link{writeGenoCSV}}, \code{\link{writeMapCSV}},
  \code{\link{writeMapframeCSV}}
}

