% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{writeReportPDF}
\alias{writeReportPDF}
\title{Write a PDF report of QTL scan results.}
\usage{
writeReportPDF(scanfile, report, analyses = NULL)
}
\arguments{
\item{scanfile}{A scan result HDF5 file.}

\item{report}{Path of output report PDF file.}

\item{analyses}{Analyses for which results should be included in the report
file. If none are specified, results are output for all available analyses.}
}
\description{
Write a PDF report of QTL scan results.
}

