% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{writeResultHDF5}
\alias{writeResultHDF5}
\title{Write QTL analysis result to an output HDF5 file.}
\usage{
writeResultHDF5(result, outfile, phenotype, name, overwrite = FALSE)
}
\arguments{
\item{result}{R object containing a QTL analysis result.}

\item{outfile}{An output HDF5 file.}

\item{phenotype}{Name of phenotype (or equivalent analysis unit).}

\item{name}{Name of QTL analysis result.}

\item{overwrite}{Option indicating whether to force overwrite of an
existing result. By default, existing results cannot be overwritten.}
}
\description{
Write QTL analysis result to an output HDF5 file.
}
\seealso{
Other HDF5 functions: \code{\link{readCrossHDF5}},
  \code{\link{readMapHDF5}},
  \code{\link{readOverviewHDF5}},
  \code{\link{readResultHDF5}},
  \code{\link{writeCrossHDF5}}, \code{\link{writeMapHDF5}},
  \code{\link{writeOverviewHDF5}}
}

